% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_abbreviations_lup.R
\docType{data}
\name{abbreviations_lup}
\alias{abbreviations_lup}
\title{Common abbreviations lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 503 rows and 3 columns.
}
\source{
\url{https://doi.org/10.7910/DVN/2Y9VF9}
}
\usage{
abbreviations_lup
}
\description{
A lookup table for abbreviations commonly used in object names in the TTUpackage.
}
\details{
A tibble

\describe{
\item{short_name_chr}{Short name (a character vector)}
\item{long_name_chr}{Long name (a character vector)}
\item{plural_lgl}{Plural (a logical vector)}
}
}
\keyword{datasets}
