% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_adol6d_scores}
\alias{add_adol6d_scores}
\title{Add adol6d scores}
\usage{
add_adol6d_scores(
  unscored_aqol_tb,
  prefix_1L_chr = "aqol6d_q",
  id_var_nm_1L_chr = "fkClientID",
  wtd_aqol_var_nm_1L_chr = "aqol6d_total_w"
)
}
\arguments{
\item{unscored_aqol_tb}{Unscored Assessment of Quality of Life (a tibble)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'aqol6d_q'}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{wtd_aqol_var_nm_1L_chr}{Wtd Assessment of Quality of Life var name (a character vector of length one), Default: 'aqol6d_total_w'}
}
\value{
Transformed Assessment of Quality of Life (a tibble)
}
\description{
add_adol6d_scores() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add adol6d scores. Function argument unscored_aqol_tb specifies the object to be updated. The function returns Transformed Assessment of Quality of Life (a tibble).
}
