% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol6d_items_to_aqol6d_tbs_ls}
\alias{add_aqol6d_items_to_aqol6d_tbs_ls}
\title{Add Assessment of Quality of Life Six Dimension items to Assessment of Quality of Life Six Dimension tibbles}
\usage{
add_aqol6d_items_to_aqol6d_tbs_ls(
  aqol6d_tbs_ls,
  aqol_items_props_tbs_ls,
  prefix_chr,
  aqol_tots_var_nms_chr,
  id_var_nm_1L_chr = "fkClientID",
  scaling_cnst_dbl = 5
)
}
\arguments{
\item{aqol6d_tbs_ls}{Assessment of Quality of Life Six Dimension tibbles (a list)}

\item{aqol_items_props_tbs_ls}{Assessment of Quality of Life items props tibbles (a list)}

\item{prefix_chr}{Prefix (a character vector)}

\item{aqol_tots_var_nms_chr}{Assessment of Quality of Life totals var names (a character vector)}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{scaling_cnst_dbl}{Scaling cnst (a double vector), Default: 5}
}
\value{
Updated Assessment of Quality of Life Six Dimension tibbles (a list)
}
\description{
add_aqol6d_items_to_aqol6d_tbs_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add assessment of quality of life six dimension items to assessment of quality of life six dimension tibbles list. Function argument aqol6d_tbs_ls specifies the object to be updated. The function returns Updated Assessment of Quality of Life Six Dimension tibbles (a list).
}
