% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol6dU_to_aqol6d_items_tb}
\alias{add_aqol6dU_to_aqol6d_items_tb}
\title{Add Assessment of Quality of Life Six Dimension Health Utility to Assessment of Quality of Life Six Dimension items}
\usage{
add_aqol6dU_to_aqol6d_items_tb(
  aqol6d_items_tb,
  coeffs_lup_tb = aqol6d_from_8d_coeffs_lup_tb
)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension items (a tibble)}

\item{coeffs_lup_tb}{Coeffs lookup table (a tibble), Default: aqol6d_from_8d_coeffs_lup_tb}
}
\value{
Assessment of Quality of Life Six Dimension items (a tibble)
}
\description{
add_aqol6dU_to_aqol6d_items_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add assessment of quality of life six dimension health utility to assessment of quality of life six dimension items tibble. Function argument aqol6d_items_tb specifies the object to be updated. The function returns Assessment of Quality of Life Six Dimension items (a tibble).
}
\keyword{internal}
