% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol6dU_to_aqol6d_tbs_ls}
\alias{add_aqol6dU_to_aqol6d_tbs_ls}
\title{Add Assessment of Quality of Life Six Dimension Health Utility to Assessment of Quality of Life Six Dimension tibbles}
\usage{
add_aqol6dU_to_aqol6d_tbs_ls(
  aqol6d_tbs_ls,
  prefix_1L_chr = "aqol6d_q",
  id_var_nm_1L_chr
)
}
\arguments{
\item{aqol6d_tbs_ls}{Assessment of Quality of Life Six Dimension tibbles (a list)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'aqol6d_q'}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one)}
}
\value{
Assessment of Quality of Life Six Dimension tibbles (a list)
}
\description{
add_aqol6dU_to_aqol6d_tbs_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add assessment of quality of life six dimension health utility to assessment of quality of life six dimension tibbles list. Function argument aqol6d_tbs_ls specifies the object to be updated. The function returns Assessment of Quality of Life Six Dimension tibbles (a list).
}
\keyword{internal}
