% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol_dim_scrg_eqs}
\alias{add_aqol_dim_scrg_eqs}
\title{Add Assessment of Quality of Life health utility dimension scoring equations}
\usage{
add_aqol_dim_scrg_eqs(unscored_aqol_tb)
}
\arguments{
\item{unscored_aqol_tb}{Unscored Assessment of Quality of Life health utility (a tibble)}
}
\value{
Unscored Assessment of Quality of Life health utility (a tibble)
}
\description{
add_aqol_dim_scrg_eqs() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add assessment of quality of life health utility dimension scoring equations. Function argument unscored_aqol_tb specifies the object to be updated. The function returns Unscored Assessment of Quality of Life health utility (a tibble).
}
