% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol_scores_tbs_ls}
\alias{add_aqol_scores_tbs_ls}
\title{Add Assessment of Quality of Life health utility scores tibbles}
\usage{
add_aqol_scores_tbs_ls(tbs_ls, means_dbl, sds_dbl, corr_dbl)
}
\arguments{
\item{tbs_ls}{Tibbles (a list)}

\item{means_dbl}{Means (a double vector)}

\item{sds_dbl}{Sds (a double vector)}

\item{corr_dbl}{Corr (a double vector)}
}
\value{
Tibbles (a list)
}
\description{
add_aqol_scores_tbs_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add assessment of quality of life health utility scores tibbles list. Function argument tbs_ls specifies the object to be updated. The function returns Tibbles (a list).
}
