% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_cors_and_utls_to_aqol6d_tbs_ls}
\alias{add_cors_and_utls_to_aqol6d_tbs_ls}
\title{Add correlations and utilities to Assessment of Quality of Life Six Dimension tibbles}
\usage{
add_cors_and_utls_to_aqol6d_tbs_ls(
  aqol6d_tbs_ls,
  aqol_scores_pars_ls,
  aqol_items_prpns_tbs_ls,
  temporal_cors_ls,
  prefix_chr,
  aqol_tots_var_nms_chr,
  id_var_nm_1L_chr = "fkClientID"
)
}
\arguments{
\item{aqol6d_tbs_ls}{Assessment of Quality of Life Six Dimension tibbles (a list)}

\item{aqol_scores_pars_ls}{Assessment of Quality of Life scores parameters (a list)}

\item{aqol_items_prpns_tbs_ls}{Assessment of Quality of Life items proportions tibbles (a list)}

\item{temporal_cors_ls}{Temporal correlations (a list)}

\item{prefix_chr}{Prefix (a character vector)}

\item{aqol_tots_var_nms_chr}{Assessment of Quality of Life totals variable names (a character vector)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}
}
\value{
Assessment of Quality of Life Six Dimension tibbles (a list)
}
\description{
add_cors_and_utls_to_aqol6d_tbs_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add correlations and utilities to assessment of quality of life six dimension tibbles list. Function argument aqol6d_tbs_ls specifies the object to be updated. The function returns Assessment of Quality of Life Six Dimension tibbles (a list).
}
\keyword{internal}
