% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_dim_disv_to_aqol6d_items_tb}
\alias{add_dim_disv_to_aqol6d_items_tb}
\title{Add dimension disvalue to Assessment of Quality of Life Six Dimension items}
\usage{
add_dim_disv_to_aqol6d_items_tb(
  aqol6d_items_tb,
  domain_items_ls,
  domains_chr,
  dim_sclg_con_lup_tb = aqol6d_dim_sclg_con_lup_tb,
  itm_wrst_wghts_lup_tb = aqol6d_adult_itm_wrst_wghts_lup_tb
)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension items (a tibble)}

\item{domain_items_ls}{Domain items (a list)}

\item{domains_chr}{Domains (a character vector)}

\item{dim_sclg_con_lup_tb}{Dimension sclg constant lookup table (a tibble), Default: aqol6d_dim_sclg_con_lup_tb}

\item{itm_wrst_wghts_lup_tb}{Itm wrst wghts lookup table (a tibble), Default: aqol6d_adult_itm_wrst_wghts_lup_tb}
}
\value{
Assessment of Quality of Life Six Dimension items (a tibble)
}
\description{
add_dim_disv_to_aqol6d_items_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add dimension disvalue to assessment of quality of life six dimension items tibble. Function argument aqol6d_items_tb specifies the object to be updated. The function returns Assessment of Quality of Life Six Dimension items (a tibble).
}
\keyword{internal}
