% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_dim_scores_to_aqol6d_items_tb}
\alias{add_dim_scores_to_aqol6d_items_tb}
\title{Add dimension scores to Assessment of Quality of Life Six Dimension items}
\usage{
add_dim_scores_to_aqol6d_items_tb(aqol6d_items_tb, domain_items_ls)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension items (a tibble)}

\item{domain_items_ls}{Domain items (a list)}
}
\value{
Assessment of Quality of Life Six Dimension items (a tibble)
}
\description{
add_dim_scores_to_aqol6d_items_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add dimension scores to assessment of quality of life six dimension items tibble. Function argument aqol6d_items_tb specifies the object to be updated. The function returns Assessment of Quality of Life Six Dimension items (a tibble).
}
\keyword{internal}
