% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_dmn_disu_to_aqol6d_items_tb_tb}
\alias{add_dmn_disu_to_aqol6d_items_tb_tb}
\title{Add dmn disu to Assessment of Quality of Life Six Dimension health utility items tibble}
\usage{
add_dmn_disu_to_aqol6d_items_tb_tb(
  aqol6d_items_tb,
  domain_items_ls,
  domains_chr,
  aqol6d_dim_sclg_cnt_lup_tb = aqol6d_dim_sclg_cnt_lup_tb,
  aqol6d_adult_itm_wrst_wghts_lup_tb = aqol6d_adult_itm_wrst_wghts_lup_tb
)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension health utility items (a tibble)}

\item{domain_items_ls}{Domain items (a list)}

\item{domains_chr}{Domains (a character vector)}

\item{aqol6d_dim_sclg_cnt_lup_tb}{Assessment of Quality of Life Six Dimension health utility dimension sclg constant lookup table (a tibble), Default: aqol6d_dim_sclg_cnt_lup_tb}

\item{aqol6d_adult_itm_wrst_wghts_lup_tb}{Assessment of Quality of Life Six Dimension health utility adult itm wrst wghts lookup table (a tibble), Default: aqol6d_adult_itm_wrst_wghts_lup_tb}
}
\value{
an Assessment of Quality of Life Six Dimension health utility items (a tibble)
}
\description{
add_dmn_disu_to_aqol6d_items_tb_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add dmn disu to assessment of quality of life six dimension health utility items tibble tibble. Function argument aqol6d_items_tb specifies the object to be updated. The function returns an Assessment of Quality of Life Six Dimension health utility items (a tibble).
}
