% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_domain_unwtd_tots_tb}
\alias{add_domain_unwtd_tots_tb}
\title{Add domain unwtd tots}
\usage{
add_domain_unwtd_tots_tb(items_tb, domain_items_ls, domain_pfx_1L_chr)
}
\arguments{
\item{items_tb}{Items (a tibble)}

\item{domain_items_ls}{Domain items (a list)}

\item{domain_pfx_1L_chr}{Domain prefix 1L (a character vector of length 1)}
}
\value{
Items and domains (a tibble)
}
\description{
add_domain_unwtd_tots_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add domain unwtd tots. Function argument items_tb specifies the object to be updated. The function returns items and domains (a tibble).
}
\keyword{internal}
