% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_interval_var}
\alias{add_interval_var}
\title{Add interval var}
\usage{
add_interval_var(
  data_tb,
  id_var_nm_1L_chr = "fkClientID",
  msrmnt_date_var_nm_1L_chr = "d_interview_date",
  time_unit_1L_chr = "days",
  bl_date_var_nm_1L_chr = "bl_date_dtm",
  interval_var_nm_1L_chr = "interval_dbl",
  temp_row_nbr_var_nm_1L_chr = "temp_row_nbr_int",
  drop_bl_date_var = F
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{msrmnt_date_var_nm_1L_chr}{Msrmnt date var name (a character vector of length one), Default: 'd_interview_date'}

\item{time_unit_1L_chr}{Time unit (a character vector of length one), Default: 'days'}

\item{bl_date_var_nm_1L_chr}{Bl date var name (a character vector of length one), Default: 'bl_date_dtm'}

\item{interval_var_nm_1L_chr}{Interval var name (a character vector of length one), Default: 'interval_dbl'}

\item{temp_row_nbr_var_nm_1L_chr}{Temp row nbr var name (a character vector of length one), Default: 'temp_row_nbr_int'}

\item{drop_bl_date_var}{PARAM_DESCRIPTION, Default: F}
}
\value{
Updated data (a tibble)
}
\description{
add_interval_var() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add interval var. Function argument data_tb specifies the object to be updated. The function returns Updated data (a tibble).
}
