% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_itm_disv_to_aqol6d_itms_tb}
\alias{add_itm_disv_to_aqol6d_itms_tb}
\title{Add itm disvalue to Assessment of Quality of Life Six Dimension itms}
\usage{
add_itm_disv_to_aqol6d_itms_tb(
  aqol6d_items_tb,
  disvalues_lup_tb = aqol6d_adult_disv_lup_tb,
  pfx_1L_chr
)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension items (a tibble)}

\item{disvalues_lup_tb}{Disvalues lookup table (a tibble), Default: aqol6d_adult_disv_lup_tb}

\item{pfx_1L_chr}{Prefix (a character vector of length one)}
}
\value{
Assessment of Quality of Life Six Dimension items (a tibble)
}
\description{
add_itm_disv_to_aqol6d_itms_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add itm disvalue to assessment of quality of life six dimension itms tibble. Function argument aqol6d_items_tb specifies the object to be updated. The function returns Assessment of Quality of Life Six Dimension items (a tibble).
}
