% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_prefd_predr_var_to_mdl_smry_ls}
\alias{add_prefd_predr_var_to_mdl_smry_ls}
\title{Add preferred predictor variable to model summary list}
\usage{
add_prefd_predr_var_to_mdl_smry_ls(mdl_smry_ls, ds_smry_ls)
}
\arguments{
\item{mdl_smry_ls}{Model summary (a list)}

\item{ds_smry_ls}{Dataset summary (a list)}
}
\value{
Model summary (a list)
}
\description{
add_prefd_predr_var_to_mdl_smry_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add preferred predictor variable to model summary list. Function argument mdl_smry_ls specifies the object to be updated. The function returns Model summary (a list).
}
\keyword{internal}
