% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_tfmd_var_to_ds}
\alias{add_tfmd_var_to_ds}
\title{Add tfmd variable to dataset}
\usage{
add_tfmd_var_to_ds(data_tb, depnt_var_nm_1L_chr, tfmn_1L_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one)}
}
\value{
Data (a tibble)
}
\description{
add_tfmd_var_to_ds() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add tfmd variable to dataset. Function argument data_tb specifies the object to be updated. The function returns Data (a tibble).
}
\keyword{internal}
