% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_uids_to_tbs_ls}
\alias{add_uids_to_tbs_ls}
\title{Add uids to tibbles}
\usage{
add_uids_to_tbs_ls(tbs_ls, prefix_1L_chr)
}
\arguments{
\item{tbs_ls}{Tibbles (a list)}

\item{prefix_1L_chr}{Prefix 1L (a character vector of length 1)}
}
\value{
Tibbles (a list)
}
\description{
add_uids_to_tbs_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add uids to a tibbles. Function argument tbs_ls specifies the object to be updated. The function returns a tibbles (a list).
}
\keyword{internal}
