% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_adol_dim_scalg_eqs_lup.R
\docType{data}
\name{adol_dim_scalg_eqs_lup}
\alias{adol_dim_scalg_eqs_lup}
\title{AQoL6D (adolescent) item worst weightings equations lookup table}
\format{
An object of class \code{data.frame} with 19 rows and 3 columns.
}
\source{
\url{https://www.aqol.com.au/index.php/scoring-algorithms}
}
\usage{
adol_dim_scalg_eqs_lup
}
\description{
Dimension scaling equations for adolescent version of AQoL6D scoring algorithm.
}
\details{
A tibble

\describe{
\item{Dim_scal}{NO MATCH}
\item{Label}{NO MATCH}
\item{Equ}{NO MATCH}
}
}
\keyword{datasets}
