% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_aqol6d_adult_disv_lup_tb.R
\docType{data}
\name{aqol6d_adult_disv_lup_tb}
\alias{aqol6d_adult_disv_lup_tb}
\title{AQoL6D (adult version) item disvalues lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 7 columns.
}
\source{
\url{https://www.aqol.com.au/index.php/scoring-algorithms}
}
\usage{
aqol6d_adult_disv_lup_tb
}
\description{
Disutility weights for individual AQoL6D (adult version) items.
}
\details{
A tibble

\describe{
\item{Question_chr}{Question (a character vector)}
\item{Answer_1_dbl}{Answer 1 (a double vector)}
\item{Answer_2_dbl}{Answer 2 (a double vector)}
\item{Answer_3_dbl}{Answer 3 (a double vector)}
\item{Answer_4_dbl}{Answer 4 (a double vector)}
\item{Answer_5_dbl}{Answer 5 (a double vector)}
\item{Answer_6_dbl}{Answer 6 (a double vector)}
}
}
\keyword{datasets}
