% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_aqol6d_adult_itm_wrst_wghts_lup_tb.R
\docType{data}
\name{aqol6d_adult_itm_wrst_wghts_lup_tb}
\alias{aqol6d_adult_itm_wrst_wghts_lup_tb}
\title{AQoL6D (adult) item worst weightings lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 2 columns.
}
\source{
\url{https://www.aqol.com.au/index.php/scoring-algorithms}
}
\usage{
aqol6d_adult_itm_wrst_wghts_lup_tb
}
\description{
Worst weightings for individual items in AQoL6D (adult version).
}
\details{
A tibble

\describe{
\item{Question_chr}{Question (a character vector)}
\item{Worst_Weight_dbl}{Worst Weight (a double vector)}
}
}
\keyword{datasets}
