% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_aqol6d_dim_sclg_con_lup_tb.R
\docType{data}
\name{aqol6d_dim_sclg_con_lup_tb}
\alias{aqol6d_dim_sclg_con_lup_tb}
\title{AQoL6D dimension scaling constants lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\source{
\url{https://www.aqol.com.au/index.php/scoring-algorithms}
}
\usage{
aqol6d_dim_sclg_con_lup_tb
}
\description{
Scaling constants for each dimension of AQoL6D.
}
\details{
A tibble

\describe{
\item{Dimension_chr}{Dimension (a character vector)}
\item{Constant_dbl}{Constant (a double vector)}
}
}
\keyword{datasets}
