% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_aqol6d_domain_qs_lup_tb.R
\docType{data}
\name{aqol6d_domain_qs_lup_tb}
\alias{aqol6d_domain_qs_lup_tb}
\title{AQoL6D dimension questions lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 2 columns.
}
\source{
\url{https://www.aqol.com.au/index.php/scoring-algorithms}
}
\usage{
aqol6d_domain_qs_lup_tb
}
\description{
Breakdown of which questions relate to which dimension of the AQoL6D.
}
\details{
A tibble

\describe{
\item{Question_dbl}{Question (a double vector)}
\item{Domain_chr}{Domain (a character vector)}
}
}
\keyword{datasets}
