% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_aqol6d_from_8d_coeffs_lup_tb.R
\docType{data}
\name{aqol6d_from_8d_coeffs_lup_tb}
\alias{aqol6d_from_8d_coeffs_lup_tb}
\title{Model 2A Coefficients To Weight AQoL6D}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.
}
\source{
\url{https://www.aqol.com.au/index.php/scoring-algorithms}
}
\usage{
aqol6d_from_8d_coeffs_lup_tb
}
\description{
Coefficients for model to predict AQoL-6D utility score from AQoL-8D. The optimal model is Model 2A (see Richardson et al (2011, 18-19)*/
}
\details{
A tibble

\describe{
\item{var_name_chr}{Var name (a character vector)}
\item{coeff_dbl}{Coeff (a double vector)}
}
}
\keyword{datasets}
