% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author-TTUReports}
\alias{author-TTUReports}
\alias{author,TTUReports-method}
\title{Author and save files}
\usage{
\S4method{author}{TTUReports}(x, download_tmpl_1L_lgl = T, what_1L_chr = "Catalogue")
}
\arguments{
\item{x}{An object of class TTUReports}

\item{download_tmpl_1L_lgl}{Download template (a logical vector of length one), Default: T}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Catalogue'}
}
\description{
author method applied to TTUReports
}
