% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_adol_aqol6d}
\alias{calculate_adol_aqol6d}
\title{Calculate adolescent Assessment of Quality of Life Six Dimension health utility}
\usage{
calculate_adol_aqol6d(
  unscored_aqol_tb,
  prefix_1L_chr = "aqol",
  id_var_nm_1L_chr
)
}
\arguments{
\item{unscored_aqol_tb}{Unscored Assessment of Quality of Life health utility (a tibble)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'aqol'}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one)}
}
\value{
Adolescent Assessment of Quality of Life Six Dimension health utility (a double vector)
}
\description{
calculate_adol_aqol6d() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate adolescent assessment of quality of life six dimension health utility. The function returns Adolescent Assessment of Quality of Life Six Dimension health utility (a double vector).
}
\keyword{internal}
