% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_adult_aqol6dU}
\alias{calculate_adult_aqol6dU}
\title{Calculate adult Assessment of Quality of Life Six Dimension Health Utility}
\usage{
calculate_adult_aqol6dU(
  aqol6d_items_tb,
  prefix_1L_chr,
  coeffs_lup_tb = aqol6d_from_8d_coeffs_lup_tb,
  dim_sclg_con_lup_tb = aqol6d_dim_sclg_con_lup_tb,
  disvalues_lup_tb = aqol6d_adult_disv_lup_tb,
  itm_wrst_wghts_lup_tb = aqol6d_adult_itm_wrst_wghts_lup_tb
)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension items (a tibble)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}

\item{coeffs_lup_tb}{Coeffs lookup table (a tibble), Default: aqol6d_from_8d_coeffs_lup_tb}

\item{dim_sclg_con_lup_tb}{Dimension sclg constant lookup table (a tibble), Default: aqol6d_dim_sclg_con_lup_tb}

\item{disvalues_lup_tb}{Disvalues lookup table (a tibble), Default: aqol6d_adult_disv_lup_tb}

\item{itm_wrst_wghts_lup_tb}{Itm wrst wghts lookup table (a tibble), Default: aqol6d_adult_itm_wrst_wghts_lup_tb}
}
\value{
Assessment of Quality of Life Six Dimension Health Utility (a double vector)
}
\description{
calculate_adult_aqol6dU() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate adult assessment of quality of life six dimension health utility. The function returns Assessment of Quality of Life Six Dimension Health Utility (a double vector).
}
