% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_aqol6d_d3_disu_dbl}
\alias{calculate_aqol6d_d3_disu_dbl}
\title{Calculate aqol6d d3 disu}
\usage{
calculate_aqol6d_d3_disu_dbl(dvQs_tb, kD_1L_dbl, w_dbl)
}
\arguments{
\item{dvQs_tb}{DvQs (a tibble)}

\item{kD_1L_dbl}{KD 1L (a double vector of length 1)}

\item{w_dbl}{W (a double vector of length 1)}
}
\value{
DvD3 (a double vector of length 1)
}
\description{
calculate_aqol6d_d3_disu_dbl() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate aqol6d d3 disu. The function returns dvd3 (a double vector of length 1).
}
\keyword{internal}
