% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_aqol6d_d4_disu_dbl}
\alias{calculate_aqol6d_d4_disu_dbl}
\title{Calculate aqol6d d4 disu}
\usage{
calculate_aqol6d_d4_disu_dbl(dvQs_tb, kD_1L_dbl, w_dbl)
}
\arguments{
\item{dvQs_tb}{DvQs (a tibble)}

\item{kD_1L_dbl}{KD (a double vector of length one)}

\item{w_dbl}{W (a double vector)}
}
\value{
DvD4 (a double vector)
}
\description{
calculate_aqol6d_d4_disu_dbl() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate aqol6d d4 disu double vector. The function returns DvD4 (a double vector).
}
\keyword{internal}
