% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_aqol6d_dim_3_disv}
\alias{calculate_aqol6d_dim_3_disv}
\title{Calculate Assessment of Quality of Life Six Dimension dimension 3 disvalue}
\usage{
calculate_aqol6d_dim_3_disv(dvQs_tb, kD_1L_dbl, w_dbl)
}
\arguments{
\item{dvQs_tb}{DvQs (a tibble)}

\item{kD_1L_dbl}{KD (a double vector of length one)}

\item{w_dbl}{W (a double vector)}
}
\value{
DvD3 (a double vector)
}
\description{
calculate_aqol6d_dim_3_disv() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate assessment of quality of life six dimension dimension 3 disvalue. The function returns DvD3 (a double vector).
}
