% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_dep_var_tfmn}
\alias{calculate_dep_var_tfmn}
\title{Calculate dep var transformation}
\usage{
calculate_dep_var_tfmn(
  dep_var_val_dbl,
  tfmn_1L_chr = "NTF",
  tfmn_is_outp_1L_lgl = F
)
}
\arguments{
\item{dep_var_val_dbl}{Dep var value (a double vector)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{tfmn_is_outp_1L_lgl}{Transformation is output (a logical vector of length one), Default: F}
}
\value{
Transformed dep var value (a double vector)
}
\description{
calculate_dep_var_tfmn() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate dep var transformation. Function argument dep_var_val_dbl specifies the calculate The function returns Transformed dep var value (a double vector).
}
\keyword{internal}
