% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_depnt_var_tfmn}
\alias{calculate_depnt_var_tfmn}
\title{Calculate dependent variable transformation}
\usage{
calculate_depnt_var_tfmn(
  depnt_var_val_dbl,
  tfmn_1L_chr = "NTF",
  tfmn_is_outp_1L_lgl = F,
  depnt_var_max_val_1L_dbl = NULL
)
}
\arguments{
\item{depnt_var_val_dbl}{Dependent variable value (a double vector)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{tfmn_is_outp_1L_lgl}{Transformation is output (a logical vector of length one), Default: F}

\item{depnt_var_max_val_1L_dbl}{Dependent variable maximum value (a double vector of length one), Default: NULL}
}
\value{
Transformed dependent variable value (a double vector)
}
\description{
calculate_depnt_var_tfmn() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate dependent variable transformation. The function returns Transformed dependent variable value (a double vector).
}
\keyword{internal}
