% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{Calculate rmse}
\usage{
calculate_rmse(y_dbl, yhat_dbl)
}
\arguments{
\item{y_dbl}{Y (a double vector)}

\item{yhat_dbl}{Yhat (a double vector)}
}
\value{
Rmse (a double vector)
}
\description{
calculate_rmse() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate rmse. Function argument y_dbl specifies the calculate The function returns Rmse (a double vector).
}
