% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_rmse_tfmn}
\alias{calculate_rmse_tfmn}
\title{Calculate root mean square error transformation}
\usage{
calculate_rmse_tfmn(y_dbl, yhat_dbl)
}
\arguments{
\item{y_dbl}{Y (a double vector)}

\item{yhat_dbl}{Yhat (a double vector)}
}
\value{
Root mean square error transformation (a double vector)
}
\description{
calculate_rmse_tfmn() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate root mean square error transformation. The function returns Root mean square error transformation (a double vector).
}
