% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_candidate_predrs_lup.R
\docType{data}
\name{candidate_predrs_lup}
\alias{candidate_predrs_lup}
\title{Candidate predictors lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 5 columns.
}
\usage{
candidate_predrs_lup
}
\description{
A lookup table of the short name and long name of each candidate predictor.
}
\details{
A tibble

\describe{
\item{short_name_chr}{Short name (a character vector)}
\item{long_name_chr}{Long name (a character vector)}
\item{min_val_dbl}{Min value (a double vector)}
\item{max_val_dbl}{Max value (a double vector)}
\item{increment_dbl}{Increment (a double vector)}
}
}
\keyword{datasets}
