% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exhibit.R
\name{exhibit-TTUProject}
\alias{exhibit-TTUProject}
\alias{exhibit,TTUProject-method}
\title{Exhibit features of a dataset by printing them to the R console}
\usage{
\S4method{exhibit}{TTUProject}(x, what_1L_chr = "predictors", ...)
}
\arguments{
\item{x}{An object of class TTUProject}

\item{what_1L_chr}{What (a character vector of length one), Default: 'predictors'}

\item{...}{Additional arguments}
}
\description{
exhibit method applied to TTUProject
}
