% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_extract.R
\name{extract_guide_box_lgd}
\alias{extract_guide_box_lgd}
\title{Extract guide box legend}
\usage{
extract_guide_box_lgd(a.gplot)
}
\arguments{
\item{a.gplot}{PARAM_DESCRIPTION}
}
\value{
NA ()
}
\description{
extract_guide_box_lgd() is an Extract function that extracts data from an object. Specifically, this function implements an algorithm to extract guide box legend. The function is called for its side effects and does not return a value.
}
