% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_FBaqol.R
\docType{package}
\name{FBaqol-package}
\alias{FBaqol}
\alias{FBaqol-package}
\title{FBaqol: Map measures routinely collected in youth mental
health services to AQOL 6D Health Utility.}
\description{
Tools for mapping measures routinely collected in
youth mental health services to AQOL 6D Health Utility. Part of the
First Bounce model of primary youth mental health services.
}
\details{
To learn more about FBaqol, start with the vignettes:
\code{browseVignettes(package = "FBaqol")}
}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
