% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_aqol6d_adol.R
\name{firstbounce_aqol6d_adol}
\alias{firstbounce_aqol6d_adol}
\title{First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))}
\usage{
firstbounce_aqol6d_adol(x = make_prototype_firstbounce_aqol6d_adol())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent)), Default: make_prototype_firstbounce_aqol6d_adol()}
}
\value{
A validated instance of the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\description{
Create a new valid instance of the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\details{
First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
