% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_bads.R
\name{firstbounce_bads}
\alias{firstbounce_bads}
\title{First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
firstbounce_bads(x = make_prototype_firstbounce_bads())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores, Default: make_prototype_firstbounce_bads()}
}
\value{
A validated instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Create a new valid instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
