% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_gad7.R
\name{firstbounce_gad7}
\alias{firstbounce_gad7}
\title{First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores}
\usage{
firstbounce_gad7(x = make_prototype_firstbounce_gad7())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores, Default: make_prototype_firstbounce_gad7()}
}
\value{
A validated instance of the First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\description{
Create a new valid instance of the First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\details{
First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
