% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_k6.R
\name{firstbounce_k6}
\alias{firstbounce_k6}
\title{First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores}
\usage{
firstbounce_k6(x = make_prototype_firstbounce_k6())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores, Default: make_prototype_firstbounce_k6()}
}
\value{
A validated instance of the First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\description{
Create a new valid instance of the First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\details{
First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
