% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_oasis.R
\name{firstbounce_oasis}
\alias{firstbounce_oasis}
\title{First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores}
\usage{
firstbounce_oasis(x = make_prototype_firstbounce_oasis())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores, Default: make_prototype_firstbounce_oasis()}
}
\value{
A validated instance of the First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\description{
Create a new valid instance of the First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
\details{
First Bounce S3 class for Overall Anxiety Severity and Impairment Scale (OASIS) scores
}
