% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_phq9.R
\name{firstbounce_phq9}
\alias{firstbounce_phq9}
\title{First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores}
\usage{
firstbounce_phq9(x = make_prototype_firstbounce_phq9())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores, Default: make_prototype_firstbounce_phq9()}
}
\value{
A validated instance of the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\description{
Create a new valid instance of the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\details{
First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
