% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_scared.R
\name{firstbounce_scared}
\alias{firstbounce_scared}
\title{First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
\usage{
firstbounce_scared(x = make_prototype_firstbounce_scared())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores, Default: make_prototype_firstbounce_scared()}
}
\value{
A validated instance of the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\description{
Create a new valid instance of the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\details{
First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
