% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fit.R
\name{fit_gsn_log_lnk}
\alias{fit_gsn_log_lnk}
\title{Fit gaussian log lnk}
\usage{
fit_gsn_log_lnk(
  data_tb,
  depnt_var_nm_1L_chr = "utl_total_w",
  predr_vars_nms_chr,
  id_var_nm_1L_chr = "fkClientID",
  backend_1L_chr = getOption("brms.backend", "rstan"),
  iters_1L_int = 4000L,
  seed_1L_int = 1000L
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{predr_vars_nms_chr}{Predictor variables names (a character vector)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: getOption("brms.backend", "rstan")}

\item{iters_1L_int}{Iterations (an integer vector of length one), Default: 4000}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1000}
}
\value{
Model list (a list of models)
}
\description{
fit_gsn_log_lnk() is a Fit function that fits a model of a specified type to a dataset Specifically, this function implements an algorithm to fit gaussian log lnk. The function returns Model list (a list of models).
}
