% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_force.R
\name{force_min_max_and_int_cnstrs}
\alias{force_min_max_and_int_cnstrs}
\title{Force minimum maximum and integer vector constraints}
\usage{
force_min_max_and_int_cnstrs(tbl_tb, var_names_chr, min_max_ls, discrete_lgl)
}
\arguments{
\item{tbl_tb}{Table (a tibble)}

\item{var_names_chr}{Variable names (a character vector)}

\item{min_max_ls}{Minimum maximum (a list)}

\item{discrete_lgl}{Discrete (a logical vector)}
}
\value{
Table (a tibble)
}
\description{
force_min_max_and_int_cnstrs() is a Force function that checks if a specified local or global environmental condition is met and if not, updates the specified environment to comply with the condition. Specifically, this function implements an algorithm to force minimum maximum and integer vector constraints. The function returns Table (a tibble).
}
\keyword{internal}
