% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_force.R
\name{force_vec_to_sum_to_int}
\alias{force_vec_to_sum_to_int}
\title{Force vec to sum to}
\usage{
force_vec_to_sum_to_int(vec_int, target_1L_int, item_ranges_dbl_ls)
}
\arguments{
\item{vec_int}{Vec (an integer vector of length 1)}

\item{target_1L_int}{Target 1L (an integer vector of length 1)}

\item{item_ranges_dbl_ls}{Item ranges (a list of double vectors of length 1)}
}
\value{
Vec (an integer vector of length 1)
}
\description{
force_vec_to_sum_to_int() is a Force function that checks if a specified local or global environmental condition is met and if not, updates the specified environment to comply with the condition. Specifically, this function implements an algorithm to force vec to sum to. The function returns vec (an integer vector of length 1).
}
\keyword{internal}
