% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_background_text}
\alias{get_background_text}
\title{Get background text}
\usage{
get_background_text(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Text (a character vector of length one)
}
\description{
get_background_text() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get background text. Function argument results_ls specifies the where to look for the required object. The function returns Text (a character vector of length one).
}
