% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_hlth_utl_nm}
\alias{get_hlth_utl_nm}
\title{Get health utility name}
\usage{
get_hlth_utl_nm(results_ls, short_nm_1L_lgl = T)
}
\arguments{
\item{results_ls}{Results (a list)}

\item{short_nm_1L_lgl}{Short name (a logical vector of length one), Default: T}
}
\value{
Health utility name (a character vector of length one)
}
\description{
get_hlth_utl_nm() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get health utility name. Function argument results_ls specifies the where to look for the required object. The function returns Health utility name (a character vector of length one).
}
\keyword{internal}
