% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_hlth_utl_stat}
\alias{get_hlth_utl_stat}
\title{Get health utility statistic}
\usage{
get_hlth_utl_stat(results_ls, stat_1L_chr = "bl_mean")
}
\arguments{
\item{results_ls}{Results (a list)}

\item{stat_1L_chr}{Statistic (a character vector of length one), Default: 'bl_mean'}
}
\value{
Health utility statistic (a character vector of length one)
}
\description{
get_hlth_utl_stat() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get health utility statistic. Function argument results_ls specifies the where to look for the required object. The function returns Health utility statistic (a character vector of length one).
}
\keyword{internal}
