% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_link_from_tfmn}
\alias{get_link_from_tfmn}
\title{Get link from transformation}
\usage{
get_link_from_tfmn(tfmn_1L_chr, is_OLS_1L_lgl = F)
}
\arguments{
\item{tfmn_1L_chr}{Transformation (a character vector of length one)}

\item{is_OLS_1L_lgl}{Is ordinary least squares (a logical vector of length one), Default: F}
}
\value{
Link (a character vector of length one)
}
\description{
get_link_from_tfmn() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get link from transformation. Function argument tfmn_1L_chr specifies the where to look for the required object. The function returns Link (a character vector of length one).
}
\keyword{internal}
