% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdl_type_from_nm}
\alias{get_mdl_type_from_nm}
\title{Get model type from name}
\usage{
get_mdl_type_from_nm(mdl_nm_1L_chr, mdl_types_lup = NULL)
}
\arguments{
\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}
}
\value{
Model type (a character vector of length one)
}
\description{
get_mdl_type_from_nm() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get model type from name. Function argument mdl_nm_1L_chr specifies the where to look for the required object. The function returns Model type (a character vector of length one).
}
\keyword{internal}
