% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdls_with_signft_covars}
\alias{get_mdls_with_signft_covars}
\title{Get models with significant covariates}
\usage{
get_mdls_with_signft_covars(outp_smry_ls, params_ls_ls)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{params_ls_ls}{Parameters (a list of lists)}
}
\value{
Models with significant covariates (a list)
}
\description{
get_mdls_with_signft_covars() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get models with significant covariates. Function argument outp_smry_ls specifies the where to look for the required object. The function returns Models with significant covariates (a list).
}
\keyword{internal}
