% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_pkg_citation}
\alias{get_pkg_citation}
\title{Get package citation}
\usage{
get_pkg_citation(pkg_nm_1L_chr)
}
\arguments{
\item{pkg_nm_1L_chr}{Package name (a character vector of length one)}
}
\value{
Citation (a character vector of length one)
}
\description{
get_pkg_citation() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get package citation. The function returns Citation (a character vector of length one).
}
\keyword{internal}
