% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_popl_descvs}
\alias{get_popl_descvs}
\title{Get population descriptives}
\usage{
get_popl_descvs(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Population descriptives (a character vector of length one)
}
\description{
get_popl_descvs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get population descriptives. Function argument results_ls specifies the where to look for the required object. The function returns Population descriptives (a character vector of length one).
}
\keyword{internal}
