% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_predrs_by_ctg}
\alias{get_predrs_by_ctg}
\title{Get predictors by category categories}
\usage{
get_predrs_by_ctg(
  results_ls,
  long_desc_1L_lgl = F,
  transform_1L_lgl = F,
  collapse_1L_lgl = F
)
}
\arguments{
\item{results_ls}{Results (a list)}

\item{long_desc_1L_lgl}{Long description (a logical vector of length one), Default: F}

\item{transform_1L_lgl}{Transform (a logical vector of length one), Default: F}

\item{collapse_1L_lgl}{Collapse (a logical vector of length one), Default: F}
}
\value{
Predictors by category categories (a list)
}
\description{
get_predrs_by_ctg() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get predictors by category categories. Function argument results_ls specifies the where to look for the required object. The function returns Predictors by category categories (a list).
}
\keyword{internal}
