% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_prefd_mdl_predrs}
\alias{get_prefd_mdl_predrs}
\title{Get preferred model predictors}
\usage{
get_prefd_mdl_predrs(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Predictors (a character vector of length one)
}
\description{
get_prefd_mdl_predrs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get preferred model predictors. Function argument results_ls specifies the where to look for the required object. The function returns Predictors (a character vector of length one).
}
