% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_random_intercept}
\alias{get_random_intercept}
\title{Get random intercept}
\usage{
get_random_intercept(mdls_smry_tb, mdl_nm_1L_chr, deterministic_1L_lgl = T)
}
\arguments{
\item{mdls_smry_tb}{Models summary (a tibble)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{deterministic_1L_lgl}{Deterministic (a logical vector of length one), Default: T}
}
\value{
Standard deviation (a double vector)
}
\description{
get_random_intercept() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get random intercept. Function argument mdls_smry_tb specifies the where to look for the required object. The function returns Standard deviation (a double vector).
}
\keyword{internal}
