% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_scndry_anlys_descs}
\alias{get_scndry_anlys_descs}
\title{Get secondary analysis descriptions}
\usage{
get_scndry_anlys_descs(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Secondary analysis descriptions (a character vector)
}
\description{
get_scndry_anlys_descs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get secondary analysis descriptions. Function argument results_ls specifies the where to look for the required object. The function returns Secondary analysis descriptions (a character vector).
}
\keyword{internal}
