% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_selected_mixed_mdls}
\alias{get_selected_mixed_mdls}
\title{Get selected mixed models}
\usage{
get_selected_mixed_mdls(results_ls, collapse_1L_lgl = T)
}
\arguments{
\item{results_ls}{Results (a list)}

\item{collapse_1L_lgl}{Collapse (a logical vector of length one), Default: T}
}
\value{
Mixed models (an output object of multiple potential types)
}
\description{
get_selected_mixed_mdls() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get selected mixed models. Function argument results_ls specifies the where to look for the required object. The function returns Mixed models (an output object of multiple potential types).
}
\keyword{internal}
