% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_signft_covars}
\alias{get_signft_covars}
\title{Get signft covars}
\usage{
get_signft_covars(mdls_with_covars_smry_tb, covar_var_nms_chr)
}
\arguments{
\item{mdls_with_covars_smry_tb}{Mdls with covars smry (a tibble)}

\item{covar_var_nms_chr}{Covar var names (a character vector)}
}
\value{
Signt covars (a character vector)
}
\description{
get_signft_covars() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get signft covars. Function argument mdls_with_covars_smry_tb specifies the where to look for the required object. The function returns Signt covars (a character vector).
}
