% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_table_predn_mdl}
\alias{get_table_predn_mdl}
\title{Get table prediction model}
\usage{
get_table_predn_mdl(mdl_nm_1L_chr, ingredients_ls, analysis_1L_chr = NULL)
}
\arguments{
\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{ingredients_ls}{Ingredients (a list)}

\item{analysis_1L_chr}{Analysis (a character vector of length one), Default: NULL}
}
\value{
Table prediction (a model)
}
\description{
get_table_predn_mdl() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get table prediction model. Function argument mdl_nm_1L_chr specifies the where to look for the required object. The function returns Table prediction (a model).
}
\keyword{internal}
