% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_impute.R
\name{impute_adol_unscrd_aqol_ds}
\alias{impute_adol_unscrd_aqol_ds}
\title{Impute adolescent unscored Assessment of Quality of Life health utility dataset}
\usage{
impute_adol_unscrd_aqol_ds(unscrd_aqol_ds_tb)
}
\arguments{
\item{unscrd_aqol_ds_tb}{Unscored Assessment of Quality of Life health utility dataset (a tibble)}
}
\value{
Imputed unscored Assessment of Quality of Life health utility dataset tibble (a tibble)
}
\description{
impute_adol_unscrd_aqol_ds() is an Impute function that imputes data. Specifically, this function implements an algorithm to impute adolescent unscored assessment of quality of life health utility dataset. The function returns Imputed unscored Assessment of Quality of Life health utility dataset tibble (a tibble).
}
\keyword{internal}
