% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_impute.R
\name{impute_adult_aqol6d_items_tb}
\alias{impute_adult_aqol6d_items_tb}
\title{Impute adult Assessment of Quality of Life Six Dimension items}
\usage{
impute_adult_aqol6d_items_tb(aqol6d_items_tb, domain_items_ls)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension items (a tibble)}

\item{domain_items_ls}{Domain items (a list)}
}
\value{
Assessment of Quality of Life Six Dimension items (a tibble)
}
\description{
impute_adult_aqol6d_items_tb() is an Impute function that imputes data. Specifically, this function implements an algorithm to impute adult assessment of quality of life six dimension items tibble. The function returns Assessment of Quality of Life Six Dimension items (a tibble).
}
\keyword{internal}
