% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_impute.R
\name{impute_unscrd_adol_aqol6d_ds}
\alias{impute_unscrd_adol_aqol6d_ds}
\title{Impute unscored adolescent Assessment of Quality of Life Six Dimension dataset}
\usage{
impute_unscrd_adol_aqol6d_ds(unscrd_aqol_ds_tb)
}
\arguments{
\item{unscrd_aqol_ds_tb}{Unscored Assessment of Quality of Life dataset (a tibble)}
}
\value{
Imputed unscored Assessment of Quality of Life dataset tibble (a tibble)
}
\description{
impute_unscrd_adol_aqol6d_ds() is an Impute function that imputes data. Specifically, this function implements an algorithm to impute unscored adolescent assessment of quality of life six dimension dataset. The function returns Imputed unscored Assessment of Quality of Life dataset tibble (a tibble).
}
\keyword{internal}
