% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_investigate.R
\name{investigate-TTUProject}
\alias{investigate-TTUProject}
\alias{investigate,TTUProject-method}
\title{Investigate solutions to an inverse problem}
\usage{
\S4method{investigate}{TTUProject}(
  x,
  backend_1L_chr = "cmdstanr",
  combinations_1L_lgl = F,
  consent_1L_chr = "",
  cores_1L_int = 1L,
  depnt_var_max_val_1L_dbl = numeric(0),
  depnt_var_min_val_1L_dbl = numeric(0),
  existing_predrs_ls = NULL,
  max_nbr_of_covars_1L_int = integer(0),
  new_dir_nm_1L_chr = "F_TS_Mdls",
  scndry_anlys_params_ls = NULL,
  session_ls = NULL,
  signft_covars_cdn_1L_chr = "any",
  ...
)
}
\arguments{
\item{x}{An object of class TTUProject}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: 'cmdstanr'}

\item{combinations_1L_lgl}{Combinations (a logical vector of length one), Default: F}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{cores_1L_int}{Cores (an integer vector of length one), Default: 1}

\item{depnt_var_max_val_1L_dbl}{Dependent variable maximum value (a double vector of length one), Default: numeric(0)}

\item{depnt_var_min_val_1L_dbl}{Dependent variable minimum value (a double vector of length one), Default: numeric(0)}

\item{existing_predrs_ls}{Existing predictors (a list), Default: NULL}

\item{max_nbr_of_covars_1L_int}{Maximum number of covariates (an integer vector of length one), Default: integer(0)}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'F_TS_Mdls'}

\item{scndry_anlys_params_ls}{Secondary analysis parameters (a list), Default: NULL}

\item{session_ls}{Session (a list), Default: NULL}

\item{signft_covars_cdn_1L_chr}{Significant covariates condition (a character vector of length one), Default: 'any'}

\item{...}{Additional arguments}
}
\value{
x (An object of class TTUProject)
}
\description{
investigate method applied to TTUProject
}
