% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_aqol6d_adol.R
\name{is_firstbounce_aqol6d_adol}
\alias{is_firstbounce_aqol6d_adol}
\title{Is First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))}
\usage{
is_firstbounce_aqol6d_adol(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\description{
Check whether an object is a valid instance of the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\details{
First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
