% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_bads.R
\name{is_firstbounce_bads}
\alias{is_firstbounce_bads}
\title{Is First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
is_firstbounce_bads(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Check whether an object is a valid instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
