% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_k6.R
\name{is_firstbounce_k6}
\alias{is_firstbounce_k6}
\title{Is First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores}
\usage{
is_firstbounce_k6(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\description{
Check whether an object is a valid instance of the First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\details{
First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
