% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_phq9.R
\name{is_firstbounce_phq9}
\alias{is_firstbounce_phq9}
\title{Is First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores}
\usage{
is_firstbounce_phq9(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\description{
Check whether an object is a valid instance of the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\details{
First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
