% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_scared.R
\name{is_firstbounce_scared}
\alias{is_firstbounce_scared}
\title{Is First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
\usage{
is_firstbounce_scared(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\description{
Check whether an object is a valid instance of the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\details{
First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
