% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_TTU_predictors_lup.R
\name{is_TTU_predictors_lup}
\alias{is_TTU_predictors_lup}
\title{Is TTU predictors lookup table TTU S3 class for candidate predictors lookup table}
\usage{
is_TTU_predictors_lup(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the TTU S3 class for candidate predictors lookup table
}
\description{
Check whether an object is a valid instance of the TTU S3 class for candidate predictors lookup table
}
\details{
TTU S3 class for candidate predictors lookup table
}
