% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_knit.R
\name{knit_from_tmpl}
\alias{knit_from_tmpl}
\title{Knit from template}
\usage{
knit_from_tmpl(params_to_expand_ls, path_to_tmpl_1L_chr)
}
\arguments{
\item{params_to_expand_ls}{Parameters to expand (a list)}

\item{path_to_tmpl_1L_chr}{Path to template (a character vector of length one)}
}
\description{
knit_from_tmpl() is a Knit function that knits a rmarkdown file Specifically, this function implements an algorithm to knit from template. The function is called for its side effects and does not return a value.
}
\keyword{internal}
