% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_knit.R
\name{knit_mdl_rprt}
\alias{knit_mdl_rprt}
\title{Knit mdl rprt}
\usage{
knit_mdl_rprt(
  knit_pars_ls,
  path_to_mdl_rprt_tmpl_1L_chr = system.file("_Model_Report_Template.Rmd", package =
    "ready4u")
)
}
\arguments{
\item{knit_pars_ls}{Knit parameters (a list)}

\item{path_to_mdl_rprt_tmpl_1L_chr}{Path to mdl rprt tmpl (a character vector of length one), Default: system.file("_Model_Report_Template.Rmd", package = "ready4u")}
}
\description{
knit_mdl_rprt() is a Knit function that knits a rmarkdown file Specifically, this function implements an algorithm to knit mdl rprt. The function is called for its side effects and does not return a value.
}
\keyword{internal}
