% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_abstract_args_ls}
\alias{make_abstract_args_ls}
\title{Make abstract arguments list}
\usage{
make_abstract_args_ls(results_ls, fl_nm_1L_chr = "abstract.txt")
}
\arguments{
\item{results_ls}{Results (a list)}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'abstract.txt'}
}
\value{
Abstract arguments (a list)
}
\description{
make_abstract_args_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make abstract arguments list. The function returns Abstract arguments (a list).
}
\keyword{internal}
