% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_analysis_ds_smry_ls}
\alias{make_analysis_ds_smry_ls}
\title{Make analysis dataset summary list}
\usage{
make_analysis_ds_smry_ls(ds_descvs_ls, candidate_covar_nms_chr, predictors_lup)
}
\arguments{
\item{ds_descvs_ls}{Dataset descriptives (a list)}

\item{candidate_covar_nms_chr}{Candidate covariate names (a character vector)}

\item{predictors_lup}{Predictors (a lookup table)}
}
\value{
Analysis dataset summary (a list)
}
\description{
make_analysis_ds_smry_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make analysis dataset summary list. The function returns Analysis dataset summary (a list).
}
