% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol6d_fns_ls}
\alias{make_aqol6d_fns_ls}
\title{Make aqol6d functions}
\usage{
make_aqol6d_fns_ls(domain_items_ls)
}
\arguments{
\item{domain_items_ls}{Domain items (a list)}
}
\value{
Aqol6d disu (a list of functions)
}
\description{
make_aqol6d_fns_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make aqol6d a functions. The function returns aqol6d disu (a list of functions).
}
\keyword{internal}
