% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_bl_fup_add_to_row_ls}
\alias{make_bl_fup_add_to_row_ls}
\title{Make baseline follow-up add to row list}
\usage{
make_bl_fup_add_to_row_ls(df, n_at_bl_1L_int, n_at_fup_1L_int)
}
\arguments{
\item{df}{Data.frame (a data.frame)}

\item{n_at_bl_1L_int}{N at baseline (an integer vector of length one)}

\item{n_at_fup_1L_int}{N at follow-up (an integer vector of length one)}
}
\value{
Add to row (a list)
}
\description{
make_bl_fup_add_to_row_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make baseline follow-up add to row list. The function returns Add to row (a list).
}
\keyword{internal}
