% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_brms_mdl_print_ls}
\alias{make_brms_mdl_print_ls}
\title{Make bayesian regression models model print list}
\usage{
make_brms_mdl_print_ls(
  mdl_ls,
  label_stub_1L_chr,
  caption_1L_chr,
  output_type_1L_chr = "PDF",
  digits_1L_dbl = 2,
  big_mark_1L_chr = " "
)
}
\arguments{
\item{mdl_ls}{Model list (a list of models)}

\item{label_stub_1L_chr}{Label stub (a character vector of length one)}

\item{caption_1L_chr}{Caption (a character vector of length one)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'PDF'}

\item{digits_1L_dbl}{Digits (a double vector of length one), Default: 2}

\item{big_mark_1L_chr}{Big mark (a character vector of length one), Default: ' '}
}
\value{
Bayesian regression models model print (a list)
}
\description{
make_brms_mdl_print_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make bayesian regression models model print list. The function returns Bayesian regression models model print (a list).
}
\keyword{internal}
