% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_brms_mdl_smry_tbl}
\alias{make_brms_mdl_smry_tbl}
\title{Make brms mdl smry table}
\usage{
make_brms_mdl_smry_tbl(smry_mdl_ls, grp_1L_chr, pop_1L_chr, fam_1L_chr)
}
\arguments{
\item{smry_mdl_ls}{Smry mdl (a list)}

\item{grp_1L_chr}{Grp (a character vector of length one)}

\item{pop_1L_chr}{Pop (a character vector of length one)}

\item{fam_1L_chr}{Fam (a character vector of length one)}
}
\value{
Brms mdl smry (a tibble)
}
\description{
make_brms_mdl_smry_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make brms mdl smry table. The function returns Brms mdl smry (a tibble).
}
