% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cmpst_sctr_and_dnsty_plt}
\alias{make_cmpst_sctr_and_dnsty_plt}
\title{Make cmpst scatter and dnsty}
\usage{
make_cmpst_sctr_and_dnsty_plt(
  outp_smry_ls,
  predr_var_nms_chr,
  labels_chr = c("A", "B", "C", "D"),
  label_x_1L_dbl = 0.1,
  label_y_1L_dbl = 0.9,
  label_size_1L_dbl = 22
)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{predr_var_nms_chr}{Predictor variable names (a character vector)}

\item{labels_chr}{Labels (a character vector), Default: c("A", "B", "C", "D")}

\item{label_x_1L_dbl}{Label x (a double vector of length one), Default: 0.1}

\item{label_y_1L_dbl}{Label y (a double vector of length one), Default: 0.9}

\item{label_size_1L_dbl}{Label size (a double vector of length one), Default: 22}
}
\description{
make_cmpst_sctr_and_dnsty_plt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cmpst scatter and dnsty plot. The function is called for its side effects and does not return a value.
}
\keyword{internal}
