% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cndt_predr_text}
\alias{make_cndt_predr_text}
\title{Make candidate predictor text}
\usage{
make_cndt_predr_text(results_ls, type_1L_chr = "description")
}
\arguments{
\item{results_ls}{Results (a list)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'description'}
}
\value{
Text (a character vector of length one)
}
\description{
make_cndt_predr_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make candidate predictor text. The function returns Text (a character vector of length one).
}
