% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cohort_ls}
\alias{make_cohort_ls}
\title{Make cohort list}
\usage{
make_cohort_ls(
  descv_tbls_ls,
  ctgl_vars_regrouping_ls = NULL,
  nbr_of_digits_1L_int = 2L
)
}
\arguments{
\item{descv_tbls_ls}{Descriptive tables (a list)}

\item{ctgl_vars_regrouping_ls}{Categorical variables regrouping (a list), Default: NULL}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}
}
\value{
Cohort (a list)
}
\description{
make_cohort_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cohort list. The function returns Cohort (a list).
}
\keyword{internal}
