% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_complete_props_tbs_ls}
\alias{make_complete_props_tbs_ls}
\title{Make complete props tibbles}
\usage{
make_complete_props_tbs_ls(
  raw_props_tbs_ls,
  question_var_nm_1L_chr = "Question"
)
}
\arguments{
\item{raw_props_tbs_ls}{Raw props tibbles (a list)}

\item{question_var_nm_1L_chr}{Question var name (a character vector of length one), Default: 'Question'}
}
\value{
Complete props tibbles (a list)
}
\description{
make_complete_props_tbs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make complete props tibbles list. The function returns Complete props tibbles (a list).
}
\keyword{internal}
