% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_correlated_data_tb}
\alias{make_correlated_data_tb}
\title{Make correlated data}
\usage{
make_correlated_data_tb(synth_data_spine_ls, synth_data_idx_1L_dbl = 1)
}
\arguments{
\item{synth_data_spine_ls}{Synthetic data spine (a list)}

\item{synth_data_idx_1L_dbl}{Synthetic data index (a double vector of length one), Default: 1}
}
\value{
Correlated data (a tibble)
}
\description{
make_correlated_data_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make correlated data tibble. The function returns Correlated data (a tibble).
}
\keyword{internal}
