% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_correlation_text}
\alias{make_correlation_text}
\title{Make correlation text}
\usage{
make_correlation_text(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Correlation text (a character vector of length one)
}
\description{
make_correlation_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make correlation text. The function returns Correlation text (a character vector of length one).
}
