% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_corstars_tbl_xx}
\alias{make_corstars_tbl_xx}
\title{Make corstars table}
\usage{
make_corstars_tbl_xx(
  x,
  method = c("pearson", "spearman"),
  removeTriangle = c("upper", "lower"),
  result = c("none", "html", "latex")
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: c("pearson", "spearman")}

\item{removeTriangle}{PARAM_DESCRIPTION, Default: c("upper", "lower")}

\item{result}{PARAM_DESCRIPTION, Default: c("none", "html", "latex")}
}
\description{
make_corstars_tbl_xx() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make corstars table output object of multiple potential types. The function is called for its side effects and does not return a value.
}
