% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_covar_ttu_tbl_refs}
\alias{make_covar_ttu_tbl_refs}
\title{Make covariate ttu table references}
\usage{
make_covar_ttu_tbl_refs(params_ls)
}
\arguments{
\item{params_ls}{Params (a list)}
}
\value{
Text (a character vector of length one)
}
\description{
make_covar_ttu_tbl_refs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make covariate ttu table references. The function returns Text (a character vector of length one).
}
