% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_covar_ttu_tbl_title}
\alias{make_covar_ttu_tbl_title}
\title{Make covariate transfer to utility algorithm table title}
\usage{
make_covar_ttu_tbl_title(results_ls, ref_1L_int = 1)
}
\arguments{
\item{results_ls}{Results (a list)}

\item{ref_1L_int}{Reference (an integer vector of length one), Default: 1}
}
\value{
Title (a character vector of length one)
}
\description{
make_covar_ttu_tbl_title() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make covariate transfer to utility algorithm table title. The function returns Title (a character vector of length one).
}
\keyword{internal}
