% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_dim_sclg_cons_dbl}
\alias{make_dim_sclg_cons_dbl}
\title{Make dim sclg cons}
\usage{
make_dim_sclg_cons_dbl(
  domains_chr,
  dim_sclg_constant_lup_tb = dim_sclg_constant_lup_tb
)
}
\arguments{
\item{domains_chr}{Domains (a character vector)}

\item{dim_sclg_constant_lup_tb}{Dim sclg constant lookup table (a tibble), Default: dim_sclg_constant_lup_tb}
}
\value{
Dim sclg cons (a double vector)
}
\description{
make_dim_sclg_cons_dbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dim sclg cons double vector. The function returns Dim sclg cons (a double vector).
}
\keyword{internal}
