% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_dnst_and_sctr_plt_title}
\alias{make_dnst_and_sctr_plt_title}
\title{Make density and scatter plot title}
\usage{
make_dnst_and_sctr_plt_title(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Title (a character vector of length one)
}
\description{
make_dnst_and_sctr_plt_title() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make density and scatter plot title. The function returns Title (a character vector of length one).
}
