% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ds_descvs_ls}
\alias{make_ds_descvs_ls}
\title{Make dataset descriptives}
\usage{
make_ds_descvs_ls(
  candidate_predrs_chr,
  cohort_descv_var_nms_chr,
  dictionary_tb,
  id_var_nm_1L_chr,
  msrmnt_date_var_nm_1L_chr,
  round_var_nm_1L_chr,
  round_vals_chr,
  maui_item_pfx_1L_chr,
  utl_wtd_var_nm_1L_chr = "wtd_utl_dbl",
  utl_unwtd_var_nm_1L_chr = "unwtd_utl_dbl"
)
}
\arguments{
\item{candidate_predrs_chr}{Candidate predictors (a character vector)}

\item{cohort_descv_var_nms_chr}{Cohort descriptive variable names (a character vector)}

\item{dictionary_tb}{Dictionary (a tibble)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one)}

\item{msrmnt_date_var_nm_1L_chr}{Measurement date variable name (a character vector of length one)}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one)}

\item{round_vals_chr}{Round values (a character vector)}

\item{maui_item_pfx_1L_chr}{Maui item prefix (a character vector of length one)}

\item{utl_wtd_var_nm_1L_chr}{Utility weighted variable name (a character vector of length one), Default: 'wtd_utl_dbl'}

\item{utl_unwtd_var_nm_1L_chr}{Utility unwtd variable name (a character vector of length one), Default: 'unwtd_utl_dbl'}
}
\value{
Dataset descriptives (a list)
}
\description{
make_ds_descvs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dataset descriptives list. The function returns Dataset descriptives (a list).
}
\keyword{internal}
