% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fake_eq5d_ds}
\alias{make_fake_eq5d_ds}
\title{Make fake eq5d dataset}
\usage{
make_fake_eq5d_ds(
  fl_nm_1L_chr = "eq5d5l_example.xlsx",
  country_1L_chr = "UK",
  version_1L_chr = "5L",
  type_1L_chr = "CW"
)
}
\arguments{
\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'eq5d5l_example.xlsx'}

\item{country_1L_chr}{Country (a character vector of length one), Default: 'UK'}

\item{version_1L_chr}{Version (a character vector of length one), Default: '5L'}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'CW'}
}
\value{
Data (a tibble)
}
\description{
make_fake_eq5d_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make fake eq5d dataset. The function returns Data (a tibble).
}
\keyword{internal}
