% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_folds_ls}
\alias{make_folds_ls}
\title{Make folds}
\usage{
make_folds_ls(
  data_tb,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  n_folds_1L_int = 10L
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{n_folds_1L_int}{N folds (an integer vector of length one), Default: 10}
}
\value{
Folds (a list)
}
\description{
make_folds_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make folds list. The function returns Folds (a list).
}
