% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_funding_text}
\alias{make_funding_text}
\title{Make funding text}
\usage{
make_funding_text(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Text (a character vector of length one)
}
\description{
make_funding_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make funding text. The function returns Text (a character vector of length one).
}
