% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_hlth_utl_and_predrs_ls}
\alias{make_hlth_utl_and_predrs_ls}
\title{Make health utility and predictors list}
\usage{
make_hlth_utl_and_predrs_ls(
  outp_smry_ls,
  descv_tbls_ls,
  nbr_of_digits_1L_int = 2L,
  old_nms_chr = NULL,
  new_nms_chr = NULL
)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{descv_tbls_ls}{Descriptive tables (a list)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}

\item{old_nms_chr}{Old names (a character vector), Default: NULL}

\item{new_nms_chr}{New names (a character vector), Default: NULL}
}
\value{
Health utility and predictors (a list)
}
\description{
make_hlth_utl_and_predrs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make health utility and predictors list. The function returns Health utility and predictors (a list).
}
\keyword{internal}
