% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_input_params}
\alias{make_input_params}
\title{Make input parameters}
\usage{
make_input_params(
  ds_tb,
  ds_descvs_ls,
  header_yaml_args_ls,
  maui_params_ls,
  predictors_lup,
  control_ls = NULL,
  dv_ds_nm_and_url_chr = NULL,
  iters_1L_int = 4000L,
  mdl_smry_ls = make_mdl_smry_ls(),
  output_format_ls = make_output_format_ls(),
  path_params_ls = NULL,
  prefd_covars_chr = NULL,
  prefd_mdl_types_chr = NULL,
  prior_ls = NULL,
  seed_1L_int = 12345,
  scndry_anlys_params_ls = NULL,
  write_new_dir_1L_lgl = T
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{ds_descvs_ls}{Dataset descriptives (a list)}

\item{header_yaml_args_ls}{Header yaml arguments (a list)}

\item{maui_params_ls}{Multi-attribute utility instrument parameters (a list)}

\item{predictors_lup}{Predictors (a lookup table)}

\item{control_ls}{Control (a list), Default: NULL}

\item{dv_ds_nm_and_url_chr}{Dataverse dataset name and url (a character vector), Default: NULL}

\item{iters_1L_int}{Iterations (an integer vector of length one), Default: 4000}

\item{mdl_smry_ls}{Model summary (a list), Default: make_mdl_smry_ls()}

\item{output_format_ls}{Output format (a list), Default: make_output_format_ls()}

\item{path_params_ls}{Path parameters (a list), Default: NULL}

\item{prefd_covars_chr}{Preferred covariates (a character vector), Default: NULL}

\item{prefd_mdl_types_chr}{Preferred model types (a character vector), Default: NULL}

\item{prior_ls}{Prior (a list), Default: NULL}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 12345}

\item{scndry_anlys_params_ls}{Secondary analysis parameters (a list), Default: NULL}

\item{write_new_dir_1L_lgl}{Write new directory (a logical vector of length one), Default: T}
}
\value{
Parameters (a list of lists)
}
\description{
make_input_params() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make input parameters. The function returns Parameters (a list of lists).
}
