% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_item_wrst_wghts_ls_ls}
\alias{make_item_wrst_wghts_ls_ls}
\title{Make item wrst wghts}
\usage{
make_item_wrst_wghts_ls_ls(domain_items_ls, itm_wrst_wghts_lup_tb)
}
\arguments{
\item{domain_items_ls}{Domain items (a list)}

\item{itm_wrst_wghts_lup_tb}{Itm wrst wghts lookup table (a tibble)}
}
\value{
Item wrst wghts (a list of lists)
}
\description{
make_item_wrst_wghts_ls_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make item wrst wghts list list. The function returns Item wrst wghts (a list of lists).
}
