% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_knit_pars_ls}
\alias{make_knit_pars_ls}
\title{Make knit parameters}
\usage{
make_knit_pars_ls(
  mdl_smry_dir_1L_chr,
  mdl_types_chr,
  predr_vars_nms_ls,
  output_type_1L_chr = "HTML",
  mdl_types_lup = NULL,
  plt_types_lup = NULL,
  plt_types_chr = c("coefs", "hetg", "dnst", "sctr_plt"),
  section_type_1L_chr = "#"
)
}
\arguments{
\item{mdl_smry_dir_1L_chr}{Mdl smry directory (a character vector of length one)}

\item{mdl_types_chr}{Mdl types (a character vector)}

\item{predr_vars_nms_ls}{Predr vars names (a list)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{mdl_types_lup}{Mdl types (a lookup table), Default: NULL}

\item{plt_types_lup}{Plt types (a lookup table), Default: NULL}

\item{plt_types_chr}{Plt types (a character vector), Default: c("coefs", "hetg", "dnst", "sctr_plt")}

\item{section_type_1L_chr}{Section type (a character vector of length one), Default: '#'}
}
\value{
Knit parameters (a list)
}
\description{
make_knit_pars_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make knit parameters list. The function returns Knit parameters (a list).
}
