% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_make_item_wrst_wts_ls_ls}
\alias{make_make_item_wrst_wts_ls_ls}
\title{Make make item worst weights}
\usage{
make_make_item_wrst_wts_ls_ls(domain_items_ls, itm_wrst_wghts_lup_tb)
}
\arguments{
\item{domain_items_ls}{Domain items (a list)}

\item{itm_wrst_wghts_lup_tb}{Item worst wghts lookup table (a tibble)}
}
\value{
Make item worst weights (a list of lists)
}
\description{
make_make_item_wrst_wts_ls_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make make item worst weights list list. The function returns Make item worst weights (a list of lists).
}
\keyword{internal}
