% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_coef_range_text}
\alias{make_mdl_coef_range_text}
\title{Make model coefficient range text}
\usage{
make_mdl_coef_range_text(coef_ratios_dbl, nbr_of_digits_1L_int = 2L)
}
\arguments{
\item{coef_ratios_dbl}{Coefficient ratios (a double vector)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}
}
\value{
Coefficient range text (a character vector)
}
\description{
make_mdl_coef_range_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model coefficient range text. The function returns Coefficient range text (a character vector).
}
\keyword{internal}
