% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_coef_ratio_ls}
\alias{make_mdl_coef_ratio_ls}
\title{Make model coefficient ratio list}
\usage{
make_mdl_coef_ratio_ls(mdl_ingredients_ls, predr_ctgs_ls = NULL)
}
\arguments{
\item{mdl_ingredients_ls}{Model ingredients (a list)}

\item{predr_ctgs_ls}{Predictor category categoriess (a list), Default: NULL}
}
\value{
Model coefficient ratios (a list)
}
\description{
make_mdl_coef_ratio_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model coefficient ratio list. The function returns Model coefficient ratios (a list).
}
\keyword{internal}
