% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_desc_lines}
\alias{make_mdl_desc_lines}
\title{Make model description lines}
\usage{
make_mdl_desc_lines(outp_smry_ls, mdl_nm_1L_chr, output_type_1L_chr = "PDF")
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'PDF'}
}
\value{
Model description lines (a character vector)
}
\description{
make_mdl_desc_lines() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model description lines. The function returns Model description lines (a character vector).
}
