% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_nms_ls}
\alias{make_mdl_nms_ls}
\title{Make mdl names}
\usage{
make_mdl_nms_ls(predictor_vars_nms_ls, mdl_types_chr)
}
\arguments{
\item{predictor_vars_nms_ls}{Predictor vars names (a list)}

\item{mdl_types_chr}{Mdl types (a character vector)}
}
\value{
Mdl names (a list)
}
\description{
make_mdl_nms_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make mdl names list. The function returns Mdl names (a list).
}
