% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_smry_elmt_tbl}
\alias{make_mdl_smry_elmt_tbl}
\title{Make model smry elmt table}
\usage{
make_mdl_smry_elmt_tbl(mat, cat_chr)
}
\arguments{
\item{mat}{Matrix (a matrix)}

\item{cat_chr}{Cat (a character vector)}
}
\value{
Model elmt sum (a tibble)
}
\description{
make_mdl_smry_elmt_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model smry elmt table. The function returns Model elmt sum (a tibble).
}
\keyword{internal}
