% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_smry_ls}
\alias{make_mdl_smry_ls}
\title{Make model summary list}
\usage{
make_mdl_smry_ls(
  mdl_types_lup = get_cndts_for_mxd_mdls(),
  mdl_types_chr = NULL,
  choose_from_pfx_chr = NULL,
  folds_1L_int = 10L,
  max_nbr_of_boruta_mdl_runs_int = 300L
)
}
\arguments{
\item{mdl_types_lup}{Model types (a lookup table), Default: get_cndts_for_mxd_mdls()}

\item{mdl_types_chr}{Model types (a character vector), Default: NULL}

\item{choose_from_pfx_chr}{Choose from prefix (a character vector), Default: NULL}

\item{folds_1L_int}{Folds (an integer vector of length one), Default: 10}

\item{max_nbr_of_boruta_mdl_runs_int}{Maximum number of boruta model runs (an integer vector), Default: 300}
}
\value{
Model summary (a list)
}
\description{
make_mdl_smry_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model summary list. The function returns Model summary (a list).
}
\keyword{internal}
