% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdls_ls}
\alias{make_mdls_ls}
\title{Make models list}
\usage{
make_mdls_ls(outp_smry_ls, mdls_tb)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{mdls_tb}{Models (a tibble)}
}
\value{
Models (a list)
}
\description{
make_mdls_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make models list. The function returns Models (a list).
}
\keyword{internal}
