% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdls_smry_tbls_ls}
\alias{make_mdls_smry_tbls_ls}
\title{Make models summary tables list}
\usage{
make_mdls_smry_tbls_ls(outp_smry_ls, nbr_of_digits_1L_int = 2L)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}
}
\value{
Models summary tables (a list)
}
\description{
make_mdls_smry_tbls_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make models summary tables list. The function returns Models summary tables (a list).
}
\keyword{internal}
