% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_nbr_at_fup_text}
\alias{make_nbr_at_fup_text}
\title{Make number at follow-up text}
\usage{
make_nbr_at_fup_text(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Number at follow-up (a character vector of length one)
}
\description{
make_nbr_at_fup_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make number at follow-up text. The function returns Number at follow-up (a character vector of length one).
}
\keyword{internal}
