% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_nbr_included_text}
\alias{make_nbr_included_text}
\title{Make number included text}
\usage{
make_nbr_included_text(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\description{
make_nbr_included_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make number included text. The function is called for its side effects and does not return a value.
}
