% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_aqol6d_adol.R
\name{make_new_firstbounce_aqol6d_adol}
\alias{make_new_firstbounce_aqol6d_adol}
\title{Make new First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))}
\usage{
make_new_firstbounce_aqol6d_adol(x)
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))}
}
\value{
An unvalidated instance of the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\description{
Create a new unvalidated instance of the First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
\details{
First Bounce S3 class for Assessment of Quality of Life Six Dimension Health Utility - Adolescent Version (AQoL6d Adolescent))
}
