% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_bads.R
\name{make_new_firstbounce_bads}
\alias{make_new_firstbounce_bads}
\title{Make new First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
make_new_firstbounce_bads(x)
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores}
}
\value{
An unvalidated instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Create a new unvalidated instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
