% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_phq9.R
\name{make_new_firstbounce_phq9}
\alias{make_new_firstbounce_phq9}
\title{Make new First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores}
\usage{
make_new_firstbounce_phq9(x)
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores}
}
\value{
An unvalidated instance of the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\description{
Create a new unvalidated instance of the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\details{
First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
