% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_scared.R
\name{make_new_firstbounce_scared}
\alias{make_new_firstbounce_scared}
\title{Make new First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
\usage{
make_new_firstbounce_scared(x)
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
}
\value{
An unvalidated instance of the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\description{
Create a new unvalidated instance of the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\details{
First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
