% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_TTU_predictors_lup.R
\name{make_new_TTU_predictors_lup}
\alias{make_new_TTU_predictors_lup}
\title{Make new TTU predictors lookup table TTU S3 class for candidate predictors lookup table}
\usage{
make_new_TTU_predictors_lup(x)
}
\arguments{
\item{x}{A prototype for the TTU S3 class for candidate predictors lookup table}
}
\value{
An unvalidated instance of the TTU S3 class for candidate predictors lookup table
}
\description{
Create a new unvalidated instance of the TTU S3 class for candidate predictors lookup table
}
\details{
TTU S3 class for candidate predictors lookup table
}
