% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_output_format_ls}
\alias{make_output_format_ls}
\title{Make output format list}
\usage{
make_output_format_ls(
  manuscript_outp_1L_chr = "Word",
  manuscript_digits_1L_int = 2L,
  supplementary_outp_1L_chr = "PDF",
  supplementary_digits_1L_int = 2L
)
}
\arguments{
\item{manuscript_outp_1L_chr}{Manuscript output (a character vector of length one), Default: 'Word'}

\item{manuscript_digits_1L_int}{Manuscript digits (an integer vector of length one), Default: 2}

\item{supplementary_outp_1L_chr}{Supplementary output (a character vector of length one), Default: 'PDF'}

\item{supplementary_digits_1L_int}{Supplementary digits (an integer vector of length one), Default: 2}
}
\value{
Output format (a list)
}
\description{
make_output_format_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make output format list. The function returns Output format (a list).
}
\keyword{internal}
