% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_paths_to_ss_plts_ls}
\alias{make_paths_to_ss_plts_ls}
\title{Make paths to ss plots list}
\usage{
make_paths_to_ss_plts_ls(
  output_data_dir_1L_chr,
  outp_smry_ls,
  additional_paths_chr = "/dens_and_sctr.png"
)
}
\arguments{
\item{output_data_dir_1L_chr}{Output data directory (a character vector of length one)}

\item{outp_smry_ls}{Output summary (a list)}

\item{additional_paths_chr}{Additional paths (a character vector), Default: '/dens_and_sctr.png'}
}
\value{
Paths to ss plots (a list)
}
\description{
make_paths_to_ss_plts_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make paths to ss plots list. The function returns Paths to ss plots (a list).
}
\keyword{internal}
