% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_predn_ds_with_one_predr}
\alias{make_predn_ds_with_one_predr}
\title{Make predn dataset with one predr}
\usage{
make_predn_ds_with_one_predr(
  mdl,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  tfmn_1L_chr = "NTF",
  predr_var_nm_1L_chr,
  predr_vals_dbl,
  pred_type_1L_chr = NULL
)
}
\arguments{
\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{tfmn_1L_chr}{Tfmn (a character vector of length one), Default: 'NTF'}

\item{predr_var_nm_1L_chr}{Predr var name (a character vector of length one)}

\item{predr_vals_dbl}{Predr values (a double vector)}

\item{pred_type_1L_chr}{Pred type (a character vector of length one), Default: NULL}

\item{...}{Additional arguments}
}
\value{
Predn dataset (a tibble)
}
\description{
make_predn_ds_with_one_predr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make predn dataset with one predr. The function returns Predn dataset (a tibble).
}
