% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_predr_ctgs_ls}
\alias{make_predr_ctgs_ls}
\title{Make predictor category categoriess list}
\usage{
make_predr_ctgs_ls(outp_smry_ls, include_idx_int = NULL)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{include_idx_int}{Include index (an integer vector), Default: NULL}
}
\value{
Predictor category categoriess (a list)
}
\description{
make_predr_ctgs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make predictor category categoriess list. The function returns Predictor category categoriess (a list).
}
\keyword{internal}
