% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_predr_vals}
\alias{make_predr_vals}
\title{Make predr values}
\usage{
make_predr_vals(predr_var_nm_1L_chr, candidate_predrs_lup = NULL)
}
\arguments{
\item{predr_var_nm_1L_chr}{Predr var name (a character vector of length one)}

\item{candidate_predrs_lup}{Candidate predrs (a lookup table), Default: NULL}
}
\value{
Predr values (a double vector)
}
\description{
make_predr_vals() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make predr values. The function returns Predr values (a double vector).
}
