% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_predr_vars_nms_ls}
\alias{make_predr_vars_nms_ls}
\title{Make predr vars names}
\usage{
make_predr_vars_nms_ls(main_predrs_chr, covars_ls)
}
\arguments{
\item{main_predrs_chr}{Main predrs (a character vector)}

\item{covars_ls}{Covars (a list)}
}
\value{
Predr vars names (a list)
}
\description{
make_predr_vars_nms_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make predr vars names list. The function returns Predr vars names (a list).
}
