% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_predrs_for_best_mdls}
\alias{make_predrs_for_best_mdls}
\title{Make predictors for best models}
\usage{
make_predrs_for_best_mdls(outp_smry_ls, old_nms_chr = NULL, new_nms_chr = NULL)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{old_nms_chr}{Old names (a character vector), Default: NULL}

\item{new_nms_chr}{New names (a character vector), Default: NULL}
}
\value{
Predictors for best models (a character vector)
}
\description{
make_predrs_for_best_mdls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make predictors for best models. The function returns Predictors for best models (a character vector).
}
\keyword{internal}
