% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_bads.R
\name{make_prototype_firstbounce_bads}
\alias{make_prototype_firstbounce_bads}
\title{Make prototype First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
make_prototype_firstbounce_bads()
}
\value{
A prototype for First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Create a new prototype for the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
