% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_gad7.R
\name{make_prototype_firstbounce_gad7}
\alias{make_prototype_firstbounce_gad7}
\title{Make prototype First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores}
\usage{
make_prototype_firstbounce_gad7()
}
\value{
A prototype for First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\description{
Create a new prototype for the First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\details{
First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
