% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_k6.R
\name{make_prototype_firstbounce_k6}
\alias{make_prototype_firstbounce_k6}
\title{Make prototype First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores}
\usage{
make_prototype_firstbounce_k6()
}
\value{
A prototype for First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\description{
Create a new prototype for the First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\details{
First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
