% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_scared.R
\name{make_prototype_firstbounce_scared}
\alias{make_prototype_firstbounce_scared}
\title{Make prototype First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
\usage{
make_prototype_firstbounce_scared()
}
\value{
A prototype for First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\description{
Create a new prototype for the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\details{
First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
