% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_psych_predrs_lup}
\alias{make_psych_predrs_lup}
\title{Make psych predictors lookup table}
\usage{
make_psych_predrs_lup()
}
\value{
Predictors (a lookup table)
}
\description{
make_psych_predrs_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make psych predictors lookup table. The function returns Predictors (a lookup table).
}
\keyword{internal}
