% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_TTU_predictors_lup.R
\name{make_pt_TTU_predictors_lup}
\alias{make_pt_TTU_predictors_lup}
\title{Make prototype TTU S3 class for candidate predictors lookup table}
\usage{
make_pt_TTU_predictors_lup(
  short_name_chr = character(0),
  long_name_chr = character(0),
  min_val_dbl = numeric(0),
  max_val_dbl = numeric(0),
  class_chr = character(0),
  increment_dbl = numeric(0),
  class_fn_chr = character(0),
  mdl_scaling_dbl = numeric(0),
  covariate_lgl = logical(0)
)
}
\arguments{
\item{short_name_chr}{Short name (a character vector), Default: character(0)}

\item{long_name_chr}{Long name (a character vector), Default: character(0)}

\item{min_val_dbl}{Min value (a double vector), Default: numeric(0)}

\item{max_val_dbl}{Max value (a double vector), Default: numeric(0)}

\item{class_chr}{Class (a character vector), Default: character(0)}

\item{increment_dbl}{Increment (a double vector), Default: numeric(0)}

\item{class_fn_chr}{Class function (a character vector), Default: character(0)}

\item{mdl_scaling_dbl}{Mdl scaling (a double vector), Default: numeric(0)}

\item{covariate_lgl}{Covariate (a logical vector), Default: logical(0)}
}
\value{
A prototype for TTU S3 class for candidate predictors lookup table
}
\description{
Create a new prototype for the TTU S3 class for candidate predictors lookup table
}
\details{
TTU S3 class for candidate predictors lookup table
}
