% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_random_forest_text}
\alias{make_random_forest_text}
\title{Make random forest text}
\usage{
make_random_forest_text(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Text (a character vector of length one)
}
\description{
make_random_forest_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make random forest text. The function returns Text (a character vector of length one).
}
\keyword{internal}
