% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ranked_predrs_ls}
\alias{make_ranked_predrs_ls}
\title{Make ranked predictors list}
\usage{
make_ranked_predrs_ls(descv_tbls_ls, old_nms_chr = NULL, new_nms_chr = NULL)
}
\arguments{
\item{descv_tbls_ls}{Descriptive tables (a list)}

\item{old_nms_chr}{Old names (a character vector), Default: NULL}

\item{new_nms_chr}{New names (a character vector), Default: NULL}
}
\value{
Ranked predictors (a list)
}
\description{
make_ranked_predrs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ranked predictors list. The function returns Ranked predictors (a list).
}
\keyword{internal}
