% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_report_fl_nms_ls}
\alias{make_report_fl_nms_ls}
\title{Make report file names list}
\usage{
make_report_fl_nms_ls()
}
\value{
Report file names (a list)
}
\description{
make_report_fl_nms_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make report file names list. The function returns Report file names (a list).
}
\keyword{internal}
