% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_results_ls}
\alias{make_results_ls}
\title{Make results}
\usage{
make_results_ls(
  spine_of_results_ls,
  ctgl_vars_regrouping_ls = NULL,
  sig_covars_some_predrs_mdls_tb,
  sig_thresh_covars_1L_chr
)
}
\arguments{
\item{spine_of_results_ls}{Spine of results (a list)}

\item{ctgl_vars_regrouping_ls}{Ctgl variables regrouping (a list), Default: NULL}

\item{sig_covars_some_predrs_mdls_tb}{Sig covariates some predictors models (a tibble)}

\item{sig_thresh_covars_1L_chr}{Sig thresh covariates (a character vector of length one)}
}
\value{
Results (a list)
}
\description{
make_results_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make results list. The function returns Results (a list).
}
