% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_results_ls_spine}
\alias{make_results_ls_spine}
\title{Make results list spine}
\usage{
make_results_ls_spine(
  output_data_dir_1L_chr,
  var_nm_change_lup,
  study_descs_ls,
  nbr_of_digits_1L_int = 2L
)
}
\arguments{
\item{output_data_dir_1L_chr}{Output data directory (a character vector of length one)}

\item{var_nm_change_lup}{Variable name change (a lookup table)}

\item{study_descs_ls}{Study descriptions (a list)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}
}
\value{
Spine of results (a list)
}
\description{
make_results_ls_spine() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make results list spine. The function returns Spine of results (a list).
}
