% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_scaling_text}
\alias{make_scaling_text}
\title{Make scaling text}
\usage{
make_scaling_text(results_ls, table_1L_chr = "cfscl")
}
\arguments{
\item{results_ls}{Results (a list)}

\item{table_1L_chr}{Table (a character vector of length one), Default: 'cfscl'}
}
\value{
Text (a character vector of length one)
}
\description{
make_scaling_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make scaling text. The function returns Text (a character vector of length one).
}
\keyword{internal}
