% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_scndry_anlys_params}
\alias{make_scndry_anlys_params}
\title{Make secondary analysis parameters}
\usage{
make_scndry_anlys_params(
  scndry_anlys_params_ls = NULL,
  candidate_covar_nms_chr = NULL,
  candidate_predrs_chr = NULL,
  predictors_lup = NULL,
  prefd_covars_chr = NA_character_
)
}
\arguments{
\item{scndry_anlys_params_ls}{Secondary analysis parameters (a list), Default: NULL}

\item{candidate_covar_nms_chr}{Candidate covariate names (a character vector), Default: NULL}

\item{candidate_predrs_chr}{Candidate predictors (a character vector), Default: NULL}

\item{predictors_lup}{Predictors (a lookup table), Default: NULL}

\item{prefd_covars_chr}{Preferred covariates (a character vector), Default: 'NA'}
}
\value{
New parameters (a list)
}
\description{
make_scndry_anlys_params() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make secondary analysis parameters. The function returns New parameters (a list).
}
