% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_smry_of_ts_mdl}
\alias{make_smry_of_ts_mdl}
\title{Make smry of ts mdl}
\usage{
make_smry_of_ts_mdl(
  data_tb,
  fn,
  predictor_vars_nms_chr,
  mdl_nm_1L_chr,
  path_to_write_to_1L_chr = NA_character_,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  iters_1L_int = 4000L,
  seed_1L_int = 1000L
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{fn}{Function (a function)}

\item{predictor_vars_nms_chr}{Predictor vars names (a character vector)}

\item{mdl_nm_1L_chr}{Mdl name (a character vector of length one)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one), Default: 'NA'}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round var name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round bl value (a character vector of length one), Default: 'Baseline'}

\item{iters_1L_int}{Iters (an integer vector of length one), Default: 4000}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1000}
}
\value{
Smry of ts mdl (a list)
}
\description{
make_smry_of_ts_mdl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make smry of ts mdl. The function returns Smry of ts mdl (a list).
}
