% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sngl_mdl_smry_tb}
\alias{make_sngl_mdl_smry_tb}
\title{Make single model summary tibble}
\usage{
make_sngl_mdl_smry_tb(
  mdls_tb,
  mdl_nm_1L_chr,
  mdl_type_1L_chr,
  add_mdl_nm_sfx_1L_lgl = T
)
}
\arguments{
\item{mdls_tb}{Models (a tibble)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{mdl_type_1L_chr}{Model type (a character vector of length one)}

\item{add_mdl_nm_sfx_1L_lgl}{Add model name suffix (a logical vector of length one), Default: T}
}
\value{
New (a tibble)
}
\description{
make_sngl_mdl_smry_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make single model summary tibble. The function returns New (a tibble).
}
\keyword{internal}
