% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_study_descs_ls}
\alias{make_study_descs_ls}
\title{Make study descriptions}
\usage{
make_study_descs_ls(
  health_utl_nm_1L_chr,
  time_btwn_bl_and_fup_1L_chr,
  predr_ctgs_ls
)
}
\arguments{
\item{health_utl_nm_1L_chr}{Health utility name (a character vector of length one)}

\item{time_btwn_bl_and_fup_1L_chr}{Time btwn baseline and follow-up (a character vector of length one)}

\item{predr_ctgs_ls}{Predictor category categoriess (a list)}
}
\value{
Study descriptions (a list)
}
\description{
make_study_descs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make study descriptions list. The function returns Study descriptions (a list).
}
\keyword{internal}
