% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ten_fold_text}
\alias{make_ten_fold_text}
\title{Make ten fold text}
\usage{
make_ten_fold_text(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Text (a character vector of length one)
}
\description{
make_ten_fold_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ten fold text. The function returns Text (a character vector of length one).
}
\keyword{internal}
