% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tfd_sngl_predr_mdls_tb}
\alias{make_tfd_sngl_predr_mdls_tb}
\title{Make transformed single predictor models tibble}
\usage{
make_tfd_sngl_predr_mdls_tb(
  outp_smry_ls,
  nbr_of_digits_1L_int = 2L,
  mdl_pfx_ls = list(OLS = "Ordinary Least Squares ", GLM =
    c("Generalised Linear Model with ", "Beta Regression Model with Binomial "))
)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}

\item{mdl_pfx_ls}{Model prefix (a list), Default: list(OLS = "Ordinary Least Squares ", GLM = c("Generalised Linear Model with ",
"Beta Regression Model with Binomial "))}
}
\value{
Transformed single predictor models (a tibble)
}
\description{
make_tfd_sngl_predr_mdls_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make transformed single predictor models tibble. The function returns Transformed single predictor models (a tibble).
}
\keyword{internal}
