% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tfmn_cmprsn_plt}
\alias{make_tfmn_cmprsn_plt}
\title{Make transformation comparison plot}
\usage{
make_tfmn_cmprsn_plt(data_tb, depnt_var_nm_1L_chr, dictionary_tb)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one)}

\item{dictionary_tb}{Dictionary (a tibble)}
}
\value{
Transformation comparison (a plot)
}
\description{
make_tfmn_cmprsn_plt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make transformation comparison plot. The function returns Transformation comparison (a plot).
}
