% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_two_mdl_types_smry_tbl}
\alias{make_two_mdl_types_smry_tbl}
\title{Make two model types summary table}
\usage{
make_two_mdl_types_smry_tbl(outp_smry_ls, mdls_tb)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{mdls_tb}{Models (a tibble)}
}
\value{
Two model types summary table (a tibble)
}
\description{
make_two_mdl_types_smry_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make two model types summary table. The function returns Two model types summary table (a tibble).
}
