% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_uid_rename_lup}
\alias{make_uid_rename_lup}
\title{Make unique identifier rename lookup table}
\usage{
make_uid_rename_lup(data_tb, id_var_nm_1L_chr = "UID")
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'UID'}
}
\value{
Unique identifier rename lookup table (a tibble)
}
\description{
make_uid_rename_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make unique identifier rename lookup table. The function returns Unique identifier rename lookup table (a tibble).
}
\keyword{internal}
