% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_unique_ls_elmt_idx_int}
\alias{make_unique_ls_elmt_idx_int}
\title{Make unique list element index integer vector}
\usage{
make_unique_ls_elmt_idx_int(data_ls)
}
\arguments{
\item{data_ls}{Data (a list)}
}
\value{
Unique list element index (an integer vector)
}
\description{
make_unique_ls_elmt_idx_int() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make unique list element index integer vector. The function returns Unique list element index (an integer vector).
}
\keyword{internal}
