% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_vec_with_sum_of_int_val}
\alias{make_vec_with_sum_of_int_val}
\title{Make vector with sum of integer vector value}
\usage{
make_vec_with_sum_of_int_val(target_int, start_int, end_int, length_int)
}
\arguments{
\item{target_int}{Target (an integer vector)}

\item{start_int}{Start (an integer vector)}

\item{end_int}{End (an integer vector)}

\item{length_int}{Length (an integer vector)}
}
\value{
Vector (an integer vector)
}
\description{
make_vec_with_sum_of_int_val() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make vector with sum of integer vector value. The function returns Vector (an integer vector).
}
\keyword{internal}
