% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_within_between_ratios_text}
\alias{make_within_between_ratios_text}
\title{Make within between ratios text}
\usage{
make_within_between_ratios_text(results_ls, exclude_covars_1L_lgl = F)
}
\arguments{
\item{results_ls}{Results (a list)}

\item{exclude_covars_1L_lgl}{Exclude covariates (a logical vector of length one), Default: F}
}
\value{
Text (a character vector of length one)
}
\description{
make_within_between_ratios_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make within between ratios text. The function returns Text (a character vector of length one).
}
