% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-TTUProject}
\alias{manufacture-TTUProject}
\alias{manufacture,TTUProject-method}
\title{Manufacture a new object}
\usage{
\S4method{manufacture}{TTUProject}(x, type_1L_chr = "dummys", what_1L_chr = "factors", ...)
}
\arguments{
\item{x}{An object of class TTUProject}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'dummys'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'factors'}

\item{...}{Additional arguments}
}
\value{
Object (an output object of multiple potential types)
}
\description{
manufacture method applied to TTUProject
}
