% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_lnr_cmprsn_sctr_plt}
\alias{plot_lnr_cmprsn_sctr_plt}
\title{Plot lnr cmprsn sctr plt}
\usage{
plot_lnr_cmprsn_sctr_plt(
  tfd_data_tb,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  predd_val_var_nm_1L_chr = "Predicted"
)
}
\arguments{
\item{tfd_data_tb}{Transformed data (a tibble)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{predd_val_var_nm_1L_chr}{Predd value var name (a character vector of length one), Default: 'Predicted'}
}
\description{
plot_lnr_cmprsn_sctr_plt() is a Plot function that plots data Specifically, this function implements an algorithm to plot lnr cmprsn sctr plt. The function is called for its side effects and does not return a value.
}
\keyword{internal}
