% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_obsd_predd_dnst}
\alias{plot_obsd_predd_dnst}
\title{Plot observed predicted density}
\usage{
plot_obsd_predd_dnst(
  tfd_data_tb,
  depnt_var_desc_1L_chr = "AQoL-6D utility score",
  predd_val_var_nm_1L_chr = "Predicted"
)
}
\arguments{
\item{tfd_data_tb}{Transformed data (a tibble)}

\item{depnt_var_desc_1L_chr}{Dependent variable description (a character vector of length one), Default: 'AQoL-6D utility score'}

\item{predd_val_var_nm_1L_chr}{Predicted value variable name (a character vector of length one), Default: 'Predicted'}
}
\description{
plot_obsd_predd_dnst() is a Plot function that plots data Specifically, this function implements an algorithm to plot observed predicted density. The function is called for its side effects and does not return a value.
}
