% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_obsd_predd_sctr}
\alias{plot_obsd_predd_sctr}
\title{Plot obsd predd sctr}
\usage{
plot_obsd_predd_sctr(
  tfd_data_tb,
  dep_var_nm_1L_chr,
  dep_var_desc_1L_chr,
  args_ls
)
}
\arguments{
\item{tfd_data_tb}{Transformed data (a tibble)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one)}

\item{dep_var_desc_1L_chr}{Dep var description (a character vector of length one)}

\item{args_ls}{Arguments (a list)}
}
\description{
plot_obsd_predd_sctr() is a Plot function that plots data Specifically, this function implements an algorithm to plot obsd predd sctr. The function is called for its side effects and does not return a value.
}
