% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_obsd_predd_sctr_cmprsn}
\alias{plot_obsd_predd_sctr_cmprsn}
\title{Plot observed predicted scatter comparison}
\usage{
plot_obsd_predd_sctr_cmprsn(
  tfd_data_tb,
  depnt_var_nm_1L_chr = "utl_total_w",
  depnt_var_desc_1L_chr = "Total weighted utility score",
  round_var_nm_1L_chr = "round",
  args_ls = NULL,
  predd_val_var_nm_1L_chr = "Predicted"
)
}
\arguments{
\item{tfd_data_tb}{Transformed data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{depnt_var_desc_1L_chr}{Dependent variable description (a character vector of length one), Default: 'Total weighted utility score'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{predd_val_var_nm_1L_chr}{Predicted value variable name (a character vector of length one), Default: 'Predicted'}
}
\description{
plot_obsd_predd_sctr_cmprsn() is a Plot function that plots data Specifically, this function implements an algorithm to plot observed predicted scatter comparison. The function is called for its side effects and does not return a value.
}
\keyword{internal}
