% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_sctr_plt_cmprsn}
\alias{plot_sctr_plt_cmprsn}
\title{Plot scatter plot comparison}
\usage{
plot_sctr_plt_cmprsn(
  tfd_data_tb,
  depnt_var_nm_1L_chr = "utl_total_w",
  predd_val_var_nm_1L_chr = "Predicted"
)
}
\arguments{
\item{tfd_data_tb}{Transformed data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{predd_val_var_nm_1L_chr}{Predicted value variable name (a character vector of length one), Default: 'Predicted'}
}
\description{
plot_sctr_plt_cmprsn() is a Plot function that plots data Specifically, this function implements an algorithm to plot scatter plot comparison. The function is called for its side effects and does not return a value.
}
\keyword{internal}
