% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_plt_types_lup.R
\docType{data}
\name{plt_types_lup}
\alias{plt_types_lup}
\title{Model plot types lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.
}
\usage{
plt_types_lup
}
\description{
A lookup table of abbreviations to describe the different model plot types supported by TTU functions
}
\details{
A tibble

\describe{
\item{short_name_chr}{Short name (a character vector)}
\item{long_name_chr}{Long name (a character vector)}
}
}
\keyword{datasets}
