% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_aqol6d}
\alias{predict_aqol6d}
\title{Predict Assessment of Quality of Life Six Dimension}
\usage{
predict_aqol6d(data_tb, tfmn_1L_chr = "NTF", mdl)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{tfmn_1L_chr}{Tfmn (a character vector of length one), Default: 'NTF'}

\item{...}{Additional arguments}
}
\value{
Predd Assessment of Quality of Life Six Dimension (a double vector)
}
\description{
predict_aqol6d() is a Predict function that makes predictions from data using a specified statistical model. Specifically, this function implements an algorithm to predict assessment of quality of life six dimension. The function returns Predd Assessment of Quality of Life Six Dimension (a double vector).
}
\keyword{internal}
