% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_from_shareable_mdl}
\alias{predict_from_shareable_mdl}
\title{Predict from shareable model}
\usage{
predict_from_shareable_mdl(
  model_mdl,
  data_tb,
  predn_type_1L_chr = "response",
  sd_dbl,
  deterministic_1L_lgl = T
)
}
\arguments{
\item{model_mdl}{Model (a model)}

\item{data_tb}{Data (a tibble)}

\item{predn_type_1L_chr}{Prediction type (a character vector of length one), Default: 'response'}

\item{sd_dbl}{Standard deviation (a double vector)}

\item{deterministic_1L_lgl}{Deterministic (a logical vector of length one), Default: T}
}
\value{
New data (a double vector)
}
\description{
predict_from_shareable_mdl() is a Predict function that makes predictions from data using a specified statistical model. Specifically, this function implements an algorithm to predict from shareable model. The function returns New data (a double vector).
}
\keyword{internal}
