% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_shrble_betareg}
\alias{predict_shrble_betareg}
\title{Predict shareable betareg}
\usage{
predict_shrble_betareg(
  object,
  newdata = NULL,
  type = c("response", "link", "precision", "variance", "quantile"),
  na.action = na.pass,
  at = 0.5,
  sd_1L_dbl,
  ...
)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{newdata}{New dataset (for prediction) (a new dataset (for prediction)), Default: NULL}

\item{type}{PARAM_DESCRIPTION, Default: c("response", "link", "precision", "variance", "quantile")}

\item{na.action}{PARAM_DESCRIPTION, Default: na.pass}

\item{at}{PARAM_DESCRIPTION, Default: 0.5}

\item{sd_1L_dbl}{Standard deviation (a double vector of length one)}

\item{...}{Additional arguments}
}
\description{
predict_shrble_betareg() is a Predict function that makes predictions from data using a specified statistical model. Specifically, this function implements an algorithm to predict shareable betareg. The function is called for its side effects and does not return a value.
}
\keyword{internal}
