% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_predictors_lup.R
\docType{data}
\name{predictors_lup}
\alias{predictors_lup}
\title{Predictors lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 9 columns.
}
\usage{
predictors_lup
}
\description{
A lookup table of the short name and long name of each predictor used in the models included with the youthu package.
}
\details{
A tibble

\describe{
\item{short_name_chr}{Short name (a character vector)}
\item{long_name_chr}{Long name (a character vector)}
\item{min_val_dbl}{Minimum value (a double vector)}
\item{max_val_dbl}{Maximum value (a double vector)}
\item{class_chr}{Class (a character vector)}
\item{increment_dbl}{Increment (a double vector)}
\item{class_fn_chr}{Class function (a character vector)}
\item{mdl_scaling_dbl}{Model scaling (a double vector)}
\item{covariate_lgl}{Covariate (a logical vector)}
}
}
\keyword{datasets}
