% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_cohort_table}
\alias{print_cohort_table}
\title{Print cohort table}
\usage{
print_cohort_table(params_ls, caption_1L_chr, mkdn_tbl_ref_1L_chr)
}
\arguments{
\item{params_ls}{Parameters (a list)}

\item{caption_1L_chr}{Caption (a character vector of length one)}

\item{mkdn_tbl_ref_1L_chr}{Markdown table reference (a character vector of length one)}
}
\description{
print_cohort_table() is a Print function that prints output to console Specifically, this function implements an algorithm to print cohort table. The function is called for its side effects and does not return a value.
}
\keyword{internal}
