% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_indpnt_predrs_lngl_mdl_coefs}
\alias{print_indpnt_predrs_lngl_mdl_coefs}
\title{Print independent predictors longitudinal model coefficients}
\usage{
print_indpnt_predrs_lngl_mdl_coefs(
  params_ls,
  caption_1L_chr,
  ref_1L_int = 1,
  table_1L_chr
)
}
\arguments{
\item{params_ls}{Parameters (a list)}

\item{caption_1L_chr}{Caption (a character vector of length one)}

\item{ref_1L_int}{Reference (an integer vector of length one), Default: 1}

\item{table_1L_chr}{Table (a character vector of length one)}
}
\description{
print_indpnt_predrs_lngl_mdl_coefs() is a Print function that prints output to console Specifically, this function implements an algorithm to print independent predictors longitudinal model coefficients. The function is called for its side effects and does not return a value.
}
