% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_lngl_ttu_tbls}
\alias{print_lngl_ttu_tbls}
\title{Print lngl ttu tables}
\usage{
print_lngl_ttu_tbls(
  table_df,
  params_ls,
  caption_1L_chr,
  table_1L_chr,
  ref_1L_int = 1
)
}
\arguments{
\item{table_df}{Table (a data.frame)}

\item{params_ls}{Params (a list)}

\item{caption_1L_chr}{Caption (a character vector of length one)}

\item{table_1L_chr}{Table (a character vector of length one)}

\item{ref_1L_int}{Reference (an integer vector of length one), Default: 1}
}
\description{
print_lngl_ttu_tbls() is a Print function that prints output to console Specifically, this function implements an algorithm to print lngl ttu tables. The function is called for its side effects and does not return a value.
}
\keyword{internal}
