% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_ts_mdl_plts}
\alias{print_ts_mdl_plts}
\title{Print time series model plots}
\usage{
print_ts_mdl_plts(paths_to_plts_chr, title_1L_chr, label_refs_chr, mdl_smry_ls)
}
\arguments{
\item{paths_to_plts_chr}{Paths to plots (a character vector)}

\item{title_1L_chr}{Title (a character vector of length one)}

\item{label_refs_chr}{Label references (a character vector)}

\item{mdl_smry_ls}{Model summary (a list)}
}
\description{
print_ts_mdl_plts() is a Print function that prints output to console Specifically, this function implements an algorithm to print time series model plots. The function is called for its side effects and does not return a value.
}
