% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_procure.R
\name{procure-TTUProject}
\alias{procure-TTUProject}
\alias{procure,TTUProject-method}
\title{Procure items from a dataset}
\usage{
\S4method{procure}{TTUProject}(
  x,
  type_1L_chr = "default",
  variables_chr = character(0),
  what_1L_chr = "records",
  ...
)
}
\arguments{
\item{x}{An object of class TTUProject}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'default'}

\item{variables_chr}{Variables (a character vector), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'records'}

\item{...}{Additional arguments}
}
\value{
Object (an output object of multiple potential types)
}
\description{
procure method applied to TTUProject
}
