% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_randomise.R
\name{randomise_change_in_fct_var}
\alias{randomise_change_in_fct_var}
\title{Randomise change in factor var}
\usage{
randomise_change_in_fct_var(vector_fct, prob_unchanged_dbl)
}
\arguments{
\item{vector_fct}{Vector (a factor)}

\item{prob_unchanged_dbl}{Prob unchanged (a double vector)}
}
\description{
randomise_change_in_fct_var() is a Randomise function that randomly samples from data. Specifically, this function implements an algorithm to randomise change in factor var. The function is called for its side effects and does not return a value.
}
