% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_randomise.R
\name{randomise_changes_in_fct_lvls}
\alias{randomise_changes_in_fct_lvls}
\title{Randomise changes in factor vector levels}
\usage{
randomise_changes_in_fct_lvls(vector_fct, prob_unchanged_dbl)
}
\arguments{
\item{vector_fct}{Vector (a factor vector)}

\item{prob_unchanged_dbl}{Probability unchanged (a double vector)}
}
\description{
randomise_changes_in_fct_lvls() is a Randomise function that randomly samples from data. Specifically, this function implements an algorithm to randomise changes in factor vector levels. The function is called for its side effects and does not return a value.
}
\keyword{internal}
