% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_randomise.R
\name{randomise_ptl_fup_fct}
\alias{randomise_ptl_fup_fct}
\title{Randomise ptl fup}
\usage{
randomise_ptl_fup_fct(vector_fct, prob_unchanged_dbl)
}
\arguments{
\item{vector_fct}{Vector (a factor)}

\item{prob_unchanged_dbl}{Prob unchanged (a double vector)}
}
\description{
randomise_ptl_fup_fct() is a Randomise function that randomly samples from data. Specifically, this function implements an algorithm to randomise ptl fup factor. The function is called for its side effects and does not return a value.
}
\keyword{internal}
