% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4u.R
\docType{package}
\name{ready4u-package}
\alias{ready4u}
\alias{ready4u-package}
\title{ready4u: Map Measures Routinely Collected in Mental Health
Services To Health Utility}
\description{
Tools for developping Transfer To Utility (TTU)
mapping algorithms to predict health utility from measures routinely
collected in mental health services.  This development version of the
ready4u package has been made available as part of the process of
testing and documenting the package. The tools contained in this
development release automate a number of tasks which MODIFY THE
DIRECTORY STRUCTURE OF YOUR LOCAL MACHINE.  Therefore you should only
trial this software if you feel confident that you understand what it
does and have created a sandpit area in which you can safely undertake
testing. If you have any questions, please contact the authors
(matthew.hamilton@orygen.org.au).  The documentation for this package
has been automatically generated by the ready4fun package and is
therefore quite rudimentary. Human authored documentation will follow
in 2021.
}
\details{
To learn more about ready4u, start with the vignettes:
\code{browseVignettes(package = "ready4u")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4u/}
  \item \url{https://github.com/ready4-dev/ready4u}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
