% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rename.R
\name{rename_from_nmd_vec}
\alias{rename_from_nmd_vec}
\title{Rename from named vector}
\usage{
rename_from_nmd_vec(data_tb, nmd_vec_chr, vec_nms_as_new_1L_lgl = T)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{nmd_vec_chr}{Named vector (a character vector)}

\item{vec_nms_as_new_1L_lgl}{Vector names as new (a logical vector of length one), Default: T}
}
\value{
Renamed data (a tibble)
}
\description{
rename_from_nmd_vec() is a Rename function that renames elements of an object based on a pre-speccified schema. Specifically, this function implements an algorithm to rename from named vector. The function returns Renamed data (a tibble).
}
\keyword{internal}
