% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew-TTUProject}
\alias{renew-TTUProject}
\alias{renew,TTUProject-method}
\title{Renew (update) values}
\usage{
\S4method{renew}{TTUProject}(
  x,
  new_val_xx = NULL,
  consent_1L_chr = "",
  depnt_var_min_val_1L_dbl = numeric(0),
  fl_nm_1L_chr = character(0),
  paths_chr = character(0),
  type_1L_chr = "default",
  y_Ready4useRepos = ready4use::Ready4useRepos(),
  what_1L_chr = "utility",
  ...
)
}
\arguments{
\item{x}{An object of class TTUProject}

\item{new_val_xx}{New value (an output object of multiple potential types), Default: NULL}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{depnt_var_min_val_1L_dbl}{Dependent variable minimum value (a double vector of length one), Default: numeric(0)}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: character(0)}

\item{paths_chr}{Paths (a character vector), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'default'}

\item{y_Ready4useRepos}{PARAM_DESCRIPTION, Default: ready4use::Ready4useRepos()}

\item{what_1L_chr}{What (a character vector of length one), Default: 'utility'}

\item{...}{Additional arguments}
}
\value{
x (An object of class TTUProject)
}
\description{
renew method applied to TTUProject
}
