% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reorder.R
\name{reorder_cndt_predrs_chr}
\alias{reorder_cndt_predrs_chr}
\title{Reorder cndt predrs}
\usage{
reorder_cndt_predrs_chr(
  candidate_predrs_chr,
  data_tb,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  method_1L_chr = "pearson"
)
}
\arguments{
\item{candidate_predrs_chr}{Candidate predrs (a character vector)}

\item{data_tb}{Data (a tibble)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{method_1L_chr}{Method (a character vector of length one), Default: 'pearson'}
}
\value{
NA ()
}
\description{
reorder_cndt_predrs_chr() is a Reorder function that reorders an object to conform to a pre-specified schema. Specifically, this function implements an algorithm to reorder cndt predrs character vector. The function is called for its side effects and does not return a value.
}
\keyword{internal}
