% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reorder.R
\name{reorder_tb_for_target_cors}
\alias{reorder_tb_for_target_cors}
\title{Reorder tibble for target cors}
\usage{
reorder_tb_for_target_cors(
  tbs_ls,
  corr_dbl,
  corr_var_1L_chr,
  id_var_to_rm_1L_chr = NA_character_,
  included_fup_idx_dbl = NA_real_
)
}
\arguments{
\item{tbs_ls}{Tibbles (a list)}

\item{corr_dbl}{Corr (a double vector of length 1)}

\item{corr_var_1L_chr}{Corr var 1L (a character vector of length 1)}

\item{id_var_to_rm_1L_chr}{Id var to rm 1L (a character vector of length 1), Default: 'NA'}

\item{included_fup_idx_dbl}{Included fup idx (a double vector of length 1), Default: NA}
}
\value{
Tibbles (a list)
}
\description{
reorder_tb_for_target_cors() is a Reorder function that reorders an object to conform to a pre-specified schema. Specifically, this function implements an algorithm to reorder a tibble for target cors. The function returns a tibbles (a list).
}
\keyword{internal}
