% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_replication_popl_tb.R
\docType{data}
\name{replication_popl_tb}
\alias{replication_popl_tb}
\title{Synthetic population replication dataset}
\format{
An object of class \code{data.frame} with 1711 rows and 46 columns.
}
\usage{
replication_popl_tb
}
\description{
A purely synthetic dataset, representative of the original study data, that can be used for replication runs of package algorithms.
}
\details{
A tibble

\describe{
\item{fkClientID}{Unique Client Identifier}
\item{round}{Round of Data Collection}
\item{c_p_diag_s}{Primary Diagnosis}
\item{s_centre}{Centre Name}
\item{c_clinical_staging_s}{Clinical Stage}
\item{d_age}{Age}
\item{d_agegroup}{Age Group}
\item{d_gender}{Gender}
\item{d_sex_birth_s}{Sex at Birth}
\item{d_sexual_ori_s}{Sexual Orientation}
\item{d_country_bir_s}{Country Of Birth}
\item{d_ATSI}{Aboriginal or Torres Strait Islander}
\item{d_english_home}{Speaks English At Home}
\item{d_english_native}{Native English Speaker}
\item{d_relation_s}{Relationship Status}
\item{d_studying_working}{Education and Employment Status}
\item{K6}{Kessler Psychological Distress Scale (6 Dimension)}
\item{PHQ9}{Patient Health Questionnaire}
\item{BADS}{Behavioural Activation for Depression Scale}
\item{GAD7}{Generalised Anxiety Disorder Scale}
\item{OASIS}{Overall Anxiety Severity and Impairment Scale}
\item{SCARED}{Screen for Child Anxiety Related Disorders}
\item{SOFAS}{Social and Occupational Functioning Assessment Scale}
\item{Gender}{Gender (Grouped)}
\item{Region}{Region of Residence (Metropolitan or Regional)}
\item{CALD}{Culturally And Linguistically Diverse}
\item{aqol6d_q1}{Assessment of Quality of Life (6 Dimension) Question 1}
\item{aqol6d_q2}{Assessment of Quality of Life (6 Dimension) Question 2}
\item{aqol6d_q3}{Assessment of Quality of Life (6 Dimension) Question 3}
\item{aqol6d_q4}{Assessment of Quality of Life (6 Dimension) Question 4}
\item{aqol6d_q5}{Assessment of Quality of Life (6 Dimension) Question 5}
\item{aqol6d_q6}{Assessment of Quality of Life (6 Dimension) Question 6}
\item{aqol6d_q7}{Assessment of Quality of Life (6 Dimension) Question 7}
\item{aqol6d_q8}{Assessment of Quality of Life (6 Dimension) Question 8}
\item{aqol6d_q9}{Assessment of Quality of Life (6 Dimension) Question 9}
\item{aqol6d_q10}{Assessment of Quality of Life (6 Dimension) Question 10}
\item{aqol6d_q11}{Assessment of Quality of Life (6 Dimension) Question 11}
\item{aqol6d_q12}{Assessment of Quality of Life (6 Dimension) Question 12}
\item{aqol6d_q13}{Assessment of Quality of Life (6 Dimension) Question 13}
\item{aqol6d_q14}{Assessment of Quality of Life (6 Dimension) Question 14}
\item{aqol6d_q15}{Assessment of Quality of Life (6 Dimension) Question 15}
\item{aqol6d_q16}{Assessment of Quality of Life (6 Dimension) Question 16}
\item{aqol6d_q17}{Assessment of Quality of Life (6 Dimension) Question 17}
\item{aqol6d_q18}{Assessment of Quality of Life (6 Dimension) Question 18}
\item{aqol6d_q19}{Assessment of Quality of Life (6 Dimension) Question 19}
\item{aqol6d_q20}{Assessment of Quality of Life (6 Dimension) Question 20}
}
}
\keyword{datasets}
