% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_repln_ds_dict_r3.R
\docType{data}
\name{repln_ds_dict_r3}
\alias{repln_ds_dict_r3}
\title{Data dictionary for study population dataset}
\format{
An object of class \code{ready4_dictionary} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 46 rows and 4 columns.
}
\usage{
repln_ds_dict_r3
}
\description{
A data dictionary of the variables used in the source and replication (synthetic) datasets for the First Bounce transfer to utility study
}
\details{
A tibble

\describe{
\item{var_nm_chr}{Var name (a character vector)}
\item{var_cat_chr}{Var cat (a character vector)}
\item{var_desc_chr}{Var description (a character vector)}
\item{var_type_chr}{Var type (a character vector)}
}
}
\keyword{datasets}
