% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_rprt_lup.R
\docType{data}
\name{rprt_lup}
\alias{rprt_lup}
\title{Report types lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 7 columns.
}
\usage{
rprt_lup
}
\description{
A lookup table of the different report types supported by TTU functions
}
\details{
A tibble

\describe{
\item{rprt_nms_chr}{Rprt names (a character vector)}
\item{title_chr}{Title (a character vector)}
\item{paths_to_RMD_dir_1L_chr}{Paths toMD directory (a character vector of length one)}
\item{pkg_dirs_chr}{Package directories (a character vector)}
\item{packages_chr}{Packages (a character vector)}
\item{nms_of_RMD_chr}{Names ofMD (a character vector)}
\item{rltv_paths_to_outpt_yaml_chr}{Rltv paths to outpt yaml (a character vector)}
}
}
\keyword{datasets}
