% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_scramble.R
\name{scramble_xx}
\alias{scramble_xx}
\title{Scramble}
\usage{
scramble_xx(vector_xx)
}
\arguments{
\item{vector_xx}{Vector (an output object of multiple potential types)}
}
\value{
Scrambled vec (an output object of multiple potential types)
}
\description{
scramble_xx() is a Scramble function that randomly reorders an object. Specifically, this function implements an algorithm to scramble output object of multiple potential types. The function returns Scrambled vec (an output object of multiple potential types).
}
