% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_share.R
\name{share-TTUProject}
\alias{share-TTUProject}
\alias{share,TTUProject-method}
\title{Share data via an online repository}
\usage{
\S4method{share}{TTUProject}(
  x,
  formats_chr = c(".docx", ".pdf", ".tex"),
  types_chr = "auto",
  what_chr = c("catalogue", "instrument", "models")
)
}
\arguments{
\item{x}{An object of class TTUProject}

\item{formats_chr}{Formats (a character vector), Default: c(".docx", ".pdf", ".tex")}

\item{types_chr}{Types (a character vector), Default: 'auto'}

\item{what_chr}{What (a character vector), Default: c("catalogue", "instrument", "models")}
}
\value{
x (An object of class TTUProject)
}
\description{
share method applied to TTUProject
}
