% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_chr_digit_pairs}
\alias{transform_chr_digit_pairs}
\title{Transform character vector digit pairs}
\usage{
transform_chr_digit_pairs(digit_pairs_chr, nbr_of_digits_1L_int = 2L)
}
\arguments{
\item{digit_pairs_chr}{Digit pairs (a character vector)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}
}
\value{
Transformed digit pairs (a character vector)
}
\description{
transform_chr_digit_pairs() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform character vector digit pairs. Function argument digit_pairs_chr specifies the object to be updated. Argument nbr_of_digits_1L_int provides the object to be updated. The function returns Transformed digit pairs (a character vector).
}
\keyword{internal}
