% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_dep_var_nm_for_cll}
\alias{transform_dep_var_nm_for_cll}
\title{Transform dep var name for cll}
\usage{
transform_dep_var_nm_for_cll(dep_var_nm_1L_chr)
}
\arguments{
\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one)}
}
\value{
Transformed dep var name (a character vector of length one)
}
\description{
transform_dep_var_nm_for_cll() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dep var name for cll. Function argument dep_var_nm_1L_chr specifies the object to be updated. The function returns Transformed dep var name (a character vector of length one).
}
