% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_depnt_var_nm}
\alias{transform_depnt_var_nm}
\title{Transform dependent variable name}
\usage{
transform_depnt_var_nm(depnt_var_nm_1L_chr, tfmn_1L_chr = "NTF")
}
\arguments{
\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}
}
\value{
Transformed dependent variable name (a character vector of length one)
}
\description{
transform_depnt_var_nm() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dependent variable name. Function argument depnt_var_nm_1L_chr specifies the object to be updated. Argument tfmn_1L_chr provides the object to be updated. The function returns Transformed dependent variable name (a character vector of length one).
}
\keyword{internal}
