% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_depnt_var_nm_for_cll}
\alias{transform_depnt_var_nm_for_cll}
\title{Transform dependent variable name for complementary log log}
\usage{
transform_depnt_var_nm_for_cll(depnt_var_nm_1L_chr)
}
\arguments{
\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one)}
}
\value{
Transformed dependent variable name (a character vector of length one)
}
\description{
transform_depnt_var_nm_for_cll() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dependent variable name for complementary log log. Function argument depnt_var_nm_1L_chr specifies the object to be updated. The function returns Transformed dependent variable name (a character vector of length one).
}
