% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_for_all_cmprsn_plts}
\alias{transform_ds_for_all_cmprsn_plts}
\title{Transform dataset for all comparison plots}
\usage{
transform_ds_for_all_cmprsn_plts(
  tfd_data_tb,
  model_mdl,
  depnt_var_nm_1L_chr,
  is_brms_mdl_1L_lgl,
  predn_type_1L_chr,
  sd_dbl,
  sfx_1L_chr = "",
  tfmn_1L_chr,
  utl_min_val_1L_dbl = -1
)
}
\arguments{
\item{tfd_data_tb}{Transformed data (a tibble)}

\item{model_mdl}{Model (a model)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one)}

\item{is_brms_mdl_1L_lgl}{Is bayesian regression models model (a logical vector of length one)}

\item{predn_type_1L_chr}{Prediction type (a character vector of length one)}

\item{sd_dbl}{Standard deviation (a double vector)}

\item{sfx_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{tfmn_1L_chr}{Transformation (a character vector of length one)}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: -1}
}
\value{
Transformed data (a tibble)
}
\description{
transform_ds_for_all_cmprsn_plts() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset for all comparison plots. Function argument tfd_data_tb specifies the object to be updated. Argument model_mdl provides the object to be updated. The function returns Transformed data (a tibble).
}
\keyword{internal}
