% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_for_mdlng}
\alias{transform_ds_for_mdlng}
\title{Transform dataset for mdlng}
\usage{
transform_ds_for_mdlng(
  data_tb,
  dep_var_nm_1L_chr = "utl_total_w",
  predr_var_nm_1L_chr,
  covar_var_nms_chr = NA_character_
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'utl_total_w'}

\item{predr_var_nm_1L_chr}{Predr var name (a character vector of length one)}

\item{covar_var_nms_chr}{Covar var names (a character vector), Default: 'NA'}
}
\value{
Transformed data (a tibble)
}
\description{
transform_ds_for_mdlng() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset for mdlng. Function argument data_tb specifies the object to be updated. Argument dep_var_nm_1L_chr provides the object to be updated. The function returns Transformed data (a tibble).
}
