% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_for_tstng}
\alias{transform_ds_for_tstng}
\title{Transform dataset for testing}
\usage{
transform_ds_for_tstng(
  data_tb,
  depnt_var_nm_1L_chr = "aqol6d_total_w",
  dep_var_max_val_1L_dbl = 0.999,
  candidate_predrs_chr = NA_character_,
  covar_var_nms_chr = NA_character_,
  round_var_nm_1L_chr = "round",
  round_val_1L_chr = "Baseline",
  remove_all_msng_1L_lgl = F
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{dep_var_max_val_1L_dbl}{Dep variable maximum value (a double vector of length one), Default: 0.999}

\item{candidate_predrs_chr}{Candidate predictors (a character vector), Default: 'NA'}

\item{covar_var_nms_chr}{Covariate variable names (a character vector), Default: 'NA'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{round_val_1L_chr}{Round value (a character vector of length one), Default: 'Baseline'}

\item{remove_all_msng_1L_lgl}{Remove all missing (a logical vector of length one), Default: F}
}
\value{
Transformed data (a tibble)
}
\description{
transform_ds_for_tstng() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset for testing. Function argument data_tb specifies the object to be updated. Argument depnt_var_nm_1L_chr provides the object to be updated. The function returns Transformed data (a tibble).
}
\keyword{internal}
