% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_to_predn_ds}
\alias{transform_ds_to_predn_ds}
\title{Transform dataset to prediction dataset}
\usage{
transform_ds_to_predn_ds(
  data_tb,
  predr_vars_nms_chr,
  tfmn_1L_chr,
  depnt_var_nm_1L_chr,
  id_var_nm_1L_chr,
  round_var_nm_1L_chr,
  round_bl_val_1L_chr,
  predictors_lup
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{predr_vars_nms_chr}{Predictor variables names (a character vector)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one)}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one)}

\item{round_bl_val_1L_chr}{Round baseline value (a character vector of length one)}

\item{predictors_lup}{Predictors (a lookup table)}
}
\value{
Data (a tibble)
}
\description{
transform_ds_to_predn_ds() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset to prediction dataset. Function argument data_tb specifies the object to be updated. Argument predr_vars_nms_chr provides the object to be updated. The function returns Data (a tibble).
}
