% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_with_rename_lup}
\alias{transform_ds_with_rename_lup}
\title{Transform dataset with rename}
\usage{
transform_ds_with_rename_lup(ds_tb, rename_lup, target_var_nms_chr = NULL)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{rename_lup}{Rename (a lookup table)}

\item{target_var_nms_chr}{Target variable names (a character vector), Default: NULL}
}
\value{
Tfmd dataset (a tibble)
}
\description{
transform_ds_with_rename_lup() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset with rename lookup table. Function argument ds_tb specifies the object to be updated. Argument rename_lup provides the object to be updated. The function returns Tfmd dataset (a tibble).
}
\keyword{internal}
