% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_names}
\alias{transform_names}
\title{Transform names}
\usage{
transform_names(names_chr, rename_lup, invert_1L_lgl = F)
}
\arguments{
\item{names_chr}{Names (a character vector)}

\item{rename_lup}{Rename (a lookup table)}

\item{invert_1L_lgl}{Invert (a logical vector of length one), Default: F}
}
\value{
New names (a character vector)
}
\description{
transform_names() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform names. Function argument names_chr specifies the object to be updated. Argument rename_lup provides the object to be updated. The function returns New names (a character vector).
}
\keyword{internal}
