% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_nms_in_mdl_tbl}
\alias{transform_nms_in_mdl_tbl}
\title{Transform names in model table}
\usage{
transform_nms_in_mdl_tbl(
  mdl_tbl_tb,
  col_nm_1L_chr = "Parameter",
  var_nm_change_lup = NULL
)
}
\arguments{
\item{mdl_tbl_tb}{Model table (a tibble)}

\item{col_nm_1L_chr}{Column name (a character vector of length one), Default: 'Parameter'}

\item{var_nm_change_lup}{Variable name change (a lookup table), Default: NULL}
}
\value{
Transformed model table (a tibble)
}
\description{
transform_nms_in_mdl_tbl() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform names in model table. Function argument mdl_tbl_tb specifies the object to be updated. Argument col_nm_1L_chr provides the object to be updated. The function returns Transformed model table (a tibble).
}
\keyword{internal}
