% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_params_ls_to_valid}
\alias{transform_params_ls_to_valid}
\title{Transform parameters list to valid}
\usage{
transform_params_ls_to_valid(
  params_ls,
  scndry_analysis_extra_vars_chr = NA_character_
)
}
\arguments{
\item{params_ls}{Parameters (a list)}

\item{scndry_analysis_extra_vars_chr}{Secondary analysis extra variables (a character vector), Default: 'NA'}
}
\value{
Valid parameters (a list of lists)
}
\description{
transform_params_ls_to_valid() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform parameters list to valid. Function argument params_ls specifies the object to be updated. Argument scndry_analysis_extra_vars_chr provides the object to be updated. The function returns Valid parameters (a list of lists).
}
\keyword{internal}
