% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_predr_nm_part_of_phrases}
\alias{transform_predr_nm_part_of_phrases}
\title{Transform predictor name part of phrases}
\usage{
transform_predr_nm_part_of_phrases(
  phrases_chr,
  old_nms_chr = NULL,
  new_nms_chr = NULL
)
}
\arguments{
\item{phrases_chr}{Phrases (a character vector)}

\item{old_nms_chr}{Old names (a character vector), Default: NULL}

\item{new_nms_chr}{New names (a character vector), Default: NULL}
}
\value{
Transformed phrases (a character vector)
}
\description{
transform_predr_nm_part_of_phrases() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform predictor name part of phrases. Function argument phrases_chr specifies the object to be updated. Argument old_nms_chr provides the object to be updated. The function returns Transformed phrases (a character vector).
}
\keyword{internal}
