% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_raw_aqol_tb_to_aqol6d_tb}
\alias{transform_raw_aqol_tb_to_aqol6d_tb}
\title{Transform raw aqol tibble to aqol6d}
\usage{
transform_raw_aqol_tb_to_aqol6d_tb(raw_aqol_tb)
}
\arguments{
\item{raw_aqol_tb}{Raw aqol (a tibble)}
}
\value{
Aqol6d (a tibble)
}
\description{
transform_raw_aqol_tb_to_aqol6d_tb() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform raw aqol a tibble to aqol6d. Function argument raw_aqol_tb specifies the object to be updated. Argument NA provides the object to be updated. The function returns aqol6d (a tibble).
}
\keyword{internal}
