% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_rprt_lup}
\alias{transform_rprt_lup}
\title{Transform report}
\usage{
transform_rprt_lup(
  rprt_lup,
  add_suplry_rprt_1L_lgl = T,
  add_sharing_rprt_1L_lgl = F
)
}
\arguments{
\item{rprt_lup}{Report (a lookup table)}

\item{add_suplry_rprt_1L_lgl}{Add suplry report (a logical vector of length one), Default: T}

\item{add_sharing_rprt_1L_lgl}{Add sharing report (a logical vector of length one), Default: F}
}
\value{
Report (a lookup table)
}
\description{
transform_rprt_lup() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform report lookup table. Function argument rprt_lup specifies the object to be updated. Argument add_suplry_rprt_1L_lgl provides the object to be updated. The function returns Report (a lookup table).
}
\keyword{internal}
