% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_tb_for_merged_col_1}
\alias{transform_tb_for_merged_col_1}
\title{Transform tibble for merged column 1}
\usage{
transform_tb_for_merged_col_1(df, output_type_1L_chr = "PDF")
}
\arguments{
\item{df}{Data.frame (a data.frame)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'PDF'}
}
\value{
Data.frame (a data.frame)
}
\description{
transform_tb_for_merged_col_1() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform tibble for merged column 1. Function argument df specifies the object to be updated. Argument output_type_1L_chr provides the object to be updated. The function returns Data.frame (a data.frame).
}
\keyword{internal}
