% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_tb_to_mdl_inp}
\alias{transform_tb_to_mdl_inp}
\title{Transform tibble to model input}
\usage{
transform_tb_to_mdl_inp(
  data_tb,
  dep_var_nm_1L_chr = "utl_total_w",
  predr_vars_nms_chr,
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  drop_all_msng_1L_lgl = T,
  scaling_fctr_1L_dbl = 0.01,
  ungroup_1L_lgl = F,
  add_cll_tfmn_1L_lgl = T
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'utl_total_w'}

\item{predr_vars_nms_chr}{Predr vars names (a character vector)}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round var name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round bl value (a character vector of length one), Default: 'Baseline'}

\item{drop_all_msng_1L_lgl}{Drop all msng (a logical vector of length one), Default: T}

\item{scaling_fctr_1L_dbl}{Scaling fctr (a double vector of length one), Default: 0.01}

\item{ungroup_1L_lgl}{Ungroup (a logical vector of length one), Default: F}

\item{add_cll_tfmn_1L_lgl}{Add cll transformation (a logical vector of length one), Default: T}
}
\value{
Transformed for model input (a tibble)
}
\description{
transform_tb_to_mdl_inp() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform tibble to model input. Function argument data_tb specifies the object to be updated. Argument dep_var_nm_1L_chr provides the object to be updated. The function returns Transformed for model input (a tibble).
}
\keyword{internal}
