% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_tbl_to_rnd_vars}
\alias{transform_tbl_to_rnd_vars}
\title{Transform table to round variables}
\usage{
transform_tbl_to_rnd_vars(ds_tb, nbr_of_digits_1L_int = 2L)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}
}
\value{
Transformed dataset (a tibble)
}
\description{
transform_tbl_to_rnd_vars() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform table to round variables. Function argument ds_tb specifies the object to be updated. Argument nbr_of_digits_1L_int provides the object to be updated. The function returns Transformed dataset (a tibble).
}
\keyword{internal}
