% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_timepoint_vals}
\alias{transform_timepoint_vals}
\title{Transform timepoint values}
\usage{
transform_timepoint_vals(
  timepoint_vals_chr,
  timepoint_levels_chr,
  bl_val_1L_chr
)
}
\arguments{
\item{timepoint_vals_chr}{Timepoint values (a character vector)}

\item{timepoint_levels_chr}{Timepoint levels (a character vector)}

\item{bl_val_1L_chr}{Baseline value (a character vector of length one)}
}
\value{
Timepoint values (a character vector)
}
\description{
transform_timepoint_vals() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform timepoint values. Function argument timepoint_vals_chr specifies the object to be updated. Argument timepoint_levels_chr provides the object to be updated. The function returns Timepoint values (a character vector).
}
\keyword{internal}
