% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_uid_var}
\alias{transform_uid_var}
\title{Transform unique identifier variable}
\usage{
transform_uid_var(
  data_tb,
  id_var_nm_1L_chr,
  rename_tb = NULL,
  old_new_chr = c("old_id_xx", "new_id_int")
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one)}

\item{rename_tb}{Rename (a tibble), Default: NULL}

\item{old_new_chr}{Old new (a character vector), Default: c("old_id_xx", "new_id_int")}
}
\value{
Transformed data (a tibble)
}
\description{
transform_uid_var() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform unique identifier variable. Function argument data_tb specifies the object to be updated. Argument id_var_nm_1L_chr provides the object to be updated. The function returns Transformed data (a tibble).
}
\keyword{internal}
