% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTU-package.R, R/pkg_TTU.R
\docType{package}
\name{TTU-package}
\alias{TTU-package}
\title{TTU: Implement Transfer to Utility Mapping Algorithms}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Modules from the readyforwhatsnext youth mental health economic model (https://readyforwhatsnext.org/) that can be used to develop, assess, report and share utility mapping models. Outputs produced through use of TTU package modules are designed to integrate with health economic models developed with the ready4 framework (https://www.ready4-dev.com/). This development version of the TTU package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Modules from the readyforwhatsnext youth mental health economic model (https://readyforwhatsnext.org/) that can be used to develop, assess, report and share utility mapping models. Outputs produced through use of TTU package modules are designed to integrate with health economic models developed with the ready4 framework (https://www.ready4-dev.com/). This development version of the TTU package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/TTU/}
  \item \url{https://github.com/ready4-dev/TTU}
  \item \url{https://ready4-dev.github.io/ready4/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/TTU/}
  \item \url{https://github.com/ready4-dev/TTU}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [funder]

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
\keyword{internal}
