% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_TTU_predictors_lup.R
\name{TTU_predictors_lup}
\alias{TTU_predictors_lup}
\title{TTU S3 class for candidate predictors lookup table}
\usage{
TTU_predictors_lup(x = make_pt_TTU_predictors_lup())
}
\arguments{
\item{x}{A prototype for the TTU S3 class for candidate predictors lookup table, Default: make_pt_TTU_predictors_lup()}
}
\value{
A validated instance of the TTU S3 class for candidate predictors lookup table
}
\description{
Create a new valid instance of the TTU S3 class for candidate predictors lookup table
}
\details{
TTU S3 class for candidate predictors lookup table
}
