% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_TTUParameters.R
\docType{class}
\name{TTUParameters-class}
\alias{TTUParameters-class}
\alias{TTUParameters}
\title{TTUParameters}
\description{
Input Parameters For Generating Utility Mapping Models.
}
\section{Slots}{

\describe{
\item{\code{a_ScorzProfile}}{(an instance of the ScorzProfile class)}

\item{\code{b_SpecificParameters}}{(an instance of the SpecificParameters class)}

\item{\code{paths_chr}}{Paths (a character vector)}

\item{\code{prefd_mdls_chr}}{Preferred models (a character vector)}

\item{\code{prefd_covars_chr}}{Preferred covariates (a character vector)}

\item{\code{scndry_anlys_params_ls}}{Secondary analysis parameters (a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

