% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_TTUProject.R
\docType{class}
\name{TTUProject-class}
\alias{TTUProject-class}
\alias{TTUProject}
\title{TTUProject}
\description{
Input And Output Data For Undertaking and Reporting Utility Mapping Studies.
}
\section{Slots}{

\describe{
\item{\code{a_ScorzProfile}}{(an instance of the ScorzProfile class)}

\item{\code{b_SpecificParameters}}{(an instance of the SpecificParameters class)}

\item{\code{c_TTUSynopsis}}{(an instance of the TTUSynopsis class)}

\item{\code{d_TTUReports}}{(an instance of the TTUReports class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

