% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_TTUReports.R
\docType{class}
\name{TTUReports-class}
\alias{TTUReports-class}
\alias{TTUReports}
\title{TTUReports}
\description{
Metadata to produce utility mapping study reports.
}
\section{Slots}{

\describe{
\item{\code{a_SpecificSynopsis}}{(an instance of the SpecificSynopsis class)}

\item{\code{catalogue_tmpl_chr}}{Catalogue template (a character vector)}

\item{\code{catalogue_fl_nms_ls}}{Catalogue file names (a list)}

\item{\code{manuscript_tmpl_chr}}{Manuscript template (a character vector)}

\item{\code{manuscript_fl_nms_ls}}{Manuscript file names (a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

