% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_TTUSynopsis.R
\docType{class}
\name{TTUSynopsis-class}
\alias{TTUSynopsis-class}
\alias{TTUSynopsis}
\title{TTUSynopsis}
\description{
Input, Output and Authorship Data For Generating Utility Mapping Study Reports.
}
\section{Slots}{

\describe{
\item{\code{a_Ready4showPaths}}{(an instance of the Ready4showPaths class)}

\item{\code{b_SpecificResults}}{(an instance of the SpecificResults class)}

\item{\code{c_SpecificParameters}}{(an instance of the SpecificParameters class)}

\item{\code{d_YouthvarsProfile}}{(an instance of the YouthvarsProfile class)}

\item{\code{e_Ready4useRepos}}{(an instance of the Ready4useRepos class)}

\item{\code{abstract_args_ls}}{Abstract arguments (a list)}

\item{\code{authors_r3}}{Authors (a ready4 S3)}

\item{\code{background_1L_chr}}{Background (a character vector of length one)}

\item{\code{coi_1L_chr}}{Conflict of interest (a character vector of length one)}

\item{\code{conclusion_1L_chr}}{Conclusion (a character vector of length one)}

\item{\code{correspondences_r3}}{Correspondences (a ready4 S3)}

\item{\code{digits_int}}{Digits (an integer vector)}

\item{\code{ethics_1L_chr}}{Ethics (a character vector of length one)}

\item{\code{fl_nm_1L_chr}}{File name (a character vector of length one)}

\item{\code{figures_in_body_lgl}}{Figures in body (a logical vector)}

\item{\code{funding_1L_chr}}{Funding (a character vector of length one)}

\item{\code{institutes_r3}}{Institutes (a ready4 S3)}

\item{\code{interval_chr}}{Interval (a character vector)}

\item{\code{keywords_chr}}{Keywords (a character vector)}

\item{\code{outp_formats_chr}}{Output formats (a character vector)}

\item{\code{rmd_fl_nms_ls}}{R Markdown file names (a list)}

\item{\code{sample_desc_1L_chr}}{Sample description (a character vector of length one)}

\item{\code{tables_in_body_lgl}}{Tables in body (a logical vector)}

\item{\code{title_1L_chr}}{Title (a character vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

