% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_bads.R
\name{validate_firstbounce_bads}
\alias{validate_firstbounce_bads}
\title{Validate First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores}
\usage{
validate_firstbounce_bads(x)
}
\arguments{
\item{x}{An unvalidated instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores}
}
\value{
A prototpe for First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\description{
Validate an instance of the First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
\details{
First Bounce S3 class for Behavioural Activation for Depression Scale (BADS) scores
}
