% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_gad7.R
\name{validate_firstbounce_gad7}
\alias{validate_firstbounce_gad7}
\title{Validate First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores}
\usage{
validate_firstbounce_gad7(x)
}
\arguments{
\item{x}{An unvalidated instance of the First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores}
}
\value{
A prototpe for First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\description{
Validate an instance of the First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
\details{
First Bounce S3 class for Generalised Anxiety Disorder Scale (GAD-7) scores
}
