% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_k6.R
\name{validate_firstbounce_k6}
\alias{validate_firstbounce_k6}
\title{Validate First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores}
\usage{
validate_firstbounce_k6(x)
}
\arguments{
\item{x}{An unvalidated instance of the First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores}
}
\value{
A prototpe for First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\description{
Validate an instance of the First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
\details{
First Bounce S3 class for Kessler Psychological Distress Scale (K6) - US Scoring System scores
}
