% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_phq9.R
\name{validate_firstbounce_phq9}
\alias{validate_firstbounce_phq9}
\title{Validate First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores}
\usage{
validate_firstbounce_phq9(x)
}
\arguments{
\item{x}{An unvalidated instance of the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores}
}
\value{
A prototpe for First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\description{
Validate an instance of the First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
\details{
First Bounce S3 class for Patient Health Questionnaire (PHQ-9) scores
}
