% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_scared.R
\name{validate_firstbounce_scared}
\alias{validate_firstbounce_scared}
\title{Validate First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
\usage{
validate_firstbounce_scared(x)
}
\arguments{
\item{x}{An unvalidated instance of the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores}
}
\value{
A prototpe for First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\description{
Validate an instance of the First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
\details{
First Bounce S3 class for Screen for Child Anxiety Related Disorders (SCARED) scores
}
