% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_TTU_predictors_lup.R
\name{validate_TTU_predictors_lup}
\alias{validate_TTU_predictors_lup}
\title{Validate TTU predictors lookup table TTU S3 class for candidate predictors lookup table}
\usage{
validate_TTU_predictors_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the TTU S3 class for candidate predictors lookup table}
}
\value{
A prototpe for TTU S3 class for candidate predictors lookup table
}
\description{
Validate an instance of the TTU S3 class for candidate predictors lookup table
}
\details{
TTU S3 class for candidate predictors lookup table
}
