% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_all_alg_outps}
\alias{write_all_alg_outps}
\title{Write all algorithm outputs}
\usage{
write_all_alg_outps(
  scored_data_tb,
  path_to_write_to_1L_chr,
  depnt_var_nm_1L_chr = "utl_total_w",
  candidate_predrs_chr,
  candidate_covar_nms_chr,
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  mdl_types_chr = NA_character_,
  prefd_mdl_types_chr = NA_character_,
  choose_from_pfx_chr = c("GLM", "OLS", "BET"),
  prefd_covars_chr = NA_character_,
  seed_1L_int = 12345,
  folds_1L_int = 10L,
  max_nbr_of_boruta_mdl_runs_int = 300L,
  mdl_types_lup = NULL,
  ds_smry_ls
)
}
\arguments{
\item{scored_data_tb}{Scored data (a tibble)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{candidate_predrs_chr}{Candidate predictors (a character vector)}

\item{candidate_covar_nms_chr}{Candidate covariate names (a character vector)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round baseline value (a character vector of length one), Default: 'Baseline'}

\item{mdl_types_chr}{Model types (a character vector), Default: 'NA'}

\item{prefd_mdl_types_chr}{Preferred model types (a character vector), Default: 'NA'}

\item{choose_from_pfx_chr}{Choose from prefix (a character vector), Default: c("GLM", "OLS", "BET")}

\item{prefd_covars_chr}{Preferred covariates (a character vector), Default: 'NA'}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 12345}

\item{folds_1L_int}{Folds (an integer vector of length one), Default: 10}

\item{max_nbr_of_boruta_mdl_runs_int}{Maximum number of boruta model runs (an integer vector), Default: 300}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}

\item{ds_smry_ls}{Dataset summary (a list)}
}
\value{
Output summary (a list)
}
\description{
write_all_alg_outps() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write all algorithm outputs. The function returns Output summary (a list).
}
