% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_analyses}
\alias{write_analyses}
\title{Write analyses}
\usage{
write_analyses(
  input_params_ls,
  abstract_args_ls = NULL,
  start_at_int = c(2, 1)
)
}
\arguments{
\item{input_params_ls}{Input parameters (a list)}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}

\item{start_at_int}{Start at (an integer vector), Default: c(2, 1)}
}
\description{
write_analyses() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write analyses. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
