% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_box_cox_tfmn}
\alias{write_box_cox_tfmn}
\title{Write box cox transformation}
\usage{
write_box_cox_tfmn(
  data_tb,
  predr_var_nm_1L_chr,
  path_to_write_to_1L_chr,
  depnt_var_nm_1L_chr = "utl_total_w",
  covar_var_nms_chr = NA_character_,
  fl_nm_pfx_1L_chr = "A_RT",
  height_1L_dbl = 6,
  width_1L_dbl = 6,
  start_1L_chr = NULL,
  mdl_types_lup = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{predr_var_nm_1L_chr}{Predictor variable name (a character vector of length one)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{covar_var_nms_chr}{Covariate variable names (a character vector), Default: 'NA'}

\item{fl_nm_pfx_1L_chr}{File name prefix (a character vector of length one), Default: 'A_RT'}

\item{height_1L_dbl}{Height (a double vector of length one), Default: 6}

\item{width_1L_dbl}{Width (a double vector of length one), Default: 6}

\item{start_1L_chr}{Start (a character vector of length one), Default: NULL}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}
}
\value{
Path to plot (a character vector of length one)
}
\description{
write_box_cox_tfmn() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write box cox transformation. The function returns Path to plot (a character vector of length one).
}
\keyword{internal}
