% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_brm_mdl_plt_fl}
\alias{write_brm_mdl_plt_fl}
\title{Write brm mdl plt file}
\usage{
write_brm_mdl_plt_fl(
  plt_fn,
  fn_args_ls = NULL,
  path_to_write_to_1L_chr,
  plt_nm_1L_chr,
  grpx_fn = grDevices::png,
  units_1L_chr = "in",
  width_1L_dbl = 6,
  height_1L_dbl = 6,
  rsl_1L_dbl = 300
)
}
\arguments{
\item{plt_fn}{Plt (a function)}

\item{fn_args_ls}{Function arguments (a list), Default: NULL}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{plt_nm_1L_chr}{Plt name (a character vector of length one)}

\item{grpx_fn}{Grpx (a function), Default: grDevices::png}

\item{units_1L_chr}{Units (a character vector of length one), Default: 'in'}

\item{width_1L_dbl}{Width (a double vector of length one), Default: 6}

\item{height_1L_dbl}{Height (a double vector of length one), Default: 6}

\item{rsl_1L_dbl}{Rsl (a double vector of length one), Default: 300}
}
\value{
Path to plot (a character vector of length one)
}
\description{
write_brm_mdl_plt_fl() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write brm mdl plt file. The function returns Path to plot (a character vector of length one).
}
\keyword{internal}
