% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_brm_model_plts}
\alias{write_brm_model_plts}
\title{Write brm model plts}
\usage{
write_brm_model_plts(
  mdl_ls,
  tfd_data_tb,
  mdl_nm_1L_chr,
  path_to_write_to_1L_chr,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  dep_var_desc_1L_chr = "AQoL-6D utility score",
  round_var_nm_1L_chr = "round",
  tfmn_fn = function(x) {     x },
  units_1L_chr = "in",
  height_dbl = c(rep(6, 2), rep(5, 2)),
  width_dbl = c(rep(6, 2), rep(6, 2)),
  rsl_dbl = rep(300, 4),
  args_ls = NULL,
  seed_1L_dbl = 23456
)
}
\arguments{
\item{mdl_ls}{Mdl (a list)}

\item{tfd_data_tb}{Transformed data (a tibble)}

\item{mdl_nm_1L_chr}{Mdl name (a character vector of length one)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{dep_var_desc_1L_chr}{Dep var description (a character vector of length one), Default: 'AQoL-6D utility score'}

\item{round_var_nm_1L_chr}{Round var name (a character vector of length one), Default: 'round'}

\item{tfmn_fn}{Tfmn (a function), Default: function(x) {
x
}}

\item{units_1L_chr}{Units (a character vector of length one), Default: 'in'}

\item{height_dbl}{Height (a double vector), Default: c(rep(6, 2), rep(5, 2))}

\item{width_dbl}{Width (a double vector), Default: c(rep(6, 2), rep(6, 2))}

\item{rsl_dbl}{Rsl (a double vector), Default: rep(300, 4)}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{seed_1L_dbl}{Seed (a double vector of length one), Default: 23456}
}
\value{
Mdl plts paths (a list)
}
\description{
write_brm_model_plts() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write brm model plts. The function returns Mdl plts paths (a list).
}
