% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_csp_output}
\alias{write_csp_output}
\title{Write csp output}
\usage{
write_csp_output(path_to_CSP_1L_chr)
}
\arguments{
\item{path_to_CSP_1L_chr}{Path to CSP (a character vector of length one)}
}
\description{
write_csp_output() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write csp output. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
