% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_main_oupt_dir}
\alias{write_main_oupt_dir}
\title{Write main oupt directory}
\usage{
write_main_oupt_dir(params_ls = NULL, use_fake_data_1L_lgl = F)
}
\arguments{
\item{params_ls}{Params (a list), Default: NULL}

\item{use_fake_data_1L_lgl}{Use fake data (a logical vector of length one), Default: F}
}
\value{
Paths (a list)
}
\description{
write_main_oupt_dir() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write main oupt directory. The function returns Paths (a list).
}
