% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_main_outp_dir}
\alias{write_main_outp_dir}
\title{Write main output directory}
\usage{
write_main_outp_dir(
  params_ls = NULL,
  use_fake_data_1L_lgl = F,
  R_fl_nm_1L_chr = "aaaaaaaaaa.txt"
)
}
\arguments{
\item{params_ls}{Parameters (a list), Default: NULL}

\item{use_fake_data_1L_lgl}{Use fake data (a logical vector of length one), Default: F}

\item{R_fl_nm_1L_chr}{R file name (a character vector of length one), Default: 'aaaaaaaaaa.txt'}
}
\value{
Paths (a list)
}
\description{
write_main_outp_dir() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write main output directory. The function returns Paths (a list).
}
\keyword{internal}
