% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_manuscript}
\alias{write_manuscript}
\title{Write manuscript}
\usage{
write_manuscript(
  results_ls,
  output_format_ls,
  path_params_ls,
  append_params_ls = NULL,
  fl_nm_1L_chr = "TTU_Manuscript",
  path_to_RMDs_chr = c("../Manuscript/Word/Word.Rmd", "../Manuscript/PDF/PDF.Rmd")
)
}
\arguments{
\item{results_ls}{Results (a list)}

\item{output_format_ls}{Output format (a list)}

\item{path_params_ls}{Path params (a list)}

\item{append_params_ls}{Append params (a list), Default: NULL}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'TTU_Manuscript'}

\item{path_to_RMDs_chr}{Path toMDs (a character vector), Default: c("../Manuscript/Word/Word.Rmd", "../Manuscript/PDF/PDF.Rmd")}
}
\description{
write_manuscript() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write manuscript. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
