% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mdl_cmprsns}
\alias{write_mdl_cmprsns}
\title{Write model cmprsns}
\usage{
write_mdl_cmprsns(
  scored_data_tb,
  ds_smry_ls,
  mdl_smry_ls,
  output_data_dir_1L_chr,
  seed_1L_int = 1234
)
}
\arguments{
\item{scored_data_tb}{Scored data (a tibble)}

\item{ds_smry_ls}{Dataset smry (a list)}

\item{mdl_smry_ls}{Model smry (a list)}

\item{output_data_dir_1L_chr}{Output data directory (a character vector of length one)}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1234}
}
\value{
Model cmprsns (a list)
}
\description{
write_mdl_cmprsns() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write model cmprsns. The function returns Model cmprsns (a list).
}
