% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mdl_plts}
\alias{write_mdl_plts}
\title{Write model plts}
\usage{
write_mdl_plts(
  data_tb,
  model_mdl,
  mdl_fl_nm_1L_chr = "OLS_NTF",
  dep_var_nm_1L_chr = "utl_total_w",
  tfmn_1L_chr = "NTF",
  predr_var_nm_1L_chr,
  predr_var_desc_1L_chr,
  predr_vals_dbl,
  covar_var_nms_chr = NA_character_,
  path_to_write_to_1L_chr,
  pred_type_1L_chr = NULL,
  tfmn_for_bnml_1L_lgl = F,
  family_1L_chr = NA_character_,
  plt_idcs_int = 1:5
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{model_mdl}{PARAM_DESCRIPTION}

\item{mdl_fl_nm_1L_chr}{Model file name (a character vector of length one), Default: 'OLS_NTF'}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'utl_total_w'}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{predr_var_nm_1L_chr}{Predr var name (a character vector of length one)}

\item{predr_var_desc_1L_chr}{Predr var description (a character vector of length one)}

\item{predr_vals_dbl}{Predr values (a double vector)}

\item{covar_var_nms_chr}{Covar var names (a character vector), Default: 'NA'}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{pred_type_1L_chr}{Pred type (a character vector of length one), Default: NULL}

\item{tfmn_for_bnml_1L_lgl}{Transformation for bnml (a logical vector of length one), Default: F}

\item{family_1L_chr}{Family (a character vector of length one), Default: 'NA'}

\item{plt_idcs_int}{Plt idcs (an integer vector), Default: 1:5}
}
\description{
write_mdl_plts() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write model plts. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
