% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mdl_smry_rprt}
\alias{write_mdl_smry_rprt}
\title{Write model summary report}
\usage{
write_mdl_smry_rprt(
  header_yaml_args_ls,
  path_params_ls,
  use_fake_data_1L_lgl = FALSE,
  output_format_ls,
  abstract_args_ls = NULL,
  dv_ls = NULL,
  reference_int = 0,
  rprt_lup = NULL,
  rcrd_nm_1L_chr = "AAA_RPRT_WRTNG_MTH",
  rprt_nm_1L_chr = "AAA_TTU_MDL_CTG",
  start_at_int = c(2, 1),
  use_shareable_mdls_1L_lgl = F
)
}
\arguments{
\item{header_yaml_args_ls}{Header yaml arguments (a list)}

\item{path_params_ls}{Path params (a list)}

\item{use_fake_data_1L_lgl}{Use fake data (a logical vector of length one), Default: FALSE}

\item{output_format_ls}{Output format (a list)}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}

\item{dv_ls}{Dataverse (a list), Default: NULL}

\item{reference_int}{Reference (an integer vector), Default: 0}

\item{rprt_lup}{Report (a lookup table), Default: NULL}

\item{rcrd_nm_1L_chr}{Rcrd name (a character vector of length one), Default: 'AAA_RPRT_WRTNG_MTH'}

\item{rprt_nm_1L_chr}{Report name (a character vector of length one), Default: 'AAA_TTU_MDL_CTG'}

\item{start_at_int}{Start at (an integer vector), Default: c(2, 1)}

\item{use_shareable_mdls_1L_lgl}{Use shareable models (a logical vector of length one), Default: F}
}
\value{
Report lups (a list)
}
\description{
write_mdl_smry_rprt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write model summary report. The function returns Report lups (a list).
}
