% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mdl_type_covars_mdls}
\alias{write_mdl_type_covars_mdls}
\title{Write mdl type covars mdls}
\usage{
write_mdl_type_covars_mdls(
  data_tb,
  predrs_var_nms_chr,
  covar_var_nms_chr,
  mdl_type_1L_chr,
  path_to_write_to_1L_chr,
  fl_nm_pfx_1L_chr = "D_CT",
  mdl_types_lup = NULL,
  start_1L_chr = NA_character_
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{predrs_var_nms_chr}{Predrs var names (a character vector)}

\item{covar_var_nms_chr}{Covar var names (a character vector)}

\item{mdl_type_1L_chr}{Mdl type (a character vector of length one)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{fl_nm_pfx_1L_chr}{File name prefix (a character vector of length one), Default: 'D_CT'}

\item{mdl_types_lup}{Mdl types (a lookup table), Default: NULL}

\item{start_1L_chr}{Start (a character vector of length one), Default: 'NA'}
}
\value{
Smry of mdls with covars (a tibble)
}
\description{
write_mdl_type_covars_mdls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write mdl type covars mdls. The function returns Smry of mdls with covars (a tibble).
}
