% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mdl_type_sngl_outps}
\alias{write_mdl_type_sngl_outps}
\title{Write model type single outputs}
\usage{
write_mdl_type_sngl_outps(
  data_tb,
  folds_1L_int = 10,
  depnt_var_nm_1L_chr = "utl_total_w",
  start_1L_chr = NULL,
  tfmn_1L_chr = "NTF",
  predr_var_nm_1L_chr,
  predr_var_desc_1L_chr,
  predr_vals_dbl,
  covar_var_nms_chr = NA_character_,
  mdl_type_1L_chr = "OLS_NTF",
  mdl_types_lup = NULL,
  path_to_write_to_1L_chr,
  mdl_fl_nm_1L_chr,
  plt_idxs_int = NA_integer_
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{folds_1L_int}{Folds (an integer vector of length one), Default: 10}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{start_1L_chr}{Start (a character vector of length one), Default: NULL}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{predr_var_nm_1L_chr}{Predictor variable name (a character vector of length one)}

\item{predr_var_desc_1L_chr}{Predictor variable description (a character vector of length one)}

\item{predr_vals_dbl}{Predictor values (a double vector)}

\item{covar_var_nms_chr}{Covariate variable names (a character vector), Default: 'NA'}

\item{mdl_type_1L_chr}{Model type (a character vector of length one), Default: 'OLS_NTF'}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{mdl_fl_nm_1L_chr}{Model file name (a character vector of length one)}

\item{plt_idxs_int}{Plot indices (an integer vector), Default: NA}
}
\value{
Summary of one predictor model (a tibble)
}
\description{
write_mdl_type_sngl_outps() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write model type single outputs. The function returns Summary of one predictor model (a tibble).
}
\keyword{internal}
