% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mdls_with_covars_cmprsn}
\alias{write_mdls_with_covars_cmprsn}
\title{Write models with covariates comparison}
\usage{
write_mdls_with_covars_cmprsn(
  scored_data_tb,
  bl_tb,
  ds_smry_ls,
  mdl_smry_ls,
  output_data_dir_1L_chr,
  seed_1L_int = 1234,
  session_data_ls = NULL
)
}
\arguments{
\item{scored_data_tb}{Scored data (a tibble)}

\item{bl_tb}{Baseline (a tibble)}

\item{ds_smry_ls}{Dataset summary (a list)}

\item{mdl_smry_ls}{Model summary (a list)}

\item{output_data_dir_1L_chr}{Output data directory (a character vector of length one)}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1234}

\item{session_data_ls}{Session data (a list), Default: NULL}
}
\value{
Output summary (a list)
}
\description{
write_mdls_with_covars_cmprsn() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write models with covariates comparison. The function returns Output summary (a list).
}
\keyword{internal}
