% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_new_outp_dir}
\alias{write_new_outp_dir}
\title{Write new output directory}
\usage{
write_new_outp_dir(path_to_write_to_1L_chr, new_dir_nm_1L_chr)
}
\arguments{
\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one)}
}
\value{
Output directory (a character vector of length one)
}
\description{
write_new_outp_dir() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write new output directory. The function returns Output directory (a character vector of length one).
}
\keyword{internal}
