% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_predr_and_covars_cmprsn}
\alias{write_predr_and_covars_cmprsn}
\title{Write predictor and covariates comparison}
\usage{
write_predr_and_covars_cmprsn(
  scored_data_tb,
  bl_tb,
  ds_smry_ls,
  mdl_smry_ls,
  output_data_dir_1L_chr,
  seed_1L_int = 1234
)
}
\arguments{
\item{scored_data_tb}{Scored data (a tibble)}

\item{bl_tb}{Baseline (a tibble)}

\item{ds_smry_ls}{Dataset summary (a list)}

\item{mdl_smry_ls}{Model summary (a list)}

\item{output_data_dir_1L_chr}{Output data directory (a character vector of length one)}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1234}
}
\value{
Predictor and covariates comparison (a list)
}
\description{
write_predr_and_covars_cmprsn() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write predictor and covariates comparison. The function returns Predictor and covariates comparison (a list).
}
\keyword{internal}
