% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_predr_cmprsn_outps}
\alias{write_predr_cmprsn_outps}
\title{Write predr cmprsn outputs}
\usage{
write_predr_cmprsn_outps(
  data_tb,
  path_to_write_to_1L_chr,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  candidate_predrs_chr,
  max_nbr_of_boruta_mdl_runs_int = 300L
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{candidate_predrs_chr}{Candidate predrs (a character vector)}

\item{max_nbr_of_boruta_mdl_runs_int}{Max nbr of boruta mdl runs (an integer vector), Default: 300}
}
\value{
Confirmed predrs (a tibble)
}
\description{
write_predr_cmprsn_outps() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write predr cmprsn outputs. The function returns Confirmed predrs (a tibble).
}
\keyword{internal}
