% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_report}
\alias{write_report}
\title{Write report}
\usage{
write_report(
  params_ls,
  paths_ls,
  rprt_nm_1L_chr,
  abstract_args_ls = NULL,
  header_yaml_args_ls = NULL,
  rprt_lup = NULL
)
}
\arguments{
\item{params_ls}{Parameters (a list)}

\item{paths_ls}{Paths (a list)}

\item{rprt_nm_1L_chr}{Report name (a character vector of length one)}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}

\item{header_yaml_args_ls}{Header yaml arguments (a list), Default: NULL}

\item{rprt_lup}{Report (a lookup table), Default: NULL}
}
\description{
write_report() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write report. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
