% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_reporting_dir}
\alias{write_reporting_dir}
\title{Write reporting directory}
\usage{
write_reporting_dir(
  path_to_write_to_1L_chr = getwd(),
  new_dir_nm_1L_chr = "TTU_Project",
  overwrite_1L_lgl = FALSE
)
}
\arguments{
\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one), Default: getwd()}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'TTU_Project'}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: FALSE}
}
\value{
Path to complete study program (a character vector of length one)
}
\description{
write_reporting_dir() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write reporting directory. The function returns Path to complete study program (a character vector of length one).
}
