% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_results_csv}
\alias{write_results_csv}
\title{Write results comma separated variables file}
\usage{
write_results_csv(synth_data_spine_ls, output_dir_1L_chr = ".")
}
\arguments{
\item{synth_data_spine_ls}{Synth data spine (a list)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one), Default: '.'}
}
\description{
write_results_csv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write results comma separated variables file. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
