% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_results_to_csv}
\alias{write_results_to_csv}
\title{Write results to comma separated variables file}
\usage{
write_results_to_csv(synth_data_spine_ls, output_dir_1L_chr = ".")
}
\arguments{
\item{synth_data_spine_ls}{Synthetic data spine (a list)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one), Default: '.'}
}
\value{
Datasets (a tibble)
}
\description{
write_results_to_csv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write results to comma separated variables file. The function returns Datasets (a tibble).
}
\keyword{internal}
