% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_rndrd_rprt}
\alias{write_rndrd_rprt}
\title{Write rndrd rprt}
\usage{
write_rndrd_rprt(
  rprt_type_ls,
  params_ls = list(output_type_1L_chr = "HTML"),
  paths_to_fls_to_copy_chr = NA_character_,
  path_to_write_dirs_to_1L_chr = NA_character_,
  nm_of_mkdn_dir_1L_chr = "Markdown",
  path_to_rprt_dir_1L_chr = "./",
  overwrite_1L_lgl = T
)
}
\arguments{
\item{rprt_type_ls}{Rprt type (a list)}

\item{params_ls}{Params (a list), Default: list(output_type_1L_chr = "HTML")}

\item{paths_to_fls_to_copy_chr}{Paths to files to copy (a character vector), Default: 'NA'}

\item{path_to_write_dirs_to_1L_chr}{Path to write directories to (a character vector of length one), Default: 'NA'}

\item{nm_of_mkdn_dir_1L_chr}{Name of mkdn directory (a character vector of length one), Default: 'Markdown'}

\item{path_to_rprt_dir_1L_chr}{Path to rprt directory (a character vector of length one), Default: './'}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: T}
}
\description{
write_rndrd_rprt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write rndrd rprt. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
