% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_rprt}
\alias{write_rprt}
\title{Write rprt}
\usage{
write_rprt(
  outp_smry_ls,
  nm_of_RMD_1L_chr = "_Mdls_Report.RMD",
  output_type_1L_chr = "PDF",
  section_type_1L_chr = "#",
  reports_dir_1L_chr = "Reports",
  markdown_dir_1L_chr = "Markdown",
  file_nm_1L_chr = "Main_Mdl_Smry"
)
}
\arguments{
\item{outp_smry_ls}{Output smry (a list)}

\item{nm_of_RMD_1L_chr}{Name ofMD (a character vector of length one), Default: '_Mdls_Report.RMD'}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'PDF'}

\item{section_type_1L_chr}{Section type (a character vector of length one), Default: '#'}

\item{reports_dir_1L_chr}{Reports directory (a character vector of length one), Default: 'Reports'}

\item{markdown_dir_1L_chr}{Markdown directory (a character vector of length one), Default: 'Markdown'}

\item{file_nm_1L_chr}{File name (a character vector of length one), Default: 'Main_Mdl_Smry'}
}
\value{
Output smry (a list)
}
\description{
write_rprt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write rprt. The function returns Output smry (a list).
}
