% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_rprt_with_rcrd}
\alias{write_rprt_with_rcrd}
\title{Write report with rcrd}
\usage{
write_rprt_with_rcrd(
  path_to_outp_fl_1L_chr,
  paths_ls,
  R_fl_nm_1L_chr,
  header_yaml_args_ls = NULL,
  rprt_lup,
  use_fake_data_1L_lgl,
  subtitle_1L_chr = "Results Supplementary Report 1: Catalogue of time series models",
 
    rprt_subtitle_1L_chr = "Methods Supplementary Report 2: Record of auto-generation of model catalogue.",
  rprt_nm_1L_chr = "TS_TTU_Mdls_Smry",
  rcrd_nm_1L_chr = "Write_Rprt_Rcrd",
  output_type_1L_chr = "PDF",
  rprt_output_type_1L_chr = "PDF",
  nbr_of_digits_1L_int = 2L,
  abstract_args_ls = NULL
)
}
\arguments{
\item{path_to_outp_fl_1L_chr}{Path to output file (a character vector of length one)}

\item{paths_ls}{Paths (a list)}

\item{R_fl_nm_1L_chr}{R file name (a character vector of length one)}

\item{header_yaml_args_ls}{Header yaml arguments (a list), Default: NULL}

\item{rprt_lup}{Report (a lookup table)}

\item{use_fake_data_1L_lgl}{Use fake data (a logical vector of length one)}

\item{subtitle_1L_chr}{Subtitle (a character vector of length one), Default: 'Results Supplementary Report 1: Catalogue of time series models'}

\item{rprt_subtitle_1L_chr}{Report subtitle (a character vector of length one), Default: 'Methods Supplementary Report 2: Record of auto-generation of model catalogue.'}

\item{rprt_nm_1L_chr}{Report name (a character vector of length one), Default: 'TS_TTU_Mdls_Smry'}

\item{rcrd_nm_1L_chr}{Rcrd name (a character vector of length one), Default: 'Write_Rprt_Rcrd'}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'PDF'}

\item{rprt_output_type_1L_chr}{Report output type (a character vector of length one), Default: 'PDF'}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}
}
\description{
write_rprt_with_rcrd() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write report with rcrd. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
