% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_shareable_mdls}
\alias{write_shareable_mdls}
\title{Write shareable mdls}
\usage{
write_shareable_mdls(outp_smry_ls)
}
\arguments{
\item{outp_smry_ls}{Output smry (a list)}
}
\value{
Output smry (a list)
}
\description{
write_shareable_mdls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write shareable mdls. The function returns Output smry (a list).
}
