% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_shareable_mdls_to_dv}
\alias{write_shareable_mdls_to_dv}
\title{Write shareable models to dataverse}
\usage{
write_shareable_mdls_to_dv(
  outp_smry_ls,
  new_dir_nm_1L_chr = "G_Shareable",
  shareable_title_detail_1L_chr = "",
  share_ingredients_1L_lgl = T,
  output_dir_chr = NA_character_
)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'G_Shareable'}

\item{shareable_title_detail_1L_chr}{Shareable title detail (a character vector of length one), Default: ''}

\item{share_ingredients_1L_lgl}{Share ingredients (a logical vector of length one), Default: T}

\item{output_dir_chr}{Output directory (a character vector), Default: 'NA'}
}
\value{
Shareable models (a tibble)
}
\description{
write_shareable_mdls_to_dv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write shareable models to dataverse. The function returns Shareable models (a tibble).
}
\keyword{internal}
