% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_sngl_predr_multi_mdls_outps}
\alias{write_sngl_predr_multi_mdls_outps}
\title{Write sngl predr multi mdls outputs}
\usage{
write_sngl_predr_multi_mdls_outps(
  data_tb,
  mdl_types_chr,
  predr_var_nm_1L_chr,
  predr_var_desc_1L_chr,
  predr_vals_dbl,
  path_to_write_to_1L_chr,
  start_1L_chr = NULL,
  covar_var_nms_chr = NA_character_,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  n_folds_1L_int = 10,
  mdl_types_lup = NULL,
  fl_nm_pfx_1L_chr = "A_RT_",
  plt_idcs_int = NA_integer_
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{mdl_types_chr}{Mdl types (a character vector)}

\item{predr_var_nm_1L_chr}{Predr var name (a character vector of length one)}

\item{predr_var_desc_1L_chr}{Predr var description (a character vector of length one)}

\item{predr_vals_dbl}{Predr values (a double vector)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{start_1L_chr}{Start (a character vector of length one), Default: NULL}

\item{covar_var_nms_chr}{Covar var names (a character vector), Default: 'NA'}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{n_folds_1L_int}{N folds (an integer vector of length one), Default: 10}

\item{mdl_types_lup}{Mdl types (a lookup table), Default: NULL}

\item{fl_nm_pfx_1L_chr}{File name prefix (a character vector of length one), Default: 'A_RT_'}

\item{plt_idcs_int}{Plt idcs (an integer vector), Default: NA}
}
\value{
Smry of sngl predr mdls (a tibble)
}
\description{
write_sngl_predr_multi_mdls_outps() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write sngl predr multi mdls outputs. The function returns Smry of sngl predr mdls (a tibble).
}
\keyword{internal}
