% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_delete_ds_copies}
\alias{write_to_delete_ds_copies}
\title{Write to delete dataset copies}
\usage{
write_to_delete_ds_copies(input_params_ls = NULL, paths_ls = NULL)
}
\arguments{
\item{input_params_ls}{Input parameters (a list), Default: NULL}

\item{paths_ls}{Paths (a list), Default: NULL}
}
\description{
write_to_delete_ds_copies() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to delete dataset copies. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
