% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_delete_mdl_fls}
\alias{write_to_delete_mdl_fls}
\title{Write to delete model files}
\usage{
write_to_delete_mdl_fls(outp_smry_ls)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}
}
\description{
write_to_delete_mdl_fls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to delete model files. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
