% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ts_mdls}
\alias{write_ts_mdls}
\title{Write ts mdls}
\usage{
write_ts_mdls(
  data_tb,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  predr_vars_nms_ls,
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  fn_ls,
  mdl_nms_ls,
  mdl_smry_dir_1L_chr,
  iters_1L_int = 4000L,
  seed_1L_int = 1000L
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{predr_vars_nms_ls}{Predr vars names (a list)}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round var name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round bl value (a character vector of length one), Default: 'Baseline'}

\item{fn_ls}{Function list (a list of functions)}

\item{mdl_nms_ls}{Mdl names (a list)}

\item{mdl_smry_dir_1L_chr}{Mdl smry directory (a character vector of length one)}

\item{iters_1L_int}{Iters (an integer vector of length one), Default: 4000}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1000}
}
\value{
Mdls smry (a tibble)
}
\description{
write_ts_mdls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write ts mdls. The function returns Mdls smry (a tibble).
}
