% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol6dU_to_tbs_ls}
\alias{add_aqol6dU_to_tbs_ls}
\title{Add aqol6dU to tibbles}
\usage{
add_aqol6dU_to_tbs_ls(tbs_ls, prefix_1L_chr = "aqol6d_q", id_var_nm_1L_chr)
}
\arguments{
\item{tbs_ls}{Tibbles (a list)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'aqol6d_q'}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one)}
}
\value{
Tibbles (a list)
}
\description{
add_aqol6dU_to_tbs_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add aqol6du to tibbles list. Function argument tbs_ls specifies the object to be updated. The function returns Tibbles (a list).
}
