% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol6dU_to_tbs_ls}
\alias{add_aqol6dU_to_tbs_ls}
\title{Add aqol6dU to tibbles}
\usage{
add_aqol6dU_to_tbs_ls(
  tbs_ls,
  prefix_1L_chr = "aqol6d_q",
  aqol6d_from_8d_coeffs_lup_tb = aqol6d_from_8d_coeffs_lup_tb,
  dim_sclg_constant_lup_tb = dim_sclg_constant_lup_tb,
  disutilities_lup_tb = disutilities_lup_tb,
  itm_wrst_wghts_lup_tb = itm_wrst_wghts_lup_tb
)
}
\arguments{
\item{tbs_ls}{Tibbles (a list)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'aqol6d_q'}

\item{aqol6d_from_8d_coeffs_lup_tb}{Assessment of Quality of Life Six Dimension health utility from 8d coeffs lookup table (a tibble), Default: aqol6d_from_8d_coeffs_lup_tb}

\item{dim_sclg_constant_lup_tb}{Dimension sclg constant lookup table (a tibble), Default: dim_sclg_constant_lup_tb}

\item{disutilities_lup_tb}{Disutilities lookup table (a tibble), Default: disutilities_lup_tb}

\item{itm_wrst_wghts_lup_tb}{Itm wrst wghts lookup table (a tibble), Default: itm_wrst_wghts_lup_tb}
}
\value{
Tibbles (a list)
}
\description{
add_aqol6dU_to_tbs_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add aqol6du to tibbles list. Function argument tbs_ls specifies the object to be updated. The function returns Tibbles (a list).
}
