% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol_items_tbs_ls}
\alias{add_aqol_items_tbs_ls}
\title{Add aqol items tibbles}
\usage{
add_aqol_items_tbs_ls(tbs_ls, aqol_items_props_tbs_ls, prefix_1L_chr)
}
\arguments{
\item{tbs_ls}{Tibbles (a list)}

\item{aqol_items_props_tbs_ls}{Aqol items props tibbles (a list)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}
}
\value{
Updated tibbles (a list)
}
\description{
add_aqol_items_tbs_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add aqol items tibbles list. Function argument tbs_ls specifies the object to be updated. The function returns Updated tibbles (a list).
}
