% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_corrs_and_uts_to_tbs_ls_ls}
\alias{add_corrs_and_uts_to_tbs_ls_ls}
\title{Add corrs and uts to tibbles}
\usage{
add_corrs_and_uts_to_tbs_ls_ls(tbs_ls, temporal_corrs_ls, prefix_chr)
}
\arguments{
\item{tbs_ls}{Tibbles (a list)}

\item{temporal_corrs_ls}{Temporal corrs (a list)}

\item{prefix_chr}{Prefix (a character vector)}
}
\value{
Tibbles (a list)
}
\description{
add_corrs_and_uts_to_tbs_ls_ls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add corrs and uts to tibbles list list. Function argument tbs_ls specifies the object to be updated. The function returns Tibbles (a list).
}
