% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_dmn_disu_to_aqol6d_items_tb_tb}
\alias{add_dmn_disu_to_aqol6d_items_tb_tb}
\title{Add dmn disu to aqol6d items tibble}
\usage{
add_dmn_disu_to_aqol6d_items_tb_tb(
  aqol6d_items_tb,
  domain_items_ls,
  domains_chr,
  dim_sclg_constant_lup_tb = dim_sclg_constant_lup_tb,
  itm_wrst_wghts_lup_tb = itm_wrst_wghts_lup_tb
)
}
\arguments{
\item{aqol6d_items_tb}{Aqol6d items (a tibble)}

\item{domain_items_ls}{Domain items (a list)}

\item{domains_chr}{Domains (a character vector)}

\item{dim_sclg_constant_lup_tb}{Dim sclg constant lookup table (a tibble), Default: dim_sclg_constant_lup_tb}

\item{itm_wrst_wghts_lup_tb}{Itm wrst wghts lookup table (a tibble), Default: itm_wrst_wghts_lup_tb}
}
\value{
Aqol6d items (a tibble)
}
\description{
add_dmn_disu_to_aqol6d_items_tb_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add dmn disu to aqol6d items tibble tibble. Function argument aqol6d_items_tb specifies the object to be updated. The function returns Aqol6d items (a tibble).
}
