% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_labels_to_aqol6d_tb}
\alias{add_labels_to_aqol6d_tb}
\title{Add labels to Assessment of Quality of Life Six Dimension}
\usage{
add_labels_to_aqol6d_tb(aqol6d_tb, labels_chr = NA_character_)
}
\arguments{
\item{aqol6d_tb}{Assessment of Quality of Life Six Dimension (a tibble)}

\item{labels_chr}{Labels (a character vector), Default: 'NA'}
}
\value{
Assessment of Quality of Life Six Dimension (a tibble)
}
\description{
add_labels_to_aqol6d_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add labels to assessment of quality of life six dimension tibble. Function argument aqol6d_tb specifies the object to be updated. The function returns Assessment of Quality of Life Six Dimension (a tibble).
}
\keyword{internal}
