% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_utility_predn_to_ds}
\alias{add_utility_predn_to_ds}
\title{Add utility predn to dataset}
\usage{
add_utility_predn_to_ds(data_tb, mdl, tfmn_1L_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{tfmn_1L_chr}{Tfmn (a character vector of length one)}

\item{...}{Additional arguments}
}
\value{
Data (a tibble)
}
\description{
add_utility_predn_to_ds() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add utility predn to dataset. Function argument data_tb specifies the object to be updated. The function returns Data (a tibble).
}
\keyword{internal}
