% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_utility_predn_to_ds}
\alias{add_utility_predn_to_ds}
\title{Add utility prediction to dataset}
\usage{
add_utility_predn_to_ds(
  data_tb,
  model_mdl,
  tfmn_1L_chr,
  depnt_var_nm_1L_chr,
  force_min_max_1L_lgl = T,
  force_new_data_1L_lgl = F,
  impute_1L_lgl = T,
  is_brms_mdl_1L_lgl = T,
  new_data_is_1L_chr = "Predicted",
  predn_type_1L_chr = NULL,
  predr_vars_nms_chr = NULL,
  rmv_tfd_depnt_var_1L_lgl = F,
  sd_dbl = NA_real_,
  utl_cls_fn = NULL,
  utl_min_val_1L_dbl = -1
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{model_mdl}{Model (a model)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one)}

\item{force_min_max_1L_lgl}{Force minimum maximum (a logical vector of length one), Default: T}

\item{force_new_data_1L_lgl}{Force new data (a logical vector of length one), Default: F}

\item{impute_1L_lgl}{Impute (a logical vector of length one), Default: T}

\item{is_brms_mdl_1L_lgl}{Is bayesian regression models model (a logical vector of length one), Default: T}

\item{new_data_is_1L_chr}{New data is (a character vector of length one), Default: 'Predicted'}

\item{predn_type_1L_chr}{Prediction type (a character vector of length one), Default: NULL}

\item{predr_vars_nms_chr}{Predictor variables names (a character vector), Default: NULL}

\item{rmv_tfd_depnt_var_1L_lgl}{Remove transformed dependent variable (a logical vector of length one), Default: F}

\item{sd_dbl}{Standard deviation (a double vector), Default: NA}

\item{utl_cls_fn}{Utility class (a function), Default: NULL}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: -1}
}
\value{
Data (a tibble)
}
\description{
add_utility_predn_to_ds() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add utility prediction to dataset. Function argument data_tb specifies the object to be updated. The function returns Data (a tibble).
}
\keyword{internal}
