% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_aqol6dU_dbl}
\alias{calculate_aqol6dU_dbl}
\title{Calculate aqol6dU}
\usage{
calculate_aqol6dU_dbl(
  aqol6d_items_tb,
  prefix_1L_chr,
  aqol6d_from_8d_coeffs_lup_tb = aqol6d_from_8d_coeffs_lup_tb,
  dim_sclg_constant_lup_tb = dim_sclg_constant_lup_tb,
  disvalues_lup_tb = disvalues_lup_tb,
  itm_wrst_wghts_lup_tb = itm_wrst_wghts_lup_tb
)
}
\arguments{
\item{aqol6d_items_tb}{Assessment of Quality of Life Six Dimension health utility items (a tibble)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}

\item{aqol6d_from_8d_coeffs_lup_tb}{Assessment of Quality of Life Six Dimension health utility from 8d coeffs lookup table (a tibble), Default: aqol6d_from_8d_coeffs_lup_tb}

\item{dim_sclg_constant_lup_tb}{Dimension sclg constant lookup table (a tibble), Default: dim_sclg_constant_lup_tb}

\item{disvalues_lup_tb}{Disvalues lookup table (a tibble), Default: disvalues_lup_tb}

\item{itm_wrst_wghts_lup_tb}{Itm wrst wghts lookup table (a tibble), Default: itm_wrst_wghts_lup_tb}
}
\value{
Aqol6dU (a double vector)
}
\description{
calculate_aqol6dU_dbl() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate aqol6du double vector. The function returns Aqol6dU (a double vector).
}
