% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_aqol6dU_dbl}
\alias{calculate_aqol6dU_dbl}
\title{Calculate aqol6dU}
\usage{
calculate_aqol6dU_dbl(
  aqol6d_items_tb,
  prefix_1L_chr,
  aqol6d_from_8d_coeffs_lup_tb = aqol6d_from_8d_coeffs_lup_tb,
  dim_sclg_constant_lup_tb = dim_sclg_constant_lup_tb,
  disutilities_lup_tb = disutilities_lup_tb,
  itm_wrst_wghts_lup_tb = itm_wrst_wghts_lup_tb
)
}
\arguments{
\item{aqol6d_items_tb}{Aqol6d items (a tibble)}

\item{prefix_1L_chr}{Prefix 1L (a character vector of length 1)}

\item{aqol6d_from_8d_coeffs_lup_tb}{Aqol6d from 8d coeffs lookup table (a tibble), Default: aqol6d_from_8d_coeffs_lup_tb}

\item{dim_sclg_constant_lup_tb}{Dim sclg constant lookup table (a tibble), Default: dim_sclg_constant_lup_tb}

\item{disutilities_lup_tb}{Disutilities lookup table (a tibble), Default: disutilities_lup_tb}

\item{itm_wrst_wghts_lup_tb}{Itm wrst wghts lookup table (a tibble), Default: itm_wrst_wghts_lup_tb}
}
\value{
Aqol6dU (a double vector of length 1)
}
\description{
calculate_aqol6dU_dbl() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate aqol6dU. The function returns aqol6du (a double vector of length 1).
}
\keyword{internal}
