% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{Calculate root mean square error}
\usage{
calculate_rmse(y_dbl, yhat_dbl)
}
\arguments{
\item{y_dbl}{Y (a double vector)}

\item{yhat_dbl}{Yhat (a double vector)}
}
\value{
Root mean square error (a double vector)
}
\description{
calculate_rmse() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate root mean square error. The function returns Root mean square error (a double vector).
}
\keyword{internal}
