% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_rmse_tfmn}
\alias{calculate_rmse_tfmn}
\title{Calculate rmse tfmn}
\usage{
calculate_rmse_tfmn(y_dbl, yhat_dbl)
}
\arguments{
\item{y_dbl}{Y (a double vector)}

\item{yhat_dbl}{Yhat (a double vector)}
}
\value{
Rmse tfmn (a double vector)
}
\description{
calculate_rmse_tfmn() is a Calculate function that calculates a numeric value. Specifically, this function implements an algorithm to calculate rmse tfmn. The function returns Rmse tfmn (a double vector).
}
\keyword{internal}
