% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fit.R
\name{fit_clg_log_tfmn}
\alias{fit_clg_log_tfmn}
\title{Fit clg log tfmn}
\usage{
fit_clg_log_tfmn(
  data_tb,
  dep_var_nm_1L_chr = "aqol6d_total_w_cloglog",
  predictor_vars_nms_chr,
  id_var_nm_1L_chr = "fkClientID",
  iters_1L_int = 4000L,
  seed_1L_int = 1000L
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w_cloglog'}

\item{predictor_vars_nms_chr}{Predictor vars names (a character vector)}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{iters_1L_int}{Iters (an integer vector of length one), Default: 4000}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1000}
}
\value{
Mdl (a list)
}
\description{
fit_clg_log_tfmn() is a Fit function that fits a model of a specified type to a dataset Specifically, this function implements an algorithm to fit clg log tfmn. The function returns Mdl (a list).
}
