% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fit.R
\name{fit_ts_model_with_brm}
\alias{fit_ts_model_with_brm}
\title{Fit time series model with bayesian regression model}
\usage{
fit_ts_model_with_brm(
  data_tb,
  depnt_var_nm_1L_chr,
  predr_vars_nms_chr,
  id_var_nm_1L_chr,
  backend_1L_chr = getOption("brms.backend", "rstan"),
  family_fn_1L_chr,
  iters_1L_int = 4000L,
  seed_1L_int = 1000L,
  prior_ls = NULL,
  control_ls = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one)}

\item{predr_vars_nms_chr}{Predictor variables names (a character vector)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one)}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: getOption("brms.backend", "rstan")}

\item{family_fn_1L_chr}{Family function (a character vector of length one)}

\item{iters_1L_int}{Iterations (an integer vector of length one), Default: 4000}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1000}

\item{prior_ls}{Prior (a list), Default: NULL}

\item{control_ls}{Control (a list), Default: NULL}
}
\value{
Model list (a list of models)
}
\description{
fit_ts_model_with_brm() is a Fit function that fits a model of a specified type to a dataset Specifically, this function implements an algorithm to fit time series model with bayesian regression model. The function returns Model list (a list of models).
}
\keyword{internal}
