% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_force.R
\name{force_vec_to_sum_to_int}
\alias{force_vec_to_sum_to_int}
\title{Force vector to sum to}
\usage{
force_vec_to_sum_to_int(vec_int, target_1L_int, item_ranges_dbl_ls)
}
\arguments{
\item{vec_int}{Vector (an integer vector)}

\item{target_1L_int}{Target (an integer vector of length one)}

\item{item_ranges_dbl_ls}{Item ranges (a list of double vectors)}
}
\value{
Vector (an integer vector)
}
\description{
force_vec_to_sum_to_int() is a Force function that checks if a specified local or global environmental condition is met and if not, updates the specified environment to comply with the condition. Specifically, this function implements an algorithm to force vector to sum to integer vector. The function returns Vector (an integer vector).
}
\keyword{internal}
