% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_impute.R
\name{impute_miss_itms_in_aqol6d_items_tb_tb}
\alias{impute_miss_itms_in_aqol6d_items_tb_tb}
\title{Impute miss itms in aqol6d items tibble}
\usage{
impute_miss_itms_in_aqol6d_items_tb_tb(aqol6d_items_tb, domain_items_ls)
}
\arguments{
\item{aqol6d_items_tb}{Aqol6d items (a tibble)}

\item{domain_items_ls}{Domain items (a list)}
}
\value{
Aqol6d items (a tibble)
}
\description{
impute_miss_itms_in_aqol6d_items_tb_tb() is an Impute function that imputes data. Specifically, this function implements an algorithm to impute miss itms in aqol6d items tibble tibble. The function returns Aqol6d items (a tibble).
}
\keyword{internal}
