% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_knit.R
\name{knit_mdl_rprt}
\alias{knit_mdl_rprt}
\title{Knit model report}
\usage{
knit_mdl_rprt(knit_pars_ls, path_to_mdl_rprt_tmpl_1L_chr)
}
\arguments{
\item{knit_pars_ls}{Knit pars (a list)}

\item{path_to_mdl_rprt_tmpl_1L_chr}{Path to model report template (a character vector of length one)}
}
\description{
knit_mdl_rprt() is a Knit function that knits a rmarkdown file Specifically, this function implements an algorithm to knit model report. The function is called for its side effects and does not return a value.
}
\keyword{internal}
