% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol6d_adol_pop_tbs_ls}
\alias{make_aqol6d_adol_pop_tbs_ls}
\title{Make Assessment of Quality of Life Six Dimension adolescent pop tibbles}
\usage{
make_aqol6d_adol_pop_tbs_ls(
  aqol_items_prpns_tbs_ls,
  aqol_scores_pars_ls,
  series_names_chr,
  synth_data_spine_ls,
  temporal_cors_ls,
  id_var_nm_1L_chr = "fkClientID",
  prefix_chr = c(uid = "Participant_", aqol_item = "aqol6d_q", domain_unwtd_pfx_1L_chr
    = "aqol6d_subtotal_c_", domain_wtd_pfx_1L_chr = "aqol6d_subtotal_w_")
)
}
\arguments{
\item{aqol_items_prpns_tbs_ls}{Assessment of Quality of Life items proportions tibbles (a list)}

\item{aqol_scores_pars_ls}{Assessment of Quality of Life scores parameters (a list)}

\item{series_names_chr}{Series names (a character vector)}

\item{synth_data_spine_ls}{Synthetic data spine (a list)}

\item{temporal_cors_ls}{Temporal correlations (a list)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{prefix_chr}{Prefix (a character vector), Default: c(uid = "Participant_", aqol_item = "aqol6d_q", domain_unwtd_pfx_1L_chr = "aqol6d_subtotal_c_",
domain_wtd_pfx_1L_chr = "aqol6d_subtotal_w_")}
}
\value{
Assessment of Quality of Life Six Dimension adolescent pop tibbles (a list)
}
\description{
make_aqol6d_adol_pop_tbs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make assessment of quality of life six dimension adolescent pop tibbles list. The function returns Assessment of Quality of Life Six Dimension adolescent pop tibbles (a list).
}
\keyword{internal}
