% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol6d_fns_ls}
\alias{make_aqol6d_fns_ls}
\title{Make Assessment of Quality of Life Six Dimension functions}
\usage{
make_aqol6d_fns_ls(domain_items_ls)
}
\arguments{
\item{domain_items_ls}{Domain items (a list)}
}
\value{
Assessment of Quality of Life Six Dimension disu (a list of functions)
}
\description{
make_aqol6d_fns_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make assessment of quality of life six dimension functions list. The function returns Assessment of Quality of Life Six Dimension disu (a list of functions).
}
\keyword{internal}
