% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol6d_items_tb}
\alias{make_aqol6d_items_tb}
\title{Make Assessment of Quality of Life Six Dimension items}
\usage{
make_aqol6d_items_tb(aqol_tb, old_pfx_1L_chr, new_pfx_1L_chr)
}
\arguments{
\item{aqol_tb}{Assessment of Quality of Life (a tibble)}

\item{old_pfx_1L_chr}{Old prefix (a character vector of length one)}

\item{new_pfx_1L_chr}{New prefix (a character vector of length one)}
}
\value{
Assessment of Quality of Life Six Dimension items (a tibble)
}
\description{
make_aqol6d_items_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make assessment of quality of life six dimension items tibble. The function returns Assessment of Quality of Life Six Dimension items (a tibble).
}
\keyword{internal}
