% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_corstars_tbl_xx}
\alias{make_corstars_tbl_xx}
\title{Make corstars table}
\usage{
make_corstars_tbl_xx(
  x,
  method_chr = c("pearson", "spearman"),
  removeTriangle_chr = c("upper", "lower"),
  result_chr = c("none", "html", "latex")
)
}
\arguments{
\item{x}{An object}

\item{method_chr}{Method (a character vector), Default: c("pearson", "spearman")}

\item{removeTriangle_chr}{RemoveTriangle (a character vector), Default: c("upper", "lower")}

\item{result_chr}{Result (a character vector), Default: c("none", "html", "latex")}
}
\description{
make_corstars_tbl_xx() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make corstars table output object of multiple potential types. The function is called for its side effects and does not return a value.
}
\keyword{internal}
