% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cs_ts_ratios_tb}
\alias{make_cs_ts_ratios_tb}
\title{Make cross-section time series ratios tibble}
\usage{
make_cs_ts_ratios_tb(
  predr_ctgs_ls,
  mdl_coef_ratios_ls,
  candidate_predrs_chr = NULL,
  nbr_of_digits_1L_int = 2L,
  fn_ls = NULL
)
}
\arguments{
\item{predr_ctgs_ls}{Predictor category categoriess (a list)}

\item{mdl_coef_ratios_ls}{Model coefficient ratios (a list)}

\item{candidate_predrs_chr}{Candidate predictors (a character vector), Default: NULL}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}

\item{fn_ls}{Function list (a list of functions), Default: NULL}
}
\value{
Cross-section time series ratios (a tibble)
}
\description{
make_cs_ts_ratios_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cross-section time series ratios tibble. The function returns Cross-section time series ratios (a tibble).
}
\keyword{internal}
