% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_dim_sclg_cons_dbl}
\alias{make_dim_sclg_cons_dbl}
\title{Make dimension scaling constants}
\usage{
make_dim_sclg_cons_dbl(domains_chr, dim_sclg_con_lup_tb)
}
\arguments{
\item{domains_chr}{Domains (a character vector)}

\item{dim_sclg_con_lup_tb}{Dimension scaling constant lookup table (a tibble)}
}
\value{
Dimension scaling constants (a double vector)
}
\description{
make_dim_sclg_cons_dbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dimension scaling constants double vector. The function returns Dimension scaling constants (a double vector).
}
\keyword{internal}
