% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fake_eq5d_ds}
\alias{make_fake_eq5d_ds}
\title{Make fake eq5d dataset}
\usage{
make_fake_eq5d_ds(
  country_1L_chr = "UK",
  version_1L_chr = "5L",
  type_1L_chr = "CW",
  prop_with_fup_data_1L_dbl = 0.65,
  seed_1L_int = 1234,
  sample_from_1L_int = 10000
)
}
\arguments{
\item{country_1L_chr}{Country (a character vector of length one), Default: 'UK'}

\item{version_1L_chr}{Version (a character vector of length one), Default: '5L'}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'CW'}

\item{prop_with_fup_data_1L_dbl}{Prop with follow-up data (a double vector of length one), Default: 0.65}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1234}

\item{sample_from_1L_int}{Sample from (an integer vector of length one), Default: 10000}
}
\value{
Data (a tibble)
}
\description{
make_fake_eq5d_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make fake eq5d dataset. The function returns Data (a tibble).
}
