% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fake_ts_data}
\alias{make_fake_ts_data}
\title{Make fake ts data}
\usage{
make_fake_ts_data(outp_smry_ls)
}
\arguments{
\item{outp_smry_ls}{Output smry (a list)}
}
\value{
Fk data (a tibble)
}
\description{
make_fake_ts_data() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make fake ts data. The function returns Fk data (a tibble).
}
\keyword{internal}
