% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_folds_ls}
\alias{make_folds_ls}
\title{Make folds list}
\usage{
make_folds_ls(data_tb, depnt_var_nm_1L_chr = "utl_total_w", folds_1L_int = 10L)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{folds_1L_int}{Folds (an integer vector of length one), Default: 10}
}
\value{
Folds (a list)
}
\description{
make_folds_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make folds list. The function returns Folds (a list).
}
\keyword{internal}
