% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_knit_pars_ls}
\alias{make_knit_pars_ls}
\title{Make knit pars list}
\usage{
make_knit_pars_ls(
  rltv_path_to_data_dir_1L_chr,
  mdl_types_chr,
  predr_vars_nms_ls,
  output_type_1L_chr = "HTML",
  mdl_types_lup = NULL,
  plt_types_lup = NULL,
  plt_types_chr = NA_character_,
  section_type_1L_chr = "#"
)
}
\arguments{
\item{rltv_path_to_data_dir_1L_chr}{Relative path to data directory (a character vector of length one)}

\item{mdl_types_chr}{Model types (a character vector)}

\item{predr_vars_nms_ls}{Predictor variables names (a list)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}

\item{plt_types_lup}{Plot types (a lookup table), Default: NULL}

\item{plt_types_chr}{Plot types (a character vector), Default: 'NA'}

\item{section_type_1L_chr}{Section type (a character vector of length one), Default: '#'}
}
\value{
Knit pars (a list)
}
\description{
make_knit_pars_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make knit pars list. The function returns Knit pars (a list).
}
