% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_lngl_ttu_r2_text}
\alias{make_lngl_ttu_r2_text}
\title{Make longitudinal transfer to utility algorithm squared text}
\usage{
make_lngl_ttu_r2_text(results_ls, part_int = 1)
}
\arguments{
\item{results_ls}{Results (a list)}

\item{part_int}{Part (an integer vector), Default: 1}
}
\value{
Text (a character vector of length one)
}
\description{
make_lngl_ttu_r2_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make longitudinal transfer to utility algorithm squared text. The function returns Text (a character vector of length one).
}
\keyword{internal}
