% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_lngl_ttu_with_covars_text}
\alias{make_lngl_ttu_with_covars_text}
\title{Make longitudinal transfer to utility algorithm with covariates text}
\usage{
make_lngl_ttu_with_covars_text(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Text (a character vector of length one)
}
\description{
make_lngl_ttu_with_covars_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make longitudinal transfer to utility algorithm with covariates text. The function returns Text (a character vector of length one).
}
\keyword{internal}
