% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_nms_ls}
\alias{make_mdl_nms_ls}
\title{Make model names list}
\usage{
make_mdl_nms_ls(predr_vars_nms_ls, mdl_types_chr)
}
\arguments{
\item{predr_vars_nms_ls}{Predictor variables names (a list)}

\item{mdl_types_chr}{Model types (a character vector)}
}
\value{
Model names (a list)
}
\description{
make_mdl_nms_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model names list. The function returns Model names (a list).
}
\keyword{internal}
