% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_predr_vals}
\alias{make_predr_vals}
\title{Make predictor values}
\usage{
make_predr_vals(predr_var_nm_1L_chr, candidate_predrs_lup = NULL)
}
\arguments{
\item{predr_var_nm_1L_chr}{Predictor variable name (a character vector of length one)}

\item{candidate_predrs_lup}{Candidate predictors (a lookup table), Default: NULL}
}
\value{
Predictor values (a double vector)
}
\description{
make_predr_vals() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make predictor values. The function returns Predictor values (a double vector).
}
\keyword{internal}
