% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_predr_vars_nms_ls}
\alias{make_predr_vars_nms_ls}
\title{Make predictor variables names list}
\usage{
make_predr_vars_nms_ls(main_predrs_chr, covars_ls, existing_predrs_ls = NULL)
}
\arguments{
\item{main_predrs_chr}{Main predictors (a character vector)}

\item{covars_ls}{Covariates (a list)}

\item{existing_predrs_ls}{Existing predictors (a list), Default: NULL}
}
\value{
Predictor variables names (a list)
}
\description{
make_predr_vars_nms_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make predictor variables names list. The function returns Predictor variables names (a list).
}
\keyword{internal}
