% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_prefd_mdls_vec}
\alias{make_prefd_mdls_vec}
\title{Make prefd mdls vec}
\usage{
make_prefd_mdls_vec(
  smry_of_sngl_predr_mdls_tb,
  choose_from_pfx_chr = c("GLM", "OLS"),
  mdl_types_lup = NULL
)
}
\arguments{
\item{smry_of_sngl_predr_mdls_tb}{Smry of sngl predr mdls (a tibble)}

\item{choose_from_pfx_chr}{Choose from prefix (a character vector), Default: c("GLM", "OLS")}

\item{mdl_types_lup}{Mdl types (a lookup table), Default: NULL}
}
\value{
Prefd mdls (a character vector)
}
\description{
make_prefd_mdls_vec() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make prefd mdls vec. The function returns Prefd mdls (a character vector).
}
