% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_prefd_mdls_vec}
\alias{make_prefd_mdls_vec}
\title{Make preferred models vector}
\usage{
make_prefd_mdls_vec(
  smry_of_sngl_predr_mdls_tb,
  choose_from_pfx_chr = c("BET", "GLM", "OLS"),
  mdl_types_lup = NULL
)
}
\arguments{
\item{smry_of_sngl_predr_mdls_tb}{Summary of single predictor models (a tibble)}

\item{choose_from_pfx_chr}{Choose from prefix (a character vector), Default: c("BET", "GLM", "OLS")}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}
}
\value{
Preferred models (a character vector)
}
\description{
make_prefd_mdls_vec() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make preferred models vector. The function returns Preferred models (a character vector).
}
\keyword{internal}
