% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_smry_of_brm_mdl}
\alias{make_smry_of_brm_mdl}
\title{Make smry of brm model}
\usage{
make_smry_of_brm_mdl(
  mdl_ls,
  data_tb,
  dep_var_nm_1L_chr = "utl_total_w",
  predr_vars_nms_chr,
  fn = calculate_rmse,
  mdl_nm_1L_chr = NA_character_,
  seed_1L_dbl = 23456
)
}
\arguments{
\item{mdl_ls}{Model (a list)}

\item{data_tb}{Data (a tibble)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'utl_total_w'}

\item{predr_vars_nms_chr}{Predr vars names (a character vector)}

\item{fn}{Function (a function), Default: calculate_rmse}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one), Default: 'NA'}

\item{seed_1L_dbl}{Seed (a double vector of length one), Default: 23456}
}
\value{
Smry of brm model (a tibble)
}
\description{
make_smry_of_brm_mdl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make smry of brm model. The function returns Smry of brm model (a tibble).
}
