% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_smry_of_brm_mdl}
\alias{make_smry_of_brm_mdl}
\title{Make summary of bayesian regression model model}
\usage{
make_smry_of_brm_mdl(
  mdl_ls,
  data_tb,
  depnt_var_nm_1L_chr = "utl_total_w",
  predr_vars_nms_chr,
  mdl_nm_1L_chr = NA_character_,
  seed_1L_dbl = 23456,
  tfmn_1L_chr
)
}
\arguments{
\item{mdl_ls}{Model list (a list of models)}

\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{predr_vars_nms_chr}{Predictor variables names (a character vector)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one), Default: 'NA'}

\item{seed_1L_dbl}{Seed (a double vector of length one), Default: 23456}

\item{tfmn_1L_chr}{Transformation (a character vector of length one)}
}
\value{
Summary of bayesian regression model model (a tibble)
}
\description{
make_smry_of_brm_mdl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make summary of bayesian regression model model. The function returns Summary of bayesian regression model model (a tibble).
}
\keyword{internal}
