% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_smry_of_ts_mdl}
\alias{make_smry_of_ts_mdl}
\title{Make smry of time series model}
\usage{
make_smry_of_ts_mdl(
  data_tb,
  fn,
  predr_vars_nms_chr,
  mdl_nm_1L_chr,
  path_to_write_to_1L_chr = NA_character_,
  dep_var_nm_1L_chr = "utl_total_w",
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  backend_1L_chr = getOption("brms.backend", "rstan"),
  iters_1L_int = 4000L,
  seed_1L_int = 1000L
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{fn}{Function (a function)}

\item{predr_vars_nms_chr}{Predr vars names (a character vector)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one), Default: 'NA'}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'utl_total_w'}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round var name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round bl value (a character vector of length one), Default: 'Baseline'}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: getOption("brms.backend", "rstan")}

\item{iters_1L_int}{Iters (an integer vector of length one), Default: 4000}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1000}
}
\value{
Smry of time series model (a list)
}
\description{
make_smry_of_ts_mdl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make smry of time series model. The function returns Smry of time series model (a list).
}
\keyword{internal}
