% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_study_descs_ls}
\alias{make_study_descs_ls}
\title{Make study descriptions}
\usage{
make_study_descs_ls(
  input_params_ls = NULL,
  time_btwn_bl_and_fup_1L_chr,
  background_1L_chr = "",
  coi_1L_chr = "None declared.",
  conclusion_1L_chr = "",
  ethics_1L_chr = NULL,
  funding_1L_chr = NULL,
  health_utl_nm_1L_chr = NULL,
  params_ls_ls = NULL,
  predr_ctgs_ls = NULL,
  sample_desc_1L_chr = NULL,
  var_nm_change_lup = NULL
)
}
\arguments{
\item{input_params_ls}{Input params (a list), Default: NULL}

\item{time_btwn_bl_and_fup_1L_chr}{Time btwn baseline and follow-up (a character vector of length one)}

\item{background_1L_chr}{Background (a character vector of length one), Default: ''}

\item{coi_1L_chr}{Coi (a character vector of length one), Default: 'None declared.'}

\item{conclusion_1L_chr}{Conclusion (a character vector of length one), Default: ''}

\item{ethics_1L_chr}{Ethics (a character vector of length one), Default: NULL}

\item{funding_1L_chr}{Funding (a character vector of length one), Default: NULL}

\item{health_utl_nm_1L_chr}{Health utility name (a character vector of length one), Default: NULL}

\item{params_ls_ls}{Params (a list of lists), Default: NULL}

\item{predr_ctgs_ls}{Predictor category categoriess (a list), Default: NULL}

\item{sample_desc_1L_chr}{Sample description (a character vector of length one), Default: NULL}

\item{var_nm_change_lup}{Variable name change (a lookup table), Default: NULL}
}
\value{
Input params (a list)
}
\description{
make_study_descs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make study descriptions list. The function returns Input params (a list).
}
