% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_synth_series_tbs_ls}
\alias{make_synth_series_tbs_ls}
\title{Make synthetic series tibbles}
\usage{
make_synth_series_tbs_ls(synth_data_spine_ls, series_names_chr)
}
\arguments{
\item{synth_data_spine_ls}{Synthetic data spine (a list)}

\item{series_names_chr}{Series names (a character vector)}
}
\value{
Synthetic series tibbles (a list)
}
\description{
make_synth_series_tbs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make synthetic series tibbles list. The function returns Synthetic series tibbles (a list).
}
\keyword{internal}
