% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ttu_cs_ls}
\alias{make_ttu_cs_ls}
\title{Make transfer to utility algorithm cross-section list}
\usage{
make_ttu_cs_ls(
  outp_smry_ls,
  sig_covars_some_predrs_mdls_tb,
  sig_thresh_covars_1L_chr
)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{sig_covars_some_predrs_mdls_tb}{Sig covariates some predictors models (a tibble)}

\item{sig_thresh_covars_1L_chr}{Sig thresh covariates (a character vector of length one)}
}
\value{
Transfer to utility algorithm cross-section (a list)
}
\description{
make_ttu_cs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make transfer to utility algorithm cross-section list. The function returns Transfer to utility algorithm cross-section (a list).
}
\keyword{internal}
