% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_obsd_predd_dnst}
\alias{plot_obsd_predd_dnst}
\title{Plot obsd predd dnst}
\usage{
plot_obsd_predd_dnst(tfd_data_tb)
}
\arguments{
\item{tfd_data_tb}{Transformed data (a tibble)}
}
\description{
plot_obsd_predd_dnst() is a Plot function that plots data Specifically, this function implements an algorithm to plot obsd predd dnst. The function is called for its side effects and does not return a value.
}
\keyword{internal}
