% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_shrble_glm}
\alias{predict_shrble_glm}
\title{Predict shareable generalised linear model}
\usage{
predict_shrble_glm(
  object,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  dispersion = NULL,
  terms = NULL,
  na.action = na.pass,
  sd_1L_dbl,
  ...
)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{newdata}{New dataset (for prediction) (a new dataset (for prediction)), Default: NULL}

\item{type}{PARAM_DESCRIPTION, Default: c("link", "response", "terms")}

\item{se.fit}{PARAM_DESCRIPTION, Default: FALSE}

\item{dispersion}{PARAM_DESCRIPTION, Default: NULL}

\item{terms}{PARAM_DESCRIPTION, Default: NULL}

\item{na.action}{PARAM_DESCRIPTION, Default: na.pass}

\item{sd_1L_dbl}{Standard deviation (a double vector of length one)}

\item{...}{Additional arguments}
}
\description{
predict_shrble_glm() is a Predict function that makes predictions from data using a specified statistical model. Specifically, this function implements an algorithm to predict shareable generalised linear model. The function is called for its side effects and does not return a value.
}
\keyword{internal}
