% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_uncnstrd_utl}
\alias{predict_uncnstrd_utl}
\title{Predict unconstrained utility}
\usage{
predict_uncnstrd_utl(
  data_tb,
  model_mdl,
  new_data_is_1L_chr = "Predicted",
  predn_type_1L_chr = NULL,
  tfmn_for_bnml_1L_lgl = F,
  deterministic_1L_lgl = T,
  family_1L_chr = NA_character_,
  tfmn_1L_chr = "NTF",
  is_brms_mdl_1L_lgl = F,
  force_new_data_1L_lgl = F,
  sd_dbl = NA_real_
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{model_mdl}{Model (a model)}

\item{new_data_is_1L_chr}{New data is (a character vector of length one), Default: 'Predicted'}

\item{predn_type_1L_chr}{Prediction type (a character vector of length one), Default: NULL}

\item{tfmn_for_bnml_1L_lgl}{Transformation for binomial (a logical vector of length one), Default: F}

\item{deterministic_1L_lgl}{Deterministic (a logical vector of length one), Default: T}

\item{family_1L_chr}{Family (a character vector of length one), Default: 'NA'}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{is_brms_mdl_1L_lgl}{Is bayesian regression models model (a logical vector of length one), Default: F}

\item{force_new_data_1L_lgl}{Force new data (a logical vector of length one), Default: F}

\item{sd_dbl}{Standard deviation (a double vector), Default: NA}
}
\value{
New data (a double vector)
}
\description{
predict_uncnstrd_utl() is a Predict function that makes predictions from data using a specified statistical model. Specifically, this function implements an algorithm to predict unconstrained utility. The function returns New data (a double vector).
}
\keyword{internal}
