% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_utility}
\alias{predict_utility}
\title{Predict utility}
\usage{
predict_utility(
  data_tb,
  tfmn_1L_chr = "NTF",
  model_mdl,
  force_min_max_1L_lgl = T,
  utl_min_val_1L_dbl = 0.03,
  impute_1L_lgl = T,
  utl_cls_fn = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{model_mdl}{Model (a model)}

\item{force_min_max_1L_lgl}{Force minimum maximum (a logical vector of length one), Default: T}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: 0.03}

\item{impute_1L_lgl}{Impute (a logical vector of length one), Default: T}

\item{utl_cls_fn}{Utility class (a function), Default: NULL}
}
\value{
Predicted utility (a double vector)
}
\description{
predict_utility() is a Predict function that makes predictions from data using a specified statistical model. Specifically, this function implements an algorithm to predict utility. The function returns Predicted utility (a double vector).
}
