% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_utility}
\alias{predict_utility}
\title{Predict utility}
\usage{
predict_utility(
  data_tb,
  tfmn_1L_chr = "NTF",
  model_mdl,
  force_min_max_1L_lgl = T,
  force_new_data_1L_lgl = F,
  utl_min_val_1L_dbl = 0.03,
  impute_1L_lgl = T,
  utl_cls_fn = NULL,
  new_data_is_1L_chr = "Predicted",
  predn_type_1L_chr = NULL,
  sd_dbl = NA_real_,
  tfmn_for_bnml_1L_lgl = F,
  family_1L_chr = NA_character_,
  is_brms_mdl_1L_lgl = T
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{model_mdl}{Model (a model)}

\item{force_min_max_1L_lgl}{Force minimum maximum (a logical vector of length one), Default: T}

\item{force_new_data_1L_lgl}{Force new data (a logical vector of length one), Default: F}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: 0.03}

\item{impute_1L_lgl}{Impute (a logical vector of length one), Default: T}

\item{utl_cls_fn}{Utility class (a function), Default: NULL}

\item{new_data_is_1L_chr}{New data is (a character vector of length one), Default: 'Predicted'}

\item{predn_type_1L_chr}{Prediction type (a character vector of length one), Default: NULL}

\item{sd_dbl}{Standard deviation (a double vector), Default: NA}

\item{tfmn_for_bnml_1L_lgl}{Transformation for binomial (a logical vector of length one), Default: F}

\item{family_1L_chr}{Family (a character vector of length one), Default: 'NA'}

\item{is_brms_mdl_1L_lgl}{Is bayesian regression models model (a logical vector of length one), Default: T}
}
\value{
Predicted utility (a double vector)
}
\description{
predict_utility() is a Predict function that makes predictions from data using a specified statistical model. Specifically, this function implements an algorithm to predict utility. The function returns Predicted utility (a double vector).
}
\keyword{internal}
