% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_table_xx}
\alias{print_table_xx}
\title{Print table}
\usage{
print_table_xx(
  data_tb,
  output_type_1L_chr = "PDF",
  caption_1L_chr = NA_character_,
  footnotes_chr = NA_character_,
  merge_row_idx_int = NA_integer_,
  digits_dbl = NULL,
  big_mark_1L_chr = " ",
  label,
  hline.after,
  addtorow,
  sanitize_fn = getOption("xtable.sanitize.text.function", NULL)
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'PDF'}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: 'NA'}

\item{footnotes_chr}{Footnotes (a character vector), Default: 'NA'}

\item{merge_row_idx_int}{Merge row index (an integer vector), Default: NA}

\item{digits_dbl}{Digits (a double vector), Default: NULL}

\item{big_mark_1L_chr}{Big mark (a character vector of length one), Default: ' '}

\item{label}{PARAM_DESCRIPTION}

\item{hline.after}{PARAM_DESCRIPTION}

\item{addtorow}{PARAM_DESCRIPTION}

\item{sanitize_fn}{Sanitize (a function), Default: getOption("xtable.sanitize.text.function", NULL)}
}
\description{
print_table_xx() is a Print function that prints output to console Specifically, this function implements an algorithm to print table output object of multiple potential types. The function is called for its side effects and does not return a value.
}
