% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reorder.R
\name{reorder_cndt_predrs_chr}
\alias{reorder_cndt_predrs_chr}
\title{Reorder candidate predictors character vector}
\usage{
reorder_cndt_predrs_chr(
  candidate_predrs_chr,
  data_tb,
  depnt_var_nm_1L_chr = "utl_total_w",
  method_1L_chr = "pearson"
)
}
\arguments{
\item{candidate_predrs_chr}{Candidate predictors (a character vector)}

\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{method_1L_chr}{Method (a character vector of length one), Default: 'pearson'}
}
\value{
Reordered candidate (predictors)
}
\description{
reorder_cndt_predrs_chr() is a Reorder function that reorders an object to conform to a pre-specified schema. Specifically, this function implements an algorithm to reorder candidate predictors character vector. The function is called for its side effects and does not return a value.
}
\keyword{internal}
