% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reorder.R
\name{reorder_tb_for_target_cors}
\alias{reorder_tb_for_target_cors}
\title{Reorder tibble for target cors}
\usage{
reorder_tb_for_target_cors(
  tbs_ls,
  corr_dbl,
  corr_var_1L_chr,
  id_var_to_rm_1L_chr = NA_character_
)
}
\arguments{
\item{tbs_ls}{Tibbles (a list)}

\item{corr_dbl}{Corr (a double vector)}

\item{corr_var_1L_chr}{Corr var (a character vector of length one)}

\item{id_var_to_rm_1L_chr}{Id var to rm (a character vector of length one), Default: 'NA'}
}
\value{
Tibbles (a list)
}
\description{
reorder_tb_for_target_cors() is a Reorder function that reorders an object to conform to a pre-specified schema. Specifically, this function implements an algorithm to reorder tibble for target cors. The function returns Tibbles (a list).
}
