% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_data_tb_for_cmprsn}
\alias{transform_data_tb_for_cmprsn}
\title{Transform data tibble for comparison}
\usage{
transform_data_tb_for_cmprsn(
  data_tb,
  model_mdl,
  depnt_var_nm_1L_chr = "utl_total_w",
  source_data_nm_1L_chr = "Original",
  new_data_is_1L_chr = "Predicted",
  predn_type_1L_chr = NULL,
  family_1L_chr = NA_character_,
  impute_1L_lgl = F,
  is_brms_mdl_1L_lgl = F,
  sd_dbl = NA_real_,
  sfx_1L_chr = "",
  tfmn_for_bnml_1L_lgl = F,
  tfmn_1L_chr = "NTF",
  utl_cls_fn = NULL,
  utl_min_val_1L_dbl = NA_real_
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{model_mdl}{Model (a model)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{source_data_nm_1L_chr}{Source data name (a character vector of length one), Default: 'Original'}

\item{new_data_is_1L_chr}{New data is (a character vector of length one), Default: 'Predicted'}

\item{predn_type_1L_chr}{Prediction type (a character vector of length one), Default: NULL}

\item{family_1L_chr}{Family (a character vector of length one), Default: 'NA'}

\item{impute_1L_lgl}{Impute (a logical vector of length one), Default: F}

\item{is_brms_mdl_1L_lgl}{Is bayesian regression models model (a logical vector of length one), Default: F}

\item{sd_dbl}{Standard deviation (a double vector), Default: NA}

\item{sfx_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{tfmn_for_bnml_1L_lgl}{Transformation for binomial (a logical vector of length one), Default: F}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{utl_cls_fn}{Utility class (a function), Default: NULL}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: NA}
}
\value{
Transformed data (a tibble)
}
\description{
transform_data_tb_for_cmprsn() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform data tibble for comparison. Function argument data_tb specifies the object to be updated. Argument model_mdl provides the object to be updated. The function returns Transformed data (a tibble).
}
