% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ts_mdl_data}
\alias{transform_ts_mdl_data}
\title{Transform time series model data}
\usage{
transform_ts_mdl_data(
  mdl_ls,
  data_tb,
  depnt_var_nm_1L_chr = "utl_total_w",
  predr_vars_nms_chr,
  id_var_nm_1L_chr = "fkClientID",
  mdl_nm_1L_chr
)
}
\arguments{
\item{mdl_ls}{Model list (a list of models)}

\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{predr_vars_nms_chr}{Predictor variables names (a character vector)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}
}
\value{
Cnfdl (a list of models)
}
\description{
transform_ts_mdl_data() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform time series model data. Function argument mdl_ls specifies the object to be updated. Argument data_tb provides the object to be updated. The function returns Cnfdl (a list of models).
}
\keyword{internal}
