% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_TTU.R
\docType{package}
\name{TTU-package}
\alias{TTU}
\alias{TTU-package}
\title{TTU: Transfer to Utility Mapping Algorithm Toolkit}
\description{
Tools for developping, sharing and reporting Transfer To
Utility (TTU) mapping algorithms that predict health utility from
other health measures.  This development version of the TTU package
has been made available as part of the process of testing and
documenting the package.  Some of the documentation for this package
has been automatically generated by the ready4fun package and is
therefore quite rudimentary. Human edits to improve the quality of
documentation will follow in 2021.  If you have any questions, please
contact the authors (matthew.hamilton@orygen.org.au).
}
\details{
To learn more about TTU, start with the vignettes:
\code{browseVignettes(package = "TTU")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/TTU/}
  \item \url{https://github.com/ready4-dev/TTU}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
