% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_brm_model_plts}
\alias{write_brm_model_plts}
\title{Write bayesian regression model model plots}
\usage{
write_brm_model_plts(
  mdl_ls,
  tfd_data_tb,
  mdl_nm_1L_chr,
  path_to_write_to_1L_chr,
  depnt_var_nm_1L_chr = "utl_total_w",
  depnt_var_desc_1L_chr = "Utility score",
  round_var_nm_1L_chr = "round",
  tfmn_1L_chr = "NTF",
  units_1L_chr = "in",
  height_dbl = c(rep(6, 2), rep(5, 3)),
  width_dbl = c(rep(6, 2), rep(6, 3)),
  rsl_dbl = rep(300, 5),
  args_ls = NULL,
  seed_1L_dbl = 23456
)
}
\arguments{
\item{mdl_ls}{Model list (a list of models)}

\item{tfd_data_tb}{Transformed data (a tibble)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{depnt_var_desc_1L_chr}{Dependent variable description (a character vector of length one), Default: 'Utility score'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{units_1L_chr}{Units (a character vector of length one), Default: 'in'}

\item{height_dbl}{Height (a double vector), Default: c(rep(6, 2), rep(5, 3))}

\item{width_dbl}{Width (a double vector), Default: c(rep(6, 2), rep(6, 3))}

\item{rsl_dbl}{Resolution (a double vector), Default: rep(300, 5)}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{seed_1L_dbl}{Seed (a double vector of length one), Default: 23456}
}
\value{
Model plots paths (a list)
}
\description{
write_brm_model_plts() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write bayesian regression model model plots. The function returns Model plots paths (a list).
}
\keyword{internal}
