% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_csp_output}
\alias{write_csp_output}
\title{Write complete study program output}
\usage{
write_csp_output(
  path_to_csp_1L_chr,
  dv_ds_doi_1L_chr = NULL,
  execute_1L_lgl = T
)
}
\arguments{
\item{path_to_csp_1L_chr}{Path to complete study program (a character vector of length one)}

\item{dv_ds_doi_1L_chr}{Dataverse dataset digital object identifier (a character vector of length one), Default: NULL}

\item{execute_1L_lgl}{Execute (a logical vector of length one), Default: T}
}
\description{
write_csp_output() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write complete study program output. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
