% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_manuscript}
\alias{write_manuscript}
\title{Write manuscript}
\usage{
write_manuscript(
  abstract_args_ls = NULL,
  input_params_ls = NULL,
  results_ls = NULL,
  figures_in_body_lgl = NULL,
  output_type_1L_chr = NULL,
  tables_in_body_lgl = NULL,
  title_1L_chr = "Scientific manuscript",
  version_1L_chr = "0.5",
  write_to_dv_1L_lgl = F
)
}
\arguments{
\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}

\item{input_params_ls}{Input parameters (a list), Default: NULL}

\item{results_ls}{Results (a list), Default: NULL}

\item{figures_in_body_lgl}{Figures in body (a logical vector), Default: NULL}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: NULL}

\item{tables_in_body_lgl}{Tables in body (a logical vector), Default: NULL}

\item{title_1L_chr}{Title (a character vector of length one), Default: 'Scientific manuscript'}

\item{version_1L_chr}{Version (a character vector of length one), Default: '0.5'}

\item{write_to_dv_1L_lgl}{Write to dataverse (a logical vector of length one), Default: F}
}
\value{
Results (a list)
}
\description{
write_manuscript() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write manuscript. The function returns Results (a list).
}
