% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mdl_smry_rprt}
\alias{write_mdl_smry_rprt}
\title{Write model summary report}
\usage{
write_mdl_smry_rprt(
  header_yaml_args_ls,
  path_params_ls,
  use_fake_data_1L_lgl = FALSE,
  output_format_ls,
  abstract_args_ls = NULL,
  dv_ls = NULL,
  reference_1L_int = NULL,
  rprt_lup = NULL,
  rcrd_nm_1L_chr = "Write_Rprt_Rcrd",
  rprt_nm_1L_chr = "TS_TTU_Mdls_Smry",
  rprt_subtitle_1L_chr = NULL,
  subtitle_1L_chr = NULL,
  use_shareable_mdls_1L_lgl = F
)
}
\arguments{
\item{header_yaml_args_ls}{Header yaml arguments (a list)}

\item{path_params_ls}{Path params (a list)}

\item{use_fake_data_1L_lgl}{Use fake data (a logical vector of length one), Default: FALSE}

\item{output_format_ls}{Output format (a list)}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}

\item{dv_ls}{Dataverse (a list), Default: NULL}

\item{reference_1L_int}{Reference (an integer vector of length one), Default: NULL}

\item{rprt_lup}{Report (a lookup table), Default: NULL}

\item{rcrd_nm_1L_chr}{Rcrd name (a character vector of length one), Default: 'Write_Rprt_Rcrd'}

\item{rprt_nm_1L_chr}{Report name (a character vector of length one), Default: 'TS_TTU_Mdls_Smry'}

\item{rprt_subtitle_1L_chr}{Report subtitle (a character vector of length one), Default: NULL}

\item{subtitle_1L_chr}{Subtitle (a character vector of length one), Default: NULL}

\item{use_shareable_mdls_1L_lgl}{Use shareable models (a logical vector of length one), Default: F}
}
\description{
write_mdl_smry_rprt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write model summary report. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
