% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mdl_type_covars_mdls}
\alias{write_mdl_type_covars_mdls}
\title{Write model type covariates models}
\usage{
write_mdl_type_covars_mdls(
  data_tb,
  depnt_var_nm_1L_chr = "utl_total_w",
  predrs_var_nms_chr,
  covar_var_nms_chr,
  mdl_type_1L_chr,
  path_to_write_to_1L_chr,
  new_dir_nm_1L_chr = "D_Covars_Selection",
  fl_nm_pfx_1L_chr = "D_CT",
  mdl_types_lup = NULL,
  start_1L_chr = NA_character_
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{predrs_var_nms_chr}{Predictors variable names (a character vector)}

\item{covar_var_nms_chr}{Covariate variable names (a character vector)}

\item{mdl_type_1L_chr}{Model type (a character vector of length one)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'D_Covars_Selection'}

\item{fl_nm_pfx_1L_chr}{File name prefix (a character vector of length one), Default: 'D_CT'}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}

\item{start_1L_chr}{Start (a character vector of length one), Default: 'NA'}
}
\value{
Summary of models with covariates (a tibble)
}
\description{
write_mdl_type_covars_mdls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write model type covariates models. The function returns Summary of models with covariates (a tibble).
}
\keyword{internal}
