% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_rndrd_rprt}
\alias{write_rndrd_rprt}
\title{Write rndrd rprt}
\usage{
write_rndrd_rprt(
  path_to_RMD_1L_chr,
  path_to_rprt_dir_1L_chr,
  file_nm_1L_chr,
  params_ls = list(output_type_1L_chr = "PDF")
)
}
\arguments{
\item{path_to_RMD_1L_chr}{Path toMD (a character vector of length one)}

\item{path_to_rprt_dir_1L_chr}{Path to rprt directory (a character vector of length one)}

\item{file_nm_1L_chr}{File name (a character vector of length one)}

\item{params_ls}{Params (a list), Default: list(output_type_1L_chr = "PDF")}
}
\description{
write_rndrd_rprt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write rndrd rprt. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
