% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_rndrd_rprt}
\alias{write_rndrd_rprt}
\title{Write rndrd rprt}
\usage{
write_rndrd_rprt(
  path_to_RMD_dir_1L_chr,
  nm_of_RMD_1L_chr = "report.RMD",
  params_ls = list(output_type_1L_chr = "HTML"),
  rltv_path_to_outpt_yaml_1L_chr = "output_yml",
  paths_to_fls_to_copy_chr = NA_character_,
  path_to_write_fls_to_1L_chr = NA_character_,
  nm_of_rprt_dir_1L_chr = "Reports",
  path_to_outpt_rtrp_1L_chr = "./",
  file_nm_1L_chr,
  overwrite_1L_lgl = T
)
}
\arguments{
\item{path_to_RMD_dir_1L_chr}{Path toMD directory (a character vector of length one)}

\item{nm_of_RMD_1L_chr}{Name ofMD (a character vector of length one), Default: 'report.RMD'}

\item{params_ls}{Params (a list), Default: list(output_type_1L_chr = "HTML")}

\item{rltv_path_to_outpt_yaml_1L_chr}{Rltv path to outpt yaml (a character vector of length one), Default: 'output_yml'}

\item{paths_to_fls_to_copy_chr}{Paths to files to copy (a character vector), Default: 'NA'}

\item{path_to_write_fls_to_1L_chr}{Path to write files to (a character vector of length one), Default: 'NA'}

\item{nm_of_rprt_dir_1L_chr}{Name of rprt directory (a character vector of length one), Default: 'Reports'}

\item{path_to_outpt_rtrp_1L_chr}{Path to outpt rtrp (a character vector of length one), Default: './'}

\item{file_nm_1L_chr}{File name (a character vector of length one)}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: T}
}
\description{
write_rndrd_rprt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write rndrd rprt. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
