% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_rprt}
\alias{write_rprt}
\title{Write rprt}
\usage{
write_rprt(
  rprt_type_ls,
  outp_smry_ls,
  output_type_1L_chr = "PDF",
  section_type_1L_chr = "#",
  path_to_outpt_dir_1L_chr = "../../../../Data/Project/Utility_Models",
  reports_dir_1L_chr = "Reports",
  rltv_path_to_data_dir_1L_chr = "../Output",
  nm_of_mkdn_dir_1L_chr = "Markdown"
)
}
\arguments{
\item{rprt_type_ls}{Rprt type (a list)}

\item{outp_smry_ls}{Output smry (a list)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'PDF'}

\item{section_type_1L_chr}{Section type (a character vector of length one), Default: '#'}

\item{path_to_outpt_dir_1L_chr}{Path to outpt directory (a character vector of length one), Default: '../../../../Data/Project/Utility_Models'}

\item{reports_dir_1L_chr}{Reports directory (a character vector of length one), Default: 'Reports'}

\item{rltv_path_to_data_dir_1L_chr}{Rltv path to data directory (a character vector of length one), Default: '../Output'}

\item{nm_of_mkdn_dir_1L_chr}{Name of mkdn directory (a character vector of length one), Default: 'Markdown'}
}
\value{
Output smry (a list)
}
\description{
write_rprt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write rprt. The function returns Output smry (a list).
}
