% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scndry_analysis}
\alias{write_scndry_analysis}
\title{Write scndry analysis}
\usage{
write_scndry_analysis(
  analysis_params_ls,
  candidate_covar_nms_chr,
  candidate_predrs_chr,
  prefd_covars_chr,
  header_yaml_args_ls,
  subtitle_1L_chr,
  reference_1L_int,
  paths_ls,
  R_fl_nm_1L_chr,
  rprt_lup,
  abstract_args_ls = NULL
)
}
\arguments{
\item{analysis_params_ls}{Analysis params (a list)}

\item{candidate_covar_nms_chr}{Candidate covariate names (a character vector)}

\item{candidate_predrs_chr}{Candidate predictors (a character vector)}

\item{prefd_covars_chr}{Preferred covariates (a character vector)}

\item{header_yaml_args_ls}{Header yaml arguments (a list)}

\item{subtitle_1L_chr}{Subtitle (a character vector of length one)}

\item{reference_1L_int}{Reference (an integer vector of length one)}

\item{paths_ls}{Paths (a list)}

\item{R_fl_nm_1L_chr}{R file name (a character vector of length one)}

\item{rprt_lup}{Report (a lookup table)}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}
}
\description{
write_scndry_analysis() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write scndry analysis. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
