% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_shareable_mdls}
\alias{write_shareable_mdls}
\title{Write shareable models}
\usage{
write_shareable_mdls(
  outp_smry_ls,
  new_dir_nm_1L_chr = "G_Shareable",
  shareable_title_detail_1L_chr = "",
  write_mdls_to_dv_1L_lgl = F
)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'G_Shareable'}

\item{shareable_title_detail_1L_chr}{Shareable title detail (a character vector of length one), Default: ''}

\item{write_mdls_to_dv_1L_lgl}{Write models to dataverse (a logical vector of length one), Default: F}
}
\value{
Output summary (a list)
}
\description{
write_shareable_mdls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write shareable models. The function returns Output summary (a list).
}
