% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_sngl_predr_multi_mdls_outps}
\alias{write_sngl_predr_multi_mdls_outps}
\title{Write single predictor multi models outputs}
\usage{
write_sngl_predr_multi_mdls_outps(
  data_tb,
  mdl_types_chr,
  predr_var_nm_1L_chr,
  predr_var_desc_1L_chr,
  predr_vals_dbl,
  path_to_write_to_1L_chr,
  new_dir_nm_1L_chr = "A_Candidate_Mdls_Cmprsn",
  start_1L_chr = NULL,
  covar_var_nms_chr = NA_character_,
  depnt_var_nm_1L_chr = "utl_total_w",
  folds_1L_int = 10,
  mdl_types_lup = NULL,
  fl_nm_pfx_1L_chr = "A_RT_",
  plt_idxs_int = NA_integer_,
  dictionary_tb
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{mdl_types_chr}{Model types (a character vector)}

\item{predr_var_nm_1L_chr}{Predictor variable name (a character vector of length one)}

\item{predr_var_desc_1L_chr}{Predictor variable description (a character vector of length one)}

\item{predr_vals_dbl}{Predictor values (a double vector)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'A_Candidate_Mdls_Cmprsn'}

\item{start_1L_chr}{Start (a character vector of length one), Default: NULL}

\item{covar_var_nms_chr}{Covariate variable names (a character vector), Default: 'NA'}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{folds_1L_int}{Folds (an integer vector of length one), Default: 10}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}

\item{fl_nm_pfx_1L_chr}{File name prefix (a character vector of length one), Default: 'A_RT_'}

\item{plt_idxs_int}{Plot indices (an integer vector), Default: NA}

\item{dictionary_tb}{Dictionary (a tibble)}
}
\value{
Summary of single predictor models (a tibble)
}
\description{
write_sngl_predr_multi_mdls_outps() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write single predictor multi models outputs. The function returns Summary of single predictor models (a tibble).
}
\keyword{internal}
