% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_study_outp_ds}
\alias{write_study_outp_ds}
\title{Write study output dataset}
\usage{
write_study_outp_ds(
  input_params_ls,
  dv_ds_nm_and_url_chr = NULL,
  rprt_lups_ls = NULL,
  output_format_ls = NULL,
  path_params_ls = NULL,
  abstract_args_ls = NULL,
 
    dv_mdl_desc_1L_chr = "This is a longitudinal transfer to utility model designed for use with the youthu R package.",
  header_yaml_args_ls = NULL,
  inc_fl_types_chr = ".pdf",
  purge_data_1L_lgl = FALSE,
  start_at_int = c(2, 1),
  use_fake_data_1L_lgl = NULL
)
}
\arguments{
\item{input_params_ls}{Input parameters (a list)}

\item{dv_ds_nm_and_url_chr}{Dataverse dataset name and url (a character vector), Default: NULL}

\item{rprt_lups_ls}{Report lookup tables (a list), Default: NULL}

\item{output_format_ls}{Output format (a list), Default: NULL}

\item{path_params_ls}{Path parameters (a list), Default: NULL}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}

\item{dv_mdl_desc_1L_chr}{Dataverse model description (a character vector of length one), Default: 'This is a longitudinal transfer to utility model designed for use with the youthu R package.'}

\item{header_yaml_args_ls}{Header yaml arguments (a list), Default: NULL}

\item{inc_fl_types_chr}{Include file types (a character vector), Default: '.pdf'}

\item{purge_data_1L_lgl}{Purge data (a logical vector of length one), Default: FALSE}

\item{start_at_int}{Start at (an integer vector), Default: c(2, 1)}

\item{use_fake_data_1L_lgl}{Use fake data (a logical vector of length one), Default: NULL}
}
\value{
Dataverse dataset name and url (a character vector)
}
\description{
write_study_outp_ds() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write study output dataset. The function returns Dataverse dataset name and url (a character vector).
}
\keyword{internal}
