% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_study_outp_ds}
\alias{write_study_outp_ds}
\title{Write study output dataset}
\usage{
write_study_outp_ds(
  dv_ls,
  output_format_ls,
  path_params_ls,
  abstract_args_ls = NULL,
 
    dv_mdl_desc_1L_chr = "This is a time series transfer to utility model designed for use with the youthu R package.",
  inc_fl_types_chr = ".pdf",
  purge_data_1L_lgl = FALSE,
  reference_1L_int = NULL,
  rprt_lup = NULL,
  subtitle_1L_chr = NULL,
  use_fake_data_1L_lgl = F
)
}
\arguments{
\item{dv_ls}{Dataverse (a list)}

\item{output_format_ls}{Output format (a list)}

\item{path_params_ls}{Path params (a list)}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}

\item{dv_mdl_desc_1L_chr}{Dataverse model description (a character vector of length one), Default: 'This is a time series transfer to utility model designed for use with the youthu R package.'}

\item{inc_fl_types_chr}{Include file types (a character vector), Default: '.pdf'}

\item{purge_data_1L_lgl}{Purge data (a logical vector of length one), Default: FALSE}

\item{reference_1L_int}{Reference (an integer vector of length one), Default: NULL}

\item{rprt_lup}{Report (a lookup table), Default: NULL}

\item{subtitle_1L_chr}{Subtitle (a character vector of length one), Default: NULL}

\item{use_fake_data_1L_lgl}{Use fake data (a logical vector of length one), Default: F}
}
\description{
write_study_outp_ds() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write study output dataset. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
