% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ts_mdls}
\alias{write_ts_mdls}
\title{Write time series models}
\usage{
write_ts_mdls(
  data_tb,
  depnt_var_nm_1L_chr = "utl_total_w",
  predr_vars_nms_ls,
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  utl_min_val_1L_dbl = -1,
  backend_1L_chr = getOption("brms.backend", "rstan"),
  mdl_nms_ls,
  mdl_smry_dir_1L_chr,
  predictors_lup,
  iters_1L_int = 4000L,
  mdl_types_lup,
  seed_1L_int = 1000L,
  prior_ls = NULL,
  control_ls = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{predr_vars_nms_ls}{Predictor variables names (a list)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round baseline value (a character vector of length one), Default: 'Baseline'}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: -1}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: getOption("brms.backend", "rstan")}

\item{mdl_nms_ls}{Model names (a list)}

\item{mdl_smry_dir_1L_chr}{Model summary directory (a character vector of length one)}

\item{predictors_lup}{Predictors (a lookup table)}

\item{iters_1L_int}{Iterations (an integer vector of length one), Default: 4000}

\item{mdl_types_lup}{Model types (a lookup table)}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1000}

\item{prior_ls}{Prior (a list), Default: NULL}

\item{control_ls}{Control (a list), Default: NULL}
}
\value{
Models summary (a tibble)
}
\description{
write_ts_mdls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write time series models. The function returns Models summary (a tibble).
}
\keyword{internal}
