% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ts_mdls_from_alg_outp}
\alias{write_ts_mdls_from_alg_outp}
\title{Write ts models from algorithm output}
\usage{
write_ts_mdls_from_alg_outp(
  outp_smry_ls,
  fn_ls,
  backend_1L_chr = getOption("brms.backend", "rstan"),
  iters_1L_int = 4000L
)
}
\arguments{
\item{outp_smry_ls}{Output smry (a list)}

\item{fn_ls}{Function list (a list of functions)}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: getOption("brms.backend", "rstan")}

\item{iters_1L_int}{Iters (an integer vector of length one), Default: 4000}
}
\value{
Output smry (a list)
}
\description{
write_ts_mdls_from_alg_outp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write ts models from algorithm output. The function returns Output smry (a list).
}
