% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ts_mdls_from_alg_outp}
\alias{write_ts_mdls_from_alg_outp}
\title{Write time series models from algorithm output}
\usage{
write_ts_mdls_from_alg_outp(
  outp_smry_ls,
  predictors_lup,
  utl_min_val_1L_dbl = -1,
  backend_1L_chr = getOption("brms.backend", "rstan"),
  iters_1L_int = 4000L,
  new_dir_nm_1L_chr = "F_TS_Mdls",
  prior_ls = NULL,
  control_ls = NULL
)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{predictors_lup}{Predictors (a lookup table)}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: -1}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: getOption("brms.backend", "rstan")}

\item{iters_1L_int}{Iterations (an integer vector of length one), Default: 4000}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'F_TS_Mdls'}

\item{prior_ls}{Prior (a list), Default: NULL}

\item{control_ls}{Control (a list), Default: NULL}
}
\value{
Output summary (a list)
}
\description{
write_ts_mdls_from_alg_outp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write time series models from algorithm output. The function returns Output summary (a list).
}
\keyword{internal}
